Extract comments from AirDAS data

Extract comments from airdas_dfr or airdas_df object

airdas_comments(x)

# S3 method for data.frame
airdas_comments(x)

# S3 method for airdas_df
airdas_comments(x)

# S3 method for airdas_dfr
airdas_comments(x)

Arguments

x
airdas_dfr or airdas_df object, or a data frame that can be coerced to a airdas_dfr object

Value

x, with the added column comment_str containing the concatenated comment strings

Details

This function recreates the comment strings by pasting the Data# columns back together for the C events (comments)

See the examples section for how to search for comments with the phrase "record" to determine what extra information (e.g. molas) was being recorded vs ignored.

Examples

y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.read <- airdas_read(y, file.type = "turtle")

airdas_comments(y.read)

#>    Event EffortDot            DateTime      Lat       Lon Data1 Data2 Data3
#> 1      C     FALSE 2015-04-09 12:30:59 39.23550 -123.1563  Reco rder:  dd  
#> 2      C     FALSE 2015-04-09 12:30:59 39.23550 -123.1563  Not  recor ding 
#> 3      C      TRUE 2015-04-09 12:31:48 39.23350 -123.1857  2 cp  <NA>  <NA>
#> 4      C      TRUE 2015-04-09 12:34:56 39.23317 -123.2968  stre aky r ed ti
#> 5      C      TRUE 2015-04-09 12:35:32 39.23317 -123.3178  shar k 11f t    
#> 6      C     FALSE 2015-04-09 12:38:00 39.23400 -123.4047  off  effor t to 
#> 7      C      TRUE 2015-04-09 12:40:36 39.23317 -123.4960  wate r got  more
#> 8      C      TRUE 2015-04-09 12:42:01 39.23267 -123.5433  fb2s  fb1m  <NA>
#> 9      C      TRUE 2015-04-09 12:52:39 39.21700 -123.4072  bell y has  some
#> 10     C      TRUE 2015-04-09 12:56:43 39.21717 -123.2718  patc hy re d tid
#>    Data4 Data5 Data6           Data7 EventNum          file_das line_num
#> 1   <NA>  <NA>  <NA>            <NA>        1 airdas_sample.das        1
#> 2  molas  <NA>  <NA>            <NA>        2 airdas_sample.das        2
#> 3   <NA>  <NA>  <NA>            <NA>       11 airdas_sample.das       11
#> 4  de     <NA>  <NA>            <NA>       21 airdas_sample.das       22
#> 5   <NA>  <NA>  <NA>            <NA>       24 airdas_sample.das       25
#> 6  circl e on  unide ntifed object         37 airdas_sample.das       38
#> 7   blue  <NA>  <NA>            <NA>       41 airdas_sample.das       42
#> 8   <NA>  <NA>  <NA>            <NA>       45 airdas_sample.das       46
#> 9   glar e      <NA>            <NA>       72 airdas_sample.das       73
#> 10 e      <NA>  <NA>            <NA>       83 airdas_sample.das       84
#>    file_type                                   comment_str
#> 1     turtle                                Recorder: dd  
#> 2     turtle                           Not recording molas
#> 3     turtle                                          2 cp
#> 4     turtle                           streaky red tide   
#> 5     turtle                                shark 11ft    
#> 6     turtle  off effort to circle on unidentifed object  
#> 7     turtle                           water got more blue
#> 8     turtle                                     fb2s fb1m
#> 9     turtle                      belly has some glare    
#> 10    turtle                           patchy red tide    

# Extract all comments containing "record"
y.comm <- airdas_comments(y.read)
y.comm[grepl("record", y.comm$comment_str, ignore.case = TRUE), ]

#>   Event EffortDot            DateTime     Lat       Lon Data1 Data2 Data3 Data4
#> 1     C     FALSE 2015-04-09 12:30:59 39.2355 -123.1563  Reco rder:  dd    <NA>
#> 2     C     FALSE 2015-04-09 12:30:59 39.2355 -123.1563  Not  recor ding  molas
#>   Data5 Data6 Data7 EventNum          file_das line_num file_type
#> 1  <NA>  <NA>  <NA>        1 airdas_sample.das        1    turtle
#> 2  <NA>  <NA>  <NA>        2 airdas_sample.das        2    turtle
#>            comment_str
#> 1       Recorder: dd  
#> 2  Not recording molas

# Extract all comments containing "record", but not "recorder"
y.comm <- airdas_comments(y.read)
y.comm[grepl("record", y.comm$comment_str, ignore.case = TRUE) &
         !grepl("recorder", y.comm$comment_str, ignore.case = TRUE), ]

#>   Event EffortDot            DateTime     Lat       Lon Data1 Data2 Data3 Data4
#> 2     C     FALSE 2015-04-09 12:30:59 39.2355 -123.1563  Not  recor ding  molas
#>   Data5 Data6 Data7 EventNum          file_das line_num file_type
#> 2  <NA>  <NA>  <NA>        2 airdas_sample.das        2    turtle
#>            comment_str
#> 2  Not recording molas