Chop AirDAS effort segment length

Determine lengths of segments into which to chop aerial DAS effort data

airdas_effort_chop(das.df, seg.km, randpicks = NULL)

Arguments

das.df
data frame; processed aerial DAS data output from airdas_process (or subsequent airdas function)

seg.km
numeric; length of effort segments

randpicks
numeric; if not NULL, segment positions that get leftover segment bits todo

Value

List of...

Details

Chop continuous effort section into effort segments (modeling segments)

Calculates dist_from_prev using swfscMisc::distance, if necessary

randpicks description

If segment length is 0, nothing..?

If segment length is less than target length, only make one segment

If segment length is greater than the target length and remainder is greater than or equal to half of the target length, the remainder is its own (randomly placed) segment

If segment length is greater than the target length and remainder is less than half of the target length, the remainder added to a random segment

Examples

y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.proc <- airdas_process(y)

airdas_effort_chop(y.proc[1:33, ], 1)

#> [[1]]
#>    Event            DateTime      Lat       Lon OnEffort Trans Bft CCover Jelly
#> 1      T 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 2      V 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 3      P 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 4      A 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 5      W 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1   1     10     0
#> 6      * 2015-04-09 12:31:00 39.23567 -123.1575     TRUE    T1   1     10     0
#> 7      S 2015-04-09 12:31:35 39.23367 -123.1783     TRUE    T1   1     10     0
#> 8      S 2015-04-09 12:31:47 39.23350 -123.1857     TRUE    T1   1     10     0
#> 9      S 2015-04-09 12:31:56 39.23333 -123.1917     TRUE    T1   1     10     0
#> 10     * 2015-04-09 12:32:00 39.23317 -123.1930     TRUE    T1   1     10     0
#> 11     S 2015-04-09 12:32:20 39.23233 -123.2053     TRUE    T1   1     10     0
#> 12     S 2015-04-09 12:32:30 39.23233 -123.2108     TRUE    T1   1     10     0
#> 13     S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 14     1 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 15     * 2015-04-09 12:33:00 39.23267 -123.2292     TRUE    T1   1     10     0
#> 16     S 2015-04-09 12:33:43 39.23333 -123.2550     TRUE    T1   1     10     0
#> 17     * 2015-04-09 12:34:00 39.23300 -123.2637     TRUE    T1   1     10     0
#> 18     W 2015-04-09 12:34:52 39.23317 -123.2945     TRUE    T1   1     10     2
#> 19     C 2015-04-09 12:34:56 39.23317 -123.2968     TRUE    T1   1     10     2
#> 20     * 2015-04-09 12:35:00 39.23317 -123.2985     TRUE    T1   1     10     2
#> 21     S 2015-04-09 12:35:13 39.23317 -123.3073     TRUE    T1   1     10     2
#> 22     C 2015-04-09 12:35:32 39.23317 -123.3178     TRUE    T1   1     10     2
#> 23     * 2015-04-09 12:36:00 39.23333 -123.3337     TRUE    T1   1     10     2
#> 24     S 2015-04-09 12:36:13 39.23333 -123.3418     TRUE    T1   1     10     2
#> 25     S 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 26     t 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 27     S 2015-04-09 12:36:43 39.23350 -123.3592     TRUE    T1   1     10     2
#> 28     * 2015-04-09 12:37:01 39.23383 -123.3685     TRUE    T1   1     10     2
#> 29     S 2015-04-09 12:37:01 39.23400 -123.3697     TRUE    T1   1     10     2
#> 30     S 2015-04-09 12:37:16 39.23433 -123.3790     TRUE    T1   1     10     2
#> 31     S 2015-04-09 12:37:37 39.23417 -123.3910     TRUE    T1   1     10     2
#> 32     W 2015-04-09 12:37:57 39.23400 -123.4033     TRUE    T1   2     10     0
#> 33     E 2015-04-09 12:38:00 39.23400 -123.4047    FALSE    T1   2     10     0
#>    HorizSun  HKR ObsL ObsB ObsR  Rec AltFt SpKnot  VLI  VLO   VB  VRI  VRO
#> 1        NA <NA> <NA> <NA> <NA> <NA>    NA     NA <NA> <NA> <NA> <NA> <NA>
#> 2        NA <NA> <NA> <NA> <NA> <NA>    NA     NA    g    g    g    g    g
#> 3        NA <NA>   aa   bb   cc   dd    NA     NA    g    g    g    g    g
#> 4        NA <NA>   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 5         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 6         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 7         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 8         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 9         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 10        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 11        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 12        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 13        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 14        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 15        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 16        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 17        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 18        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 19        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 20        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 21        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 22        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 23        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 24        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 25        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 26        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 27        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 28        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 29        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 30        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 31        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 32        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 33        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#>    Data1 Data2 Data3 Data4 Data5 Data6 Data7 EffortDot          file_das
#> 1     T1  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 2      g     g     g     g     g  <NA>  <NA>      TRUE airdas_sample.das
#> 3     aa    bb    cc    dd  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 4    650   100  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 5      n    10     1     0     6  <NA>  <NA>      TRUE airdas_sample.das
#> 6   <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 7      1    cc    67     6    mn  <NA>  <NA>      TRUE airdas_sample.das
#> 8      2    bb   -60     6    mn  <NA>  <NA>      TRUE airdas_sample.das
#> 9      3    cc    38     2    mn  <NA>  <NA>      TRUE airdas_sample.das
#> 10  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 11     4    aa   -50     5    mn  <NA>  <NA>      TRUE airdas_sample.das
#> 12     5    cc    49     4    bm  <NA>  <NA>      TRUE airdas_sample.das
#> 13     6    cc    42    10    gg    er  <NA>      TRUE airdas_sample.das
#> 14  <NA>  <NA>  <NA>  <NA>    80    20  <NA>     FALSE airdas_sample.das
#> 15  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 16     7    aa   -53     5    bm  <NA>  <NA>      TRUE airdas_sample.das
#> 17  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 18     r    10     1     2     6  <NA>  <NA>      TRUE airdas_sample.das
#> 19  stre aky r ed ti de     <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 20  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 21     8    dd    24     2    bm  <NA>  <NA>      TRUE airdas_sample.das
#> 22  shar k 11f t      <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 23  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 24     9    aa   -57     4    mn  <NA>  <NA>      TRUE airdas_sample.das
#> 25    10    cc    65     5    gg  <NA>  <NA>      TRUE airdas_sample.das
#> 26    aa   -20    dc     5    90     N  <NA>      TRUE airdas_sample.das
#> 27    11    bb    87     3    pp  <NA>  <NA>      TRUE airdas_sample.das
#> 28  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#> 29    12    bb    82     5    gg  <NA>  <NA>      TRUE airdas_sample.das
#> 30    13    bb    54     6    pp  <NA>  <NA>      TRUE airdas_sample.das
#> 31    14    cc    62     2    gg  <NA>  <NA>      TRUE airdas_sample.das
#> 32     n    10     2     0     6  <NA>  <NA>      TRUE airdas_sample.das
#> 33  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>      TRUE airdas_sample.das
#>    event_num line_num dist_from_prev effort_seg
#> 1          1        1      0.0000000          1
#> 2          2        2      0.0000000          1
#> 3          3        3      0.0000000          1
#> 4          4        4      0.0000000          1
#> 5          5        5      0.0000000          1
#> 6          6        6      0.1024135          1
#> 7          7        7      1.8123929          2
#> 8          8        8      0.6334366          3
#> 9          9        9      0.5183767          4
#> 10        10       10      0.1165992          4
#> 11        11       11      1.0688945          4
#> 12        12       12      0.4748831          5
#> 13        13       13      0.7483007          6
#> 14        NA       14      0.0000000          6
#> 15        14       15      0.8354611          6
#> 16        15       16      2.2317183          9
#> 17        16       17      0.7492064          9
#> 18        17       18      2.6622597         12
#> 19        18       19      0.2014632         12
#> 20        19       20      0.1439023         12
#> 21        20       21      0.7626821         13
#> 22        21       22      0.9065844         14
#> 23        22       23      1.3671953         15
#> 24        23       24      0.7051195         16
#> 25        24       25      0.7053614         17
#> 26        25       26      0.0000000         17
#> 27        26       27      0.7914588         18
#> 28        27       28      0.8066963         18
#> 29        28       29      0.1024159         18
#> 30        29       30      0.8066906         19
#> 31        31       32      1.0362457         20
#> 32        32       33      1.0650238         21
#> 33        33       34      0.1151205         21
#> 
#> [[2]]
#>  [1] 1.000000 1.000000 1.000000 1.469903 1.000000 1.000000 1.000000 1.000000
#>  [9] 1.000000 1.000000 1.000000 1.000000 1.000000 1.000000 1.000000 1.000000
#> [17] 1.000000 1.000000 1.000000 1.000000 1.000000
#> 
#> [[3]]
#> [1] 4
#>