Process aerial survey DAS data

Process AirDAS data (the output of airdas_read), including extracting state and condition information for each AirDAS event

airdas_process(x, ...)

# S3 method for character
airdas_process(x, ...)

# S3 method for data.frame
airdas_process(
  x,
  days.gap.part = 0.5/24,
  days.gap.full = 12/24,
  gap.message = FALSE,
  ...
)

Arguments

x
either a data frame (the output of airdas_read), or a character (filepath) which is first passed to airdas_read

...
ignored

days.gap.part
numeric of length 1; time gap (in days) used to identify when a 'partial reset' is performed, i.e. when propogated info (weather, observers, etc) is reset. Default is 30 minutes; must be less than or equal to days.gap.full

days.gap.full
numeric of length 1; time gap (in days) used to identify when a 'full reset; is performed, i.e. when all info (transect number and propogated info) is reset. Default is 12 hours

gap.message
logical; should messages be printed detailing which row(s) of the output data frame were partially or fully reset?

Value

Data frame; the input data frame, i.e. the output of airdas_read, with the following columns added:

State/condition	Column name	Data source

On/off effort	OnEffort	T/R and O/E events

Transect code	Trans	Event: T; Column: Data1

Beaufort sea state	Bft	Event: W; Column: Data3

Percent overcast (cloud cover)	CCover	Event: W; Column: Data2

Jellyfish code	Trans	Event: W; Column: Data4

Horizontal sun (clock system)	HorizSun	Event: W; Column: Data5

Haze/Kelp/Red tide code	HKR	Event: W; Column: Data1

Left observer	ObsL	Event: P; Column: Data1

Belly observer	ObsB	Event: P; Column: Data2

Right observer	ObsR	Event: P; Column: Data3

Data recorder	Rec	Event: P; Column: Data4

Altitude (feet)	AltFt	Event: A; Column: Data1

Speed (knots)	SpKnot	Event: A; Column: Data2

Viewing condition - left inside	VLI	Event: V; Column: Data1

Viewing condition - left outside	VLO	Event: V; Column: Data2

Viewing condition - belly	VB	Event: V; Column: Data3

Viewing condition - right inside	VRI	Event: V; Column: Data4

Viewing condition - right outside	VRO	Event: V; Column: Data5

Details

If x is a character, it is assumed to be a filepath and first passed to airdas_read. This output is then passed to airdas_process.

AirDAS data is event-based, meaning most events indicate when a state or weather condition changes. For instance, a 'W' event indicates when the Beaufort sea state changes, and the Beaufort is the same for subsequent events until the next 'W' event. For each state/condition: a new column is created, the state/condition information is extracted from relevant events, and extracted information is propagated to appropriate subsequent rows (events). Thus, each row in the output data frame contains all pertinent state/condition information for that row.

The following assumptions/decisions are made during processing:

All '#' events (deleted events) are removed

'Datetime', 'Lat', and 'Lon' information are added to '1' events

Effort is determined as follows: T/R events turns effort on, and O/E events turn effort off. T/R events themselves will be on effort, while O/E events will be off effort. The 'EffortDot' column is ignored

Missing values are NA rather than -1

Examples

y <- system.file("airdas_sample.das", package = "swfscAirDAS")
airdas_process(y)

#>    Event            DateTime      Lat       Lon OnEffort Trans Bft CCover Jelly
#> 1      T 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 2      V 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 3      P 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 4      A 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 5      W 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1   1     10     0
#> 6      * 2015-04-09 12:31:00 39.23567 -123.1575     TRUE    T1   1     10     0
#> 7      S 2015-04-09 12:31:35 39.23367 -123.1783     TRUE    T1   1     10     0
#> 8      S 2015-04-09 12:31:47 39.23350 -123.1857     TRUE    T1   1     10     0
#> 9      S 2015-04-09 12:31:56 39.23333 -123.1917     TRUE    T1   1     10     0
#> 10     * 2015-04-09 12:32:00 39.23317 -123.1930     TRUE    T1   1     10     0
#> 11     S 2015-04-09 12:32:20 39.23233 -123.2053     TRUE    T1   1     10     0
#> 12     S 2015-04-09 12:32:30 39.23233 -123.2108     TRUE    T1   1     10     0
#> 13     S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 14     1 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 15     * 2015-04-09 12:33:00 39.23267 -123.2292     TRUE    T1   1     10     0
#> 16     S 2015-04-09 12:33:43 39.23333 -123.2550     TRUE    T1   1     10     0
#> 17     * 2015-04-09 12:34:00 39.23300 -123.2637     TRUE    T1   1     10     0
#> 18     W 2015-04-09 12:34:52 39.23317 -123.2945     TRUE    T1   1     10     2
#> 19     C 2015-04-09 12:34:56 39.23317 -123.2968     TRUE    T1   1     10     2
#> 20     * 2015-04-09 12:35:00 39.23317 -123.2985     TRUE    T1   1     10     2
#> 21     S 2015-04-09 12:35:13 39.23317 -123.3073     TRUE    T1   1     10     2
#> 22     C 2015-04-09 12:35:32 39.23317 -123.3178     TRUE    T1   1     10     2
#> 23     * 2015-04-09 12:36:00 39.23333 -123.3337     TRUE    T1   1     10     2
#> 24     S 2015-04-09 12:36:13 39.23333 -123.3418     TRUE    T1   1     10     2
#> 25     S 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 26     t 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 27     S 2015-04-09 12:36:43 39.23350 -123.3592     TRUE    T1   1     10     2
#> 28     * 2015-04-09 12:37:01 39.23383 -123.3685     TRUE    T1   1     10     2
#> 29     S 2015-04-09 12:37:01 39.23400 -123.3697     TRUE    T1   1     10     2
#> 30     S 2015-04-09 12:37:16 39.23433 -123.3790     TRUE    T1   1     10     2
#> 31     S 2015-04-09 12:37:37 39.23417 -123.3910     TRUE    T1   1     10     2
#> 32     W 2015-04-09 12:37:57 39.23400 -123.4033     TRUE    T1   2     10     0
#> 33     E 2015-04-09 12:38:00 39.23400 -123.4047    FALSE    T1   2     10     0
#> 34     C 2015-04-09 12:38:00 39.23400 -123.4047    FALSE    T1   2     10     0
#> 35     * 2015-04-09 12:39:00 39.23417 -123.4400    FALSE    T1   2     10     0
#> 36     R 2015-04-09 12:40:00 39.23400 -123.4742     TRUE    T1   2     10     0
#> 37     W 2015-04-09 12:40:28 39.23333 -123.4915     TRUE    T1   2     20     0
#> 38     C 2015-04-09 12:40:36 39.23317 -123.4960     TRUE    T1   2     20     0
#> 39     * 2015-04-09 12:41:00 39.23317 -123.5092     TRUE    T1   2     20     0
#> 40     W 2015-04-09 12:41:08 39.23317 -123.5143     TRUE    T1   1     20     0
#> 41     * 2015-04-09 12:42:00 39.23267 -123.5433     TRUE    T1   1     20     0
#> 42     S 2015-04-09 12:42:13 39.23250 -123.5515     TRUE    T1   1     20     0
#> 43     * 2015-04-09 12:43:00 39.23233 -123.5778     TRUE    T1   1     20     0
#> 44     O 2015-04-09 12:43:43 39.23283 -123.6033    FALSE  <NA>   1     20     0
#> 45     * 2015-04-09 12:44:00 39.23300 -123.6118    FALSE  <NA>   1     20     0
#> 46     * 2015-04-09 12:45:00 39.24450 -123.6415    FALSE  <NA>   1     20     0
#> 47     * 2015-04-09 12:46:00 39.22167 -123.6373    FALSE  <NA>   1     20     0
#> 48     T 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 49     V 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 50     P 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 51     A 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 52     W 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 53     * 2015-04-09 12:47:00 39.21950 -123.6025     TRUE    T2   1     20     0
#> 54     * 2015-04-09 12:48:00 39.21783 -123.5682     TRUE    T2   1     20     0
#> 55     A 2015-04-09 12:48:29 39.21767 -123.5515     TRUE    T2   1     20     0
#> 56     A 2015-04-09 12:48:36 39.21783 -123.5468     TRUE    T2   1     20     0
#> 57     * 2015-04-09 12:49:00 39.21783 -123.5333     TRUE    T2   1     20     0
#> 58     * 2015-04-09 12:50:00 39.21617 -123.4987     TRUE    T2   1     20     0
#> 59     A 2015-04-09 12:50:32 39.21683 -123.4797     TRUE    T2   1     20     0
#> 60     * 2015-04-09 12:51:00 39.21733 -123.4653     TRUE    T2   1     20     0
#> 61     S 2015-04-09 12:51:40 39.21750 -123.4410     TRUE    T2   1     20     0
#> 62     S 2015-04-09 12:51:55 39.21750 -123.4317     TRUE    T2   1     20     0
#> 63     * 2015-04-09 12:52:00 39.21733 -123.4293     TRUE    T2   1     20     0
#> 64     S 2015-04-09 12:52:03 39.21733 -123.4270     TRUE    T2   1     20     0
#> 65     S 2015-04-09 12:52:04 39.21733 -123.4270     TRUE    T2   1     20     0
#> 66     S 2015-04-09 12:52:07 39.21733 -123.4258     TRUE    T2   1     20     0
#> 67     S 2015-04-09 12:52:17 39.21700 -123.4185     TRUE    T2   1     20     0
#> 68     C 2015-04-09 12:52:39 39.21700 -123.4072     TRUE    T2   1     20     0
#> 69     S 2015-04-09 12:52:54 39.21700 -123.3982     TRUE    T2   1     20     0
#> 70     * 2015-04-09 12:53:00 39.21717 -123.3960     TRUE    T2   1     20     0
#> 71     * 2015-04-09 12:54:00 39.21683 -123.3617     TRUE    T2   1     20     0
#> 72     V 2015-04-09 12:54:01 39.21683 -123.3605     TRUE    T2   1     20     0
#> 73     W 2015-04-09 12:54:04 39.21683 -123.3605     TRUE    T2   2     20     0
#> 74     S 2015-04-09 12:54:34 39.21667 -123.3422     TRUE    T2   2     20     0
#> 75     * 2015-04-09 12:55:00 39.21667 -123.3285     TRUE    T2   2     20     0
#> 76     S 2015-04-09 12:55:20 39.21700 -123.3172     TRUE    T2   2     20     0
#> 77     * 2015-04-09 12:56:00 39.21700 -123.2960     TRUE    T2   2     20     0
#> 78     W 2015-04-09 12:56:40 39.21733 -123.2730     TRUE    T2   2     20     0
#> 79     C 2015-04-09 12:56:43 39.21717 -123.2718     TRUE    T2   2     20     0
#> 80     S 2015-04-09 12:56:49 39.21717 -123.2685     TRUE    T2   2     20     0
#> 81     * 2015-04-09 12:57:00 39.21700 -123.2617     TRUE    T2   2     20     0
#> 82     O 2015-04-09 12:57:30 39.21700 -123.2417    FALSE  <NA>   2     20     0
#>    HorizSun  HKR ObsL ObsB ObsR  Rec AltFt SpKnot  VLI  VLO   VB  VRI  VRO
#> 1        NA <NA> <NA> <NA> <NA> <NA>    NA     NA <NA> <NA> <NA> <NA> <NA>
#> 2        NA <NA> <NA> <NA> <NA> <NA>    NA     NA    g    g    g    g    g
#> 3        NA <NA>   aa   bb   cc   dd    NA     NA    g    g    g    g    g
#> 4        NA <NA>   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 5         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 6         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 7         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 8         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 9         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 10        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 11        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 12        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 13        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 14        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 15        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 16        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 17        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 18        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 19        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 20        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 21        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 22        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 23        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 24        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 25        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 26        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 27        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 28        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 29        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 30        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 31        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 32        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 33        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 34        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 35        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 36        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 37        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 38        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 39        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 40        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 41        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 42        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 43        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 44        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 45        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 46        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 47        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 48        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 49        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 50        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 51        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 52       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 53       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 54       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 55       12    n   aa   bb   cc   dd   680    100    g    g    g    g    g
#> 56       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 57       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 58       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 59       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 60       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 61       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 62       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 63       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 64       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 65       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 66       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 67       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 68       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 69       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 70       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 71       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 72       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 73       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 74       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 75       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 76       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 77       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 78       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 79       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 80       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 81       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 82       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#>    Data1 Data2 Data3 Data4 Data5 Data6           Data7 EffortDot
#> 1     T1  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 2      g     g     g     g     g  <NA>            <NA>      TRUE
#> 3     aa    bb    cc    dd  <NA>  <NA>            <NA>      TRUE
#> 4    650   100  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 5      n    10     1     0     6  <NA>            <NA>      TRUE
#> 6   <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 7      1    cc    67     6    mn  <NA>            <NA>      TRUE
#> 8      2    bb   -60     6    mn  <NA>            <NA>      TRUE
#> 9      3    cc    38     2    mn  <NA>            <NA>      TRUE
#> 10  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 11     4    aa   -50     5    mn  <NA>            <NA>      TRUE
#> 12     5    cc    49     4    bm  <NA>            <NA>      TRUE
#> 13     6    cc    42    10    gg    er            <NA>      TRUE
#> 14  <NA>  <NA>  <NA>  <NA>    80    20            <NA>     FALSE
#> 15  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 16     7    aa   -53     5    bm  <NA>            <NA>      TRUE
#> 17  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 18     r    10     1     2     6  <NA>            <NA>      TRUE
#> 19  stre aky r ed ti de     <NA>  <NA>            <NA>      TRUE
#> 20  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 21     8    dd    24     2    bm  <NA>            <NA>      TRUE
#> 22  shar k 11f t      <NA>  <NA>  <NA>            <NA>      TRUE
#> 23  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 24     9    aa   -57     4    mn  <NA>            <NA>      TRUE
#> 25    10    cc    65     5    gg  <NA>            <NA>      TRUE
#> 26    aa   -20    dc     5    90     N            <NA>      TRUE
#> 27    11    bb    87     3    pp  <NA>            <NA>      TRUE
#> 28  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 29    12    bb    82     5    gg  <NA>            <NA>      TRUE
#> 30    13    bb    54     6    pp  <NA>            <NA>      TRUE
#> 31    14    cc    62     2    gg  <NA>            <NA>      TRUE
#> 32     n    10     2     0     6  <NA>            <NA>      TRUE
#> 33  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 34  off  effor t to  circl e on  unide ntifed object        TRUE
#> 35  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 36  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 37     n    20     2     0     6  <NA>            <NA>      TRUE
#> 38  wate r got  more  blue  <NA>  <NA>            <NA>      TRUE
#> 39  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 40     n    20     1     0     6  <NA>            <NA>      TRUE
#> 41  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 42    15    bb    50     3    mn  <NA>            <NA>      TRUE
#> 43  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 44  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>     FALSE
#> 45  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>     FALSE
#> 46  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>     FALSE
#> 47  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>     FALSE
#> 48    T2  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 49     g     g     g     g     g  <NA>            <NA>      TRUE
#> 50    aa    bb    cc    dd  <NA>  <NA>            <NA>      TRUE
#> 51   650   100  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 52     n    20     1     0    12  <NA>            <NA>      TRUE
#> 53  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 54  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 55   680   100  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 56   650   100  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 57  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 58  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 59   680    90  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 60  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 61    16    aa   -31     6    mn  <NA>            <NA>      TRUE
#> 62    17    bb   -59     6    bm  <NA>            <NA>      TRUE
#> 63  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 64    18    aa   -62     2    gg  <NA>            <NA>      TRUE
#> 65    19    cc    42     3    bm  <NA>            <NA>      TRUE
#> 66    20    aa   -48     4    bm  <NA>            <NA>      TRUE
#> 67    21    cc    63     6    pp  <NA>            <NA>      TRUE
#> 68  bell y has  some  glar e      <NA>            <NA>      TRUE
#> 69    22    cc    43     2    pp  <NA>            <NA>      TRUE
#> 70  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 71  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 72     g     g     p     g     g  <NA>            <NA>      TRUE
#> 73     n    20     2     0    12  <NA>            <NA>      TRUE
#> 74    23    cc    56     6    mn  <NA>            <NA>      TRUE
#> 75  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 76    24    cc    37     6    mn  <NA>            <NA>      TRUE
#> 77  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 78     r    20     2     0    12  <NA>            <NA>      TRUE
#> 79  patc hy re d tid e      <NA>  <NA>            <NA>      TRUE
#> 80    25    bb   -76     1    pp  <NA>            <NA>      TRUE
#> 81  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 82  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#>             file_das event_num line_num
#> 1  airdas_sample.das         1        1
#> 2  airdas_sample.das         2        2
#> 3  airdas_sample.das         3        3
#> 4  airdas_sample.das         4        4
#> 5  airdas_sample.das         5        5
#> 6  airdas_sample.das         6        6
#> 7  airdas_sample.das         7        7
#> 8  airdas_sample.das         8        8
#> 9  airdas_sample.das         9        9
#> 10 airdas_sample.das        10       10
#> 11 airdas_sample.das        11       11
#> 12 airdas_sample.das        12       12
#> 13 airdas_sample.das        13       13
#> 14 airdas_sample.das        NA       14
#> 15 airdas_sample.das        14       15
#> 16 airdas_sample.das        15       16
#> 17 airdas_sample.das        16       17
#> 18 airdas_sample.das        17       18
#> 19 airdas_sample.das        18       19
#> 20 airdas_sample.das        19       20
#> 21 airdas_sample.das        20       21
#> 22 airdas_sample.das        21       22
#> 23 airdas_sample.das        22       23
#> 24 airdas_sample.das        23       24
#> 25 airdas_sample.das        24       25
#> 26 airdas_sample.das        25       26
#> 27 airdas_sample.das        26       27
#> 28 airdas_sample.das        27       28
#> 29 airdas_sample.das        28       29
#> 30 airdas_sample.das        29       30
#> 31 airdas_sample.das        31       32
#> 32 airdas_sample.das        32       33
#> 33 airdas_sample.das        33       34
#> 34 airdas_sample.das        34       35
#> 35 airdas_sample.das        35       36
#> 36 airdas_sample.das        36       37
#> 37 airdas_sample.das        37       38
#> 38 airdas_sample.das        38       39
#> 39 airdas_sample.das        39       40
#> 40 airdas_sample.das        40       41
#> 41 airdas_sample.das        41       42
#> 42 airdas_sample.das        42       43
#> 43 airdas_sample.das        43       44
#> 44 airdas_sample.das        44       45
#> 45 airdas_sample.das        45       46
#> 46 airdas_sample.das        46       47
#> 47 airdas_sample.das        47       48
#> 48 airdas_sample.das        48       49
#> 49 airdas_sample.das        49       50
#> 50 airdas_sample.das        50       51
#> 51 airdas_sample.das        51       52
#> 52 airdas_sample.das        52       53
#> 53 airdas_sample.das        53       54
#> 54 airdas_sample.das        54       55
#> 55 airdas_sample.das        55       56
#> 56 airdas_sample.das        56       57
#> 57 airdas_sample.das        57       58
#> 58 airdas_sample.das        58       59
#> 59 airdas_sample.das        59       60
#> 60 airdas_sample.das        60       61
#> 61 airdas_sample.das        61       62
#> 62 airdas_sample.das        62       63
#> 63 airdas_sample.das        63       64
#> 64 airdas_sample.das        64       65
#> 65 airdas_sample.das        65       66
#> 66 airdas_sample.das        66       67
#> 67 airdas_sample.das        67       68
#> 68 airdas_sample.das        68       69
#> 69 airdas_sample.das        69       70
#> 70 airdas_sample.das        70       71
#> 71 airdas_sample.das        71       72
#> 72 airdas_sample.das        72       73
#> 73 airdas_sample.das        73       74
#> 74 airdas_sample.das        74       75
#> 75 airdas_sample.das        75       76
#> 76 airdas_sample.das        76       77
#> 77 airdas_sample.das        77       78
#> 78 airdas_sample.das        78       79
#> 79 airdas_sample.das        79       80
#> 80 airdas_sample.das        80       81
#> 81 airdas_sample.das        81       82
#> 82 airdas_sample.das        82       83

y.read <- airdas_read(y)
airdas_process(y.read)

#>    Event            DateTime      Lat       Lon OnEffort Trans Bft CCover Jelly
#> 1      T 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 2      V 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 3      P 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 4      A 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 5      W 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1   1     10     0
#> 6      * 2015-04-09 12:31:00 39.23567 -123.1575     TRUE    T1   1     10     0
#> 7      S 2015-04-09 12:31:35 39.23367 -123.1783     TRUE    T1   1     10     0
#> 8      S 2015-04-09 12:31:47 39.23350 -123.1857     TRUE    T1   1     10     0
#> 9      S 2015-04-09 12:31:56 39.23333 -123.1917     TRUE    T1   1     10     0
#> 10     * 2015-04-09 12:32:00 39.23317 -123.1930     TRUE    T1   1     10     0
#> 11     S 2015-04-09 12:32:20 39.23233 -123.2053     TRUE    T1   1     10     0
#> 12     S 2015-04-09 12:32:30 39.23233 -123.2108     TRUE    T1   1     10     0
#> 13     S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 14     1 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 15     * 2015-04-09 12:33:00 39.23267 -123.2292     TRUE    T1   1     10     0
#> 16     S 2015-04-09 12:33:43 39.23333 -123.2550     TRUE    T1   1     10     0
#> 17     * 2015-04-09 12:34:00 39.23300 -123.2637     TRUE    T1   1     10     0
#> 18     W 2015-04-09 12:34:52 39.23317 -123.2945     TRUE    T1   1     10     2
#> 19     C 2015-04-09 12:34:56 39.23317 -123.2968     TRUE    T1   1     10     2
#> 20     * 2015-04-09 12:35:00 39.23317 -123.2985     TRUE    T1   1     10     2
#> 21     S 2015-04-09 12:35:13 39.23317 -123.3073     TRUE    T1   1     10     2
#> 22     C 2015-04-09 12:35:32 39.23317 -123.3178     TRUE    T1   1     10     2
#> 23     * 2015-04-09 12:36:00 39.23333 -123.3337     TRUE    T1   1     10     2
#> 24     S 2015-04-09 12:36:13 39.23333 -123.3418     TRUE    T1   1     10     2
#> 25     S 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 26     t 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 27     S 2015-04-09 12:36:43 39.23350 -123.3592     TRUE    T1   1     10     2
#> 28     * 2015-04-09 12:37:01 39.23383 -123.3685     TRUE    T1   1     10     2
#> 29     S 2015-04-09 12:37:01 39.23400 -123.3697     TRUE    T1   1     10     2
#> 30     S 2015-04-09 12:37:16 39.23433 -123.3790     TRUE    T1   1     10     2
#> 31     S 2015-04-09 12:37:37 39.23417 -123.3910     TRUE    T1   1     10     2
#> 32     W 2015-04-09 12:37:57 39.23400 -123.4033     TRUE    T1   2     10     0
#> 33     E 2015-04-09 12:38:00 39.23400 -123.4047    FALSE    T1   2     10     0
#> 34     C 2015-04-09 12:38:00 39.23400 -123.4047    FALSE    T1   2     10     0
#> 35     * 2015-04-09 12:39:00 39.23417 -123.4400    FALSE    T1   2     10     0
#> 36     R 2015-04-09 12:40:00 39.23400 -123.4742     TRUE    T1   2     10     0
#> 37     W 2015-04-09 12:40:28 39.23333 -123.4915     TRUE    T1   2     20     0
#> 38     C 2015-04-09 12:40:36 39.23317 -123.4960     TRUE    T1   2     20     0
#> 39     * 2015-04-09 12:41:00 39.23317 -123.5092     TRUE    T1   2     20     0
#> 40     W 2015-04-09 12:41:08 39.23317 -123.5143     TRUE    T1   1     20     0
#> 41     * 2015-04-09 12:42:00 39.23267 -123.5433     TRUE    T1   1     20     0
#> 42     S 2015-04-09 12:42:13 39.23250 -123.5515     TRUE    T1   1     20     0
#> 43     * 2015-04-09 12:43:00 39.23233 -123.5778     TRUE    T1   1     20     0
#> 44     O 2015-04-09 12:43:43 39.23283 -123.6033    FALSE  <NA>   1     20     0
#> 45     * 2015-04-09 12:44:00 39.23300 -123.6118    FALSE  <NA>   1     20     0
#> 46     * 2015-04-09 12:45:00 39.24450 -123.6415    FALSE  <NA>   1     20     0
#> 47     * 2015-04-09 12:46:00 39.22167 -123.6373    FALSE  <NA>   1     20     0
#> 48     T 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 49     V 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 50     P 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 51     A 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 52     W 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 53     * 2015-04-09 12:47:00 39.21950 -123.6025     TRUE    T2   1     20     0
#> 54     * 2015-04-09 12:48:00 39.21783 -123.5682     TRUE    T2   1     20     0
#> 55     A 2015-04-09 12:48:29 39.21767 -123.5515     TRUE    T2   1     20     0
#> 56     A 2015-04-09 12:48:36 39.21783 -123.5468     TRUE    T2   1     20     0
#> 57     * 2015-04-09 12:49:00 39.21783 -123.5333     TRUE    T2   1     20     0
#> 58     * 2015-04-09 12:50:00 39.21617 -123.4987     TRUE    T2   1     20     0
#> 59     A 2015-04-09 12:50:32 39.21683 -123.4797     TRUE    T2   1     20     0
#> 60     * 2015-04-09 12:51:00 39.21733 -123.4653     TRUE    T2   1     20     0
#> 61     S 2015-04-09 12:51:40 39.21750 -123.4410     TRUE    T2   1     20     0
#> 62     S 2015-04-09 12:51:55 39.21750 -123.4317     TRUE    T2   1     20     0
#> 63     * 2015-04-09 12:52:00 39.21733 -123.4293     TRUE    T2   1     20     0
#> 64     S 2015-04-09 12:52:03 39.21733 -123.4270     TRUE    T2   1     20     0
#> 65     S 2015-04-09 12:52:04 39.21733 -123.4270     TRUE    T2   1     20     0
#> 66     S 2015-04-09 12:52:07 39.21733 -123.4258     TRUE    T2   1     20     0
#> 67     S 2015-04-09 12:52:17 39.21700 -123.4185     TRUE    T2   1     20     0
#> 68     C 2015-04-09 12:52:39 39.21700 -123.4072     TRUE    T2   1     20     0
#> 69     S 2015-04-09 12:52:54 39.21700 -123.3982     TRUE    T2   1     20     0
#> 70     * 2015-04-09 12:53:00 39.21717 -123.3960     TRUE    T2   1     20     0
#> 71     * 2015-04-09 12:54:00 39.21683 -123.3617     TRUE    T2   1     20     0
#> 72     V 2015-04-09 12:54:01 39.21683 -123.3605     TRUE    T2   1     20     0
#> 73     W 2015-04-09 12:54:04 39.21683 -123.3605     TRUE    T2   2     20     0
#> 74     S 2015-04-09 12:54:34 39.21667 -123.3422     TRUE    T2   2     20     0
#> 75     * 2015-04-09 12:55:00 39.21667 -123.3285     TRUE    T2   2     20     0
#> 76     S 2015-04-09 12:55:20 39.21700 -123.3172     TRUE    T2   2     20     0
#> 77     * 2015-04-09 12:56:00 39.21700 -123.2960     TRUE    T2   2     20     0
#> 78     W 2015-04-09 12:56:40 39.21733 -123.2730     TRUE    T2   2     20     0
#> 79     C 2015-04-09 12:56:43 39.21717 -123.2718     TRUE    T2   2     20     0
#> 80     S 2015-04-09 12:56:49 39.21717 -123.2685     TRUE    T2   2     20     0
#> 81     * 2015-04-09 12:57:00 39.21700 -123.2617     TRUE    T2   2     20     0
#> 82     O 2015-04-09 12:57:30 39.21700 -123.2417    FALSE  <NA>   2     20     0
#>    HorizSun  HKR ObsL ObsB ObsR  Rec AltFt SpKnot  VLI  VLO   VB  VRI  VRO
#> 1        NA <NA> <NA> <NA> <NA> <NA>    NA     NA <NA> <NA> <NA> <NA> <NA>
#> 2        NA <NA> <NA> <NA> <NA> <NA>    NA     NA    g    g    g    g    g
#> 3        NA <NA>   aa   bb   cc   dd    NA     NA    g    g    g    g    g
#> 4        NA <NA>   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 5         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 6         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 7         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 8         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 9         6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 10        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 11        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 12        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 13        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 14        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 15        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 16        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 17        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 18        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 19        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 20        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 21        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 22        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 23        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 24        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 25        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 26        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 27        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 28        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 29        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 30        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 31        6    r   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 32        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 33        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 34        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 35        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 36        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 37        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 38        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 39        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 40        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 41        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 42        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 43        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 44        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 45        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 46        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 47        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 48        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 49        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 50        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 51        6    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 52       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 53       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 54       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 55       12    n   aa   bb   cc   dd   680    100    g    g    g    g    g
#> 56       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 57       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 58       12    n   aa   bb   cc   dd   650    100    g    g    g    g    g
#> 59       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 60       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 61       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 62       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 63       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 64       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 65       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 66       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 67       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 68       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 69       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 70       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 71       12    n   aa   bb   cc   dd   680     90    g    g    g    g    g
#> 72       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 73       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 74       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 75       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 76       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 77       12    n   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 78       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 79       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 80       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 81       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#> 82       12    r   aa   bb   cc   dd   680     90    g    g    p    g    g
#>    Data1 Data2 Data3 Data4 Data5 Data6           Data7 EffortDot
#> 1     T1  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 2      g     g     g     g     g  <NA>            <NA>      TRUE
#> 3     aa    bb    cc    dd  <NA>  <NA>            <NA>      TRUE
#> 4    650   100  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 5      n    10     1     0     6  <NA>            <NA>      TRUE
#> 6   <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 7      1    cc    67     6    mn  <NA>            <NA>      TRUE
#> 8      2    bb   -60     6    mn  <NA>            <NA>      TRUE
#> 9      3    cc    38     2    mn  <NA>            <NA>      TRUE
#> 10  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 11     4    aa   -50     5    mn  <NA>            <NA>      TRUE
#> 12     5    cc    49     4    bm  <NA>            <NA>      TRUE
#> 13     6    cc    42    10    gg    er            <NA>      TRUE
#> 14  <NA>  <NA>  <NA>  <NA>    80    20            <NA>     FALSE
#> 15  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 16     7    aa   -53     5    bm  <NA>            <NA>      TRUE
#> 17  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 18     r    10     1     2     6  <NA>            <NA>      TRUE
#> 19  stre aky r ed ti de     <NA>  <NA>            <NA>      TRUE
#> 20  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 21     8    dd    24     2    bm  <NA>            <NA>      TRUE
#> 22  shar k 11f t      <NA>  <NA>  <NA>            <NA>      TRUE
#> 23  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 24     9    aa   -57     4    mn  <NA>            <NA>      TRUE
#> 25    10    cc    65     5    gg  <NA>            <NA>      TRUE
#> 26    aa   -20    dc     5    90     N            <NA>      TRUE
#> 27    11    bb    87     3    pp  <NA>            <NA>      TRUE
#> 28  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 29    12    bb    82     5    gg  <NA>            <NA>      TRUE
#> 30    13    bb    54     6    pp  <NA>            <NA>      TRUE
#> 31    14    cc    62     2    gg  <NA>            <NA>      TRUE
#> 32     n    10     2     0     6  <NA>            <NA>      TRUE
#> 33  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 34  off  effor t to  circl e on  unide ntifed object        TRUE
#> 35  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 36  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 37     n    20     2     0     6  <NA>            <NA>      TRUE
#> 38  wate r got  more  blue  <NA>  <NA>            <NA>      TRUE
#> 39  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 40     n    20     1     0     6  <NA>            <NA>      TRUE
#> 41  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 42    15    bb    50     3    mn  <NA>            <NA>      TRUE
#> 43  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 44  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>     FALSE
#> 45  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>     FALSE
#> 46  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>     FALSE
#> 47  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>     FALSE
#> 48    T2  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 49     g     g     g     g     g  <NA>            <NA>      TRUE
#> 50    aa    bb    cc    dd  <NA>  <NA>            <NA>      TRUE
#> 51   650   100  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 52     n    20     1     0    12  <NA>            <NA>      TRUE
#> 53  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 54  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 55   680   100  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 56   650   100  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 57  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 58  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 59   680    90  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 60  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 61    16    aa   -31     6    mn  <NA>            <NA>      TRUE
#> 62    17    bb   -59     6    bm  <NA>            <NA>      TRUE
#> 63  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 64    18    aa   -62     2    gg  <NA>            <NA>      TRUE
#> 65    19    cc    42     3    bm  <NA>            <NA>      TRUE
#> 66    20    aa   -48     4    bm  <NA>            <NA>      TRUE
#> 67    21    cc    63     6    pp  <NA>            <NA>      TRUE
#> 68  bell y has  some  glar e      <NA>            <NA>      TRUE
#> 69    22    cc    43     2    pp  <NA>            <NA>      TRUE
#> 70  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 71  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 72     g     g     p     g     g  <NA>            <NA>      TRUE
#> 73     n    20     2     0    12  <NA>            <NA>      TRUE
#> 74    23    cc    56     6    mn  <NA>            <NA>      TRUE
#> 75  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 76    24    cc    37     6    mn  <NA>            <NA>      TRUE
#> 77  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 78     r    20     2     0    12  <NA>            <NA>      TRUE
#> 79  patc hy re d tid e      <NA>  <NA>            <NA>      TRUE
#> 80    25    bb   -76     1    pp  <NA>            <NA>      TRUE
#> 81  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#> 82  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>      TRUE
#>             file_das event_num line_num
#> 1  airdas_sample.das         1        1
#> 2  airdas_sample.das         2        2
#> 3  airdas_sample.das         3        3
#> 4  airdas_sample.das         4        4
#> 5  airdas_sample.das         5        5
#> 6  airdas_sample.das         6        6
#> 7  airdas_sample.das         7        7
#> 8  airdas_sample.das         8        8
#> 9  airdas_sample.das         9        9
#> 10 airdas_sample.das        10       10
#> 11 airdas_sample.das        11       11
#> 12 airdas_sample.das        12       12
#> 13 airdas_sample.das        13       13
#> 14 airdas_sample.das        NA       14
#> 15 airdas_sample.das        14       15
#> 16 airdas_sample.das        15       16
#> 17 airdas_sample.das        16       17
#> 18 airdas_sample.das        17       18
#> 19 airdas_sample.das        18       19
#> 20 airdas_sample.das        19       20
#> 21 airdas_sample.das        20       21
#> 22 airdas_sample.das        21       22
#> 23 airdas_sample.das        22       23
#> 24 airdas_sample.das        23       24
#> 25 airdas_sample.das        24       25
#> 26 airdas_sample.das        25       26
#> 27 airdas_sample.das        26       27
#> 28 airdas_sample.das        27       28
#> 29 airdas_sample.das        28       29
#> 30 airdas_sample.das        29       30
#> 31 airdas_sample.das        31       32
#> 32 airdas_sample.das        32       33
#> 33 airdas_sample.das        33       34
#> 34 airdas_sample.das        34       35
#> 35 airdas_sample.das        35       36
#> 36 airdas_sample.das        36       37
#> 37 airdas_sample.das        37       38
#> 38 airdas_sample.das        38       39
#> 39 airdas_sample.das        39       40
#> 40 airdas_sample.das        40       41
#> 41 airdas_sample.das        41       42
#> 42 airdas_sample.das        42       43
#> 43 airdas_sample.das        43       44
#> 44 airdas_sample.das        44       45
#> 45 airdas_sample.das        45       46
#> 46 airdas_sample.das        46       47
#> 47 airdas_sample.das        47       48
#> 48 airdas_sample.das        48       49
#> 49 airdas_sample.das        49       50
#> 50 airdas_sample.das        50       51
#> 51 airdas_sample.das        51       52
#> 52 airdas_sample.das        52       53
#> 53 airdas_sample.das        53       54
#> 54 airdas_sample.das        54       55
#> 55 airdas_sample.das        55       56
#> 56 airdas_sample.das        56       57
#> 57 airdas_sample.das        57       58
#> 58 airdas_sample.das        58       59
#> 59 airdas_sample.das        59       60
#> 60 airdas_sample.das        60       61
#> 61 airdas_sample.das        61       62
#> 62 airdas_sample.das        62       63
#> 63 airdas_sample.das        63       64
#> 64 airdas_sample.das        64       65
#> 65 airdas_sample.das        65       66
#> 66 airdas_sample.das        66       67
#> 67 airdas_sample.das        67       68
#> 68 airdas_sample.das        68       69
#> 69 airdas_sample.das        69       70
#> 70 airdas_sample.das        70       71
#> 71 airdas_sample.das        71       72
#> 72 airdas_sample.das        72       73
#> 73 airdas_sample.das        73       74
#> 74 airdas_sample.das        74       75
#> 75 airdas_sample.das        75       76
#> 76 airdas_sample.das        76       77
#> 77 airdas_sample.das        77       78
#> 78 airdas_sample.das        78       79
#> 79 airdas_sample.das        79       80
#> 80 airdas_sample.das        80       81
#> 81 airdas_sample.das        81       82
#> 82 airdas_sample.das        82       83