Read AirDAS file

Read one (or more - todo) fixed-width aerial survey DAS text file(s) generated by turtle4d (or an associated program) into a data frame, where each line is data for a specific event

airdas_read(file, tz = "UTC")

Arguments

file
filename of an aerial survey DAS file

tz
character; see strptime. Default is UTC

Value

Data frame with AirDAS data read into columns. The data are read into the data frame as characters as described in 'Details', with the following exceptions:

Name	Class	Details

EffortDot	logical	TRUE if "." was present, and FALSE otherwise

DateTime	POSIXct	combination of 'Date' and 'Time' columns with time zone tz

Lat	numeric	'Latitude' column converted to decimal degrees in range [-90, 90]

Lon	numeric	'Longitude' column converted to decimal degrees in range [-180, 180]

Data#	character	leading/trailing whitespace trimmed for non-comment events (i.e. rows where 'Event' is not "C" or "c")

file_das	character	filename, extracted from the file argument using basename

event_num	integer	'Event number' coerced to an integer using as.integer

line_num	integer	line number of each data row

Details

Reads/parses aerial survey DAS data into columns. The provided AirDAS file must adhere to the following column number and format specifications: (see the "AirDAS_format" vignette for more information):

Item	Columns	Format

Event number	1-3	Event

4	Effort dot	5

Time	6-11	HHMMSS

Date	13-18	MMDDYY

Latitude	20-28	NDD:MM.MM

Longitude	30-39	WDDD:MM.MM

Data1	40-44	Data2

45-49	Data3	50-54

Data4	55-59	Data5

60-64	Data6	65-69

Data7	70+

See the "AirDAS_format" vignette for more information

Examples

airdas_read(system.file("airdas_sample.das", package = "swfscAirDAS"))

#>    Event EffortDot            DateTime      Lat       Lon Data1 Data2 Data3
#> 1      T      TRUE 2015-04-09 12:30:59 39.23550 -123.1563    T1  <NA>  <NA>
#> 2      V      TRUE 2015-04-09 12:30:59 39.23550 -123.1563     g     g     g
#> 3      P      TRUE 2015-04-09 12:30:59 39.23550 -123.1563    aa    bb    cc
#> 4      A      TRUE 2015-04-09 12:30:59 39.23550 -123.1563   650   100  <NA>
#> 5      W      TRUE 2015-04-09 12:30:59 39.23550 -123.1563     n    10     1
#> 6      *      TRUE 2015-04-09 12:31:00 39.23567 -123.1575  <NA>  <NA>  <NA>
#> 7      S      TRUE 2015-04-09 12:31:35 39.23367 -123.1783     1    cc    67
#> 8      S      TRUE 2015-04-09 12:31:47 39.23350 -123.1857     2    bb   -60
#> 9      S      TRUE 2015-04-09 12:31:56 39.23333 -123.1917     3    cc    38
#> 10     *      TRUE 2015-04-09 12:32:00 39.23317 -123.1930  <NA>  <NA>  <NA>
#> 11     S      TRUE 2015-04-09 12:32:20 39.23233 -123.2053     4    aa   -50
#> 12     S      TRUE 2015-04-09 12:32:30 39.23233 -123.2108     5    cc    49
#> 13     S      TRUE 2015-04-09 12:32:42 39.23233 -123.2195     6    cc    42
#> 14     1     FALSE                <NA>       NA        NA  <NA>  <NA>  <NA>
#> 15     *      TRUE 2015-04-09 12:33:00 39.23267 -123.2292  <NA>  <NA>  <NA>
#> 16     S      TRUE 2015-04-09 12:33:43 39.23333 -123.2550     7    aa   -53
#> 17     *      TRUE 2015-04-09 12:34:00 39.23300 -123.2637  <NA>  <NA>  <NA>
#> 18     W      TRUE 2015-04-09 12:34:52 39.23317 -123.2945     r    10     1
#> 19     C      TRUE 2015-04-09 12:34:56 39.23317 -123.2968  stre aky r ed ti
#> 20     *      TRUE 2015-04-09 12:35:00 39.23317 -123.2985  <NA>  <NA>  <NA>
#> 21     S      TRUE 2015-04-09 12:35:13 39.23317 -123.3073     8    dd    24
#> 22     C      TRUE 2015-04-09 12:35:32 39.23317 -123.3178  shar k 11f t    
#> 23     *      TRUE 2015-04-09 12:36:00 39.23333 -123.3337  <NA>  <NA>  <NA>
#> 24     S      TRUE 2015-04-09 12:36:13 39.23333 -123.3418     9    aa   -57
#> 25     S      TRUE 2015-04-09 12:36:26 39.23350 -123.3500    10    cc    65
#> 26     t      TRUE 2015-04-09 12:36:26 39.23350 -123.3500    aa   -20    dc
#> 27     S      TRUE 2015-04-09 12:36:43 39.23350 -123.3592    11    bb    87
#> 28     *      TRUE 2015-04-09 12:37:01 39.23383 -123.3685  <NA>  <NA>  <NA>
#> 29     S      TRUE 2015-04-09 12:37:01 39.23400 -123.3697    12    bb    82
#> 30     S      TRUE 2015-04-09 12:37:16 39.23433 -123.3790    13    bb    54
#> 31     #      TRUE 2015-04-09 12:37:17 39.23433 -123.3790  <NA>  <NA>  <NA>
#> 32     S      TRUE 2015-04-09 12:37:37 39.23417 -123.3910    14    cc    62
#> 33     W      TRUE 2015-04-09 12:37:57 39.23400 -123.4033     n    10     2
#> 34     E      TRUE 2015-04-09 12:38:00 39.23400 -123.4047  <NA>  <NA>  <NA>
#> 35     C      TRUE 2015-04-09 12:38:00 39.23400 -123.4047  off  effor t to 
#> 36     *      TRUE 2015-04-09 12:39:00 39.23417 -123.4400  <NA>  <NA>  <NA>
#> 37     R      TRUE 2015-04-09 12:40:00 39.23400 -123.4742  <NA>  <NA>  <NA>
#> 38     W      TRUE 2015-04-09 12:40:28 39.23333 -123.4915     n    20     2
#> 39     C      TRUE 2015-04-09 12:40:36 39.23317 -123.4960  wate r got  more
#> 40     *      TRUE 2015-04-09 12:41:00 39.23317 -123.5092  <NA>  <NA>  <NA>
#> 41     W      TRUE 2015-04-09 12:41:08 39.23317 -123.5143     n    20     1
#> 42     *      TRUE 2015-04-09 12:42:00 39.23267 -123.5433  <NA>  <NA>  <NA>
#> 43     S      TRUE 2015-04-09 12:42:13 39.23250 -123.5515    15    bb    50
#> 44     *      TRUE 2015-04-09 12:43:00 39.23233 -123.5778  <NA>  <NA>  <NA>
#> 45     O     FALSE 2015-04-09 12:43:43 39.23283 -123.6033  <NA>  <NA>  <NA>
#> 46     *     FALSE 2015-04-09 12:44:00 39.23300 -123.6118  <NA>  <NA>  <NA>
#> 47     *     FALSE 2015-04-09 12:45:00 39.24450 -123.6415  <NA>  <NA>  <NA>
#> 48     *     FALSE 2015-04-09 12:46:00 39.22167 -123.6373  <NA>  <NA>  <NA>
#> 49     T      TRUE 2015-04-09 12:46:51 39.22000 -123.6078    T2  <NA>  <NA>
#> 50     V      TRUE 2015-04-09 12:46:51 39.22000 -123.6078     g     g     g
#> 51     P      TRUE 2015-04-09 12:46:51 39.22000 -123.6078    aa    bb    cc
#> 52     A      TRUE 2015-04-09 12:46:51 39.22000 -123.6078   650   100  <NA>
#> 53     W      TRUE 2015-04-09 12:46:51 39.22000 -123.6078     n    20     1
#> 54     *      TRUE 2015-04-09 12:47:00 39.21950 -123.6025  <NA>  <NA>  <NA>
#> 55     *      TRUE 2015-04-09 12:48:00 39.21783 -123.5682  <NA>  <NA>  <NA>
#> 56     A      TRUE 2015-04-09 12:48:29 39.21767 -123.5515   680   100  <NA>
#> 57     A      TRUE 2015-04-09 12:48:36 39.21783 -123.5468   650   100  <NA>
#> 58     *      TRUE 2015-04-09 12:49:00 39.21783 -123.5333  <NA>  <NA>  <NA>
#> 59     *      TRUE 2015-04-09 12:50:00 39.21617 -123.4987  <NA>  <NA>  <NA>
#> 60     A      TRUE 2015-04-09 12:50:32 39.21683 -123.4797   680    90  <NA>
#> 61     *      TRUE 2015-04-09 12:51:00 39.21733 -123.4653  <NA>  <NA>  <NA>
#> 62     S      TRUE 2015-04-09 12:51:40 39.21750 -123.4410    16    aa   -31
#> 63     S      TRUE 2015-04-09 12:51:55 39.21750 -123.4317    17    bb   -59
#> 64     *      TRUE 2015-04-09 12:52:00 39.21733 -123.4293  <NA>  <NA>  <NA>
#> 65     S      TRUE 2015-04-09 12:52:03 39.21733 -123.4270    18    aa   -62
#> 66     S      TRUE 2015-04-09 12:52:04 39.21733 -123.4270    19    cc    42
#> 67     S      TRUE 2015-04-09 12:52:07 39.21733 -123.4258    20    aa   -48
#> 68     S      TRUE 2015-04-09 12:52:17 39.21700 -123.4185    21    cc    63
#> 69     C      TRUE 2015-04-09 12:52:39 39.21700 -123.4072  bell y has  some
#> 70     S      TRUE 2015-04-09 12:52:54 39.21700 -123.3982    22    cc    43
#> 71     *      TRUE 2015-04-09 12:53:00 39.21717 -123.3960  <NA>  <NA>  <NA>
#> 72     *      TRUE 2015-04-09 12:54:00 39.21683 -123.3617  <NA>  <NA>  <NA>
#> 73     V      TRUE 2015-04-09 12:54:01 39.21683 -123.3605     g     g     p
#> 74     W      TRUE 2015-04-09 12:54:04 39.21683 -123.3605     n    20     2
#> 75     S      TRUE 2015-04-09 12:54:34 39.21667 -123.3422    23    cc    56
#> 76     *      TRUE 2015-04-09 12:55:00 39.21667 -123.3285  <NA>  <NA>  <NA>
#> 77     S      TRUE 2015-04-09 12:55:20 39.21700 -123.3172    24    cc    37
#> 78     *      TRUE 2015-04-09 12:56:00 39.21700 -123.2960  <NA>  <NA>  <NA>
#> 79     W      TRUE 2015-04-09 12:56:40 39.21733 -123.2730     r    20     2
#> 80     C      TRUE 2015-04-09 12:56:43 39.21717 -123.2718  patc hy re d tid
#> 81     S      TRUE 2015-04-09 12:56:49 39.21717 -123.2685    25    bb   -76
#> 82     *      TRUE 2015-04-09 12:57:00 39.21700 -123.2617  <NA>  <NA>  <NA>
#> 83     O      TRUE 2015-04-09 12:57:30 39.21700 -123.2417  <NA>  <NA>  <NA>
#>    Data4 Data5 Data6           Data7          file_das event_num line_num
#> 1   <NA>  <NA>  <NA>            <NA> airdas_sample.das         1        1
#> 2      g     g  <NA>            <NA> airdas_sample.das         2        2
#> 3     dd  <NA>  <NA>            <NA> airdas_sample.das         3        3
#> 4   <NA>  <NA>  <NA>            <NA> airdas_sample.das         4        4
#> 5      0     6  <NA>            <NA> airdas_sample.das         5        5
#> 6   <NA>  <NA>  <NA>            <NA> airdas_sample.das         6        6
#> 7      6    mn  <NA>            <NA> airdas_sample.das         7        7
#> 8      6    mn  <NA>            <NA> airdas_sample.das         8        8
#> 9      2    mn  <NA>            <NA> airdas_sample.das         9        9
#> 10  <NA>  <NA>  <NA>            <NA> airdas_sample.das        10       10
#> 11     5    mn  <NA>            <NA> airdas_sample.das        11       11
#> 12     4    bm  <NA>            <NA> airdas_sample.das        12       12
#> 13    10    gg    er            <NA> airdas_sample.das        13       13
#> 14  <NA>    80    20            <NA> airdas_sample.das        NA       14
#> 15  <NA>  <NA>  <NA>            <NA> airdas_sample.das        14       15
#> 16     5    bm  <NA>            <NA> airdas_sample.das        15       16
#> 17  <NA>  <NA>  <NA>            <NA> airdas_sample.das        16       17
#> 18     2     6  <NA>            <NA> airdas_sample.das        17       18
#> 19 de     <NA>  <NA>            <NA> airdas_sample.das        18       19
#> 20  <NA>  <NA>  <NA>            <NA> airdas_sample.das        19       20
#> 21     2    bm  <NA>            <NA> airdas_sample.das        20       21
#> 22  <NA>  <NA>  <NA>            <NA> airdas_sample.das        21       22
#> 23  <NA>  <NA>  <NA>            <NA> airdas_sample.das        22       23
#> 24     4    mn  <NA>            <NA> airdas_sample.das        23       24
#> 25     5    gg  <NA>            <NA> airdas_sample.das        24       25
#> 26     5    90     N            <NA> airdas_sample.das        25       26
#> 27     3    pp  <NA>            <NA> airdas_sample.das        26       27
#> 28  <NA>  <NA>  <NA>            <NA> airdas_sample.das        27       28
#> 29     5    gg  <NA>            <NA> airdas_sample.das        28       29
#> 30     6    pp  <NA>            <NA> airdas_sample.das        29       30
#> 31  <NA>  <NA>  <NA>            <NA> airdas_sample.das        30       31
#> 32     2    gg  <NA>            <NA> airdas_sample.das        31       32
#> 33     0     6  <NA>            <NA> airdas_sample.das        32       33
#> 34  <NA>  <NA>  <NA>            <NA> airdas_sample.das        33       34
#> 35 circl e on  unide ntifed object   airdas_sample.das        34       35
#> 36  <NA>  <NA>  <NA>            <NA> airdas_sample.das        35       36
#> 37  <NA>  <NA>  <NA>            <NA> airdas_sample.das        36       37
#> 38     0     6  <NA>            <NA> airdas_sample.das        37       38
#> 39  blue  <NA>  <NA>            <NA> airdas_sample.das        38       39
#> 40  <NA>  <NA>  <NA>            <NA> airdas_sample.das        39       40
#> 41     0     6  <NA>            <NA> airdas_sample.das        40       41
#> 42  <NA>  <NA>  <NA>            <NA> airdas_sample.das        41       42
#> 43     3    mn  <NA>            <NA> airdas_sample.das        42       43
#> 44  <NA>  <NA>  <NA>            <NA> airdas_sample.das        43       44
#> 45  <NA>  <NA>  <NA>            <NA> airdas_sample.das        44       45
#> 46  <NA>  <NA>  <NA>            <NA> airdas_sample.das        45       46
#> 47  <NA>  <NA>  <NA>            <NA> airdas_sample.das        46       47
#> 48  <NA>  <NA>  <NA>            <NA> airdas_sample.das        47       48
#> 49  <NA>  <NA>  <NA>            <NA> airdas_sample.das        48       49
#> 50     g     g  <NA>            <NA> airdas_sample.das        49       50
#> 51    dd  <NA>  <NA>            <NA> airdas_sample.das        50       51
#> 52  <NA>  <NA>  <NA>            <NA> airdas_sample.das        51       52
#> 53     0    12  <NA>            <NA> airdas_sample.das        52       53
#> 54  <NA>  <NA>  <NA>            <NA> airdas_sample.das        53       54
#> 55  <NA>  <NA>  <NA>            <NA> airdas_sample.das        54       55
#> 56  <NA>  <NA>  <NA>            <NA> airdas_sample.das        55       56
#> 57  <NA>  <NA>  <NA>            <NA> airdas_sample.das        56       57
#> 58  <NA>  <NA>  <NA>            <NA> airdas_sample.das        57       58
#> 59  <NA>  <NA>  <NA>            <NA> airdas_sample.das        58       59
#> 60  <NA>  <NA>  <NA>            <NA> airdas_sample.das        59       60
#> 61  <NA>  <NA>  <NA>            <NA> airdas_sample.das        60       61
#> 62     6    mn  <NA>            <NA> airdas_sample.das        61       62
#> 63     6    bm  <NA>            <NA> airdas_sample.das        62       63
#> 64  <NA>  <NA>  <NA>            <NA> airdas_sample.das        63       64
#> 65     2    gg  <NA>            <NA> airdas_sample.das        64       65
#> 66     3    bm  <NA>            <NA> airdas_sample.das        65       66
#> 67     4    bm  <NA>            <NA> airdas_sample.das        66       67
#> 68     6    pp  <NA>            <NA> airdas_sample.das        67       68
#> 69  glar e      <NA>            <NA> airdas_sample.das        68       69
#> 70     2    pp  <NA>            <NA> airdas_sample.das        69       70
#> 71  <NA>  <NA>  <NA>            <NA> airdas_sample.das        70       71
#> 72  <NA>  <NA>  <NA>            <NA> airdas_sample.das        71       72
#> 73     g     g  <NA>            <NA> airdas_sample.das        72       73
#> 74     0    12  <NA>            <NA> airdas_sample.das        73       74
#> 75     6    mn  <NA>            <NA> airdas_sample.das        74       75
#> 76  <NA>  <NA>  <NA>            <NA> airdas_sample.das        75       76
#> 77     6    mn  <NA>            <NA> airdas_sample.das        76       77
#> 78  <NA>  <NA>  <NA>            <NA> airdas_sample.das        77       78
#> 79     0    12  <NA>            <NA> airdas_sample.das        78       79
#> 80 e      <NA>  <NA>            <NA> airdas_sample.das        79       80
#> 81     1    pp  <NA>            <NA> airdas_sample.das        80       81
#> 82  <NA>  <NA>  <NA>            <NA> airdas_sample.das        81       82
#> 83  <NA>  <NA>  <NA>            <NA> airdas_sample.das        82       83