Check AirDAS file

Check that AirDAS file has accepted formatting and values

airdas_check(
  file,
  file.type = "turtle",
  skip = 0,
  file.out = NULL,
  sp.codes = NULL,
  print.transect = TRUE
)

Arguments

file
filename(s) of one or more AirDAS files

file.type
character; indicates the program used to create file. Must be one of: "turtle", "caretta", "survey", or "phocoena" (case sensitive). Default is "turtle". Passed to airdas_read

skip
integer: see read_fwf. Default is 0. Passed to airdas_read

file.out
character; filename to which to write the error log. Should be a text or CSV file. Default is NULL

sp.codes
character; filename of .dat file from which to read accepted species codes. If NULL, default (internal) file will be used. Default is NULL

print.transect
logical; indicates if a table with all the transect numbers in the x should be printed using table. Default is TRUE

Value

A data frame with five columns that list information about errors found in the AirDAS files: the file name, line number, index (row number) from the airdas_read(file) data frame, 'ID' (pre-Data# columns from the DAS file), and description of the issue. This data frame is sorted by the 'Description' column. If there are multiple issues with the same line, the issue descriptions are concatenated together using paste(..., collapse = "; ")

If print.transect is TRUE, then the output of table(x$Data1[x$Event == "T"], useNA = "always"), where x is the output of airdas_read(file, ...) is printed

If file.out is not NULL, then the error log is also written to the file (e.g., a .txt or .csv file) specified by file.out

Details

The default (internal) sp.codes file is located at system.file("SpCodesAirDAS.dat", package = "swfscAirDAS").

This function checks that the following is true:

Event codes are one of the following: #, *, 1, A, C, E, O, P, R, s, S, t, T, V, W

The effort dot matches effort determined using T, R, O, and E events

There is an O event between each T event, and vice versa, i.e. the data does not start a new transect while still on a transect and does not end a transect when already not on a transect

A T or R event does not occur while already on effort

An E event does not occur while already off effort

When the file ends, the data must be off effort and not still on a transect (i.e. an O event must occurred more recently than a T event)

All Data# columns for non-C events are right-justified

NAs are allowed for all data values when off effort except: altitude, speed, species, and group size

The following events have NA (blank) Data# columns: *, R, E, O

For PHOCOENA data, all non-C events have NA (blank) Data6 and Data7 columns

Event/column pairs meet the following requirements:

Item	Requirement

Viewing conditions	Must be one of: e, g, p, o, or NA (blank). Not case sensitive

Altitude	Can be converted to a numeric value, and is not NA

Speed	Can be converted to a numeric value, and is not NA

HKR	Characters must consist of: n, h, k, r, or NA (blank). Not case sensitive; y is also accepted for PHOCOENA data

Percent overcast	Must be a whole number between 0 and 100 (leading zeros are ok)

Beaufort	Must be a whole number between 0 and 9

Jellyfish	Must be one of 0, 1, 2, or 3

Horizontal sun	Must be one of 0:12 (leading zeros are ok)

Vertical sun	Must be one of 0:4 (leading zeros are ok)

Observers	Each entry must be two characters, and no observer code can be used twice in the same P event

Sighting (mammal)	Angle must be a whole number between -90 and 90 (leading zeros are ok)

Sighting (mammal)	Group size must be a whole number between 1 and 5000 (leading zeros are ok)

Sighting (mammal)	Species codes must be specified in sp.codes, and the first must not be NA

Sighting (mammal)	Observer code must be exactly two characters, and one of the current observers as specified by the most recent P event

Sighting (mammal)	Angle must be negative if sighting made by left observer, and positive if made by right observer

Sighting info	Species percentages can be converted to numeric values, and sum to 100

Sighting info	Unused columns (DateTime, Lat, Lon, and Data1-4) of a '1' event must be NA (blank)

Sighting info	Every S code with multiple species has a 1 code immediately after it

Sighting info	For every 1 code with n non-NA sighting percentages, the event before is an S event with n non-NA species codes

Resight	Angle can be converted to a numeric value (leading zeros are ok)

Resight	Unused resight columns (Data3-7) must be NA (blank)

Turtle sighting	Angle must be a whole number between -90 and 90 (leading zeros are ok)

Turtle sighting	Group size must be a whole number between 1 and 10 (only included in CARETTA data; leading zeros are ok))

Turtle sighting	Species code must be not NA and specified in sp.codes

Turtle sighting	Observer code must be exactly two characters

Turtle sighting	Turtle size must be a whole number between 1 and 9 (leading zeros are ok); one of s, m, l is also accepted for CARETTA data

Turtle sighting	Travel direction must be a whole number between 0 and 360 (only included in TURTLE data; leading zeros are ok))

Turtle sighting	Tail visible must be one of y, n, u, or NA (blank). Case sensitive

Turtle sighting	In TURTLE data, the Data7 column must be NA (blank)

Outstanding questions/todo:

Check for valid fish ball/mola/jelly/crab pot codes in comments?

Examples

if (FALSE) {
y <- system.file("airdas_sample.das", package = "swfscAirDAS")
airdas_check(y)
}