Extract comments from AirDAS data

Extract comments from airdas_dfr or airdas_df object

airdas_comments(x)

# S3 method for data.frame
airdas_comments(x)

# S3 method for airdas_df
airdas_comments(x)

# S3 method for airdas_dfr
airdas_comments(x)

Arguments

x
airdas_dfr or airdas_df object, or a data frame that can be coerced to a airdas_dfr object

Value

A data frame with the comment strings in column comment_str, as well as the following columns from x: file_das, line_num, and DateTime

Details

This function recreates the comment strings by pasting the Data# columns back together for the C events (comments)

See the examples section for how to search for comments with the phrase "record" to determine what extra information (e.g. molas) was being recorded vs ignored.

Examples

y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.read <- airdas_read(y, file.type = "turtle")

airdas_comments(y.read)

#>             file_das line_num            DateTime
#> 1  airdas_sample.das        1 2015-04-09 12:30:59
#> 2  airdas_sample.das        2 2015-04-09 12:30:59
#> 11 airdas_sample.das       11 2015-04-09 12:31:48
#> 22 airdas_sample.das       22 2015-04-09 12:34:56
#> 25 airdas_sample.das       25 2015-04-09 12:35:32
#> 38 airdas_sample.das       38 2015-04-09 12:38:00
#> 42 airdas_sample.das       42 2015-04-09 12:40:36
#> 46 airdas_sample.das       46 2015-04-09 12:42:01
#> 73 airdas_sample.das       73 2015-04-09 12:52:39
#> 84 airdas_sample.das       84 2015-04-09 12:56:43
#>                                      comment_str
#> 1                                 Recorder: dd  
#> 2                            Not recording molas
#> 11                                          2 cp
#> 22                           streaky red tide   
#> 25                                shark 11ft    
#> 38  off effort to circle on unidentifed object  
#> 42                           water got more blue
#> 46                                     fb2s fb1m
#> 73                      belly has some glare    
#> 84                           patchy red tide    

# Extract all comments containing "record"
y.comm <- airdas_comments(y.read)
y.comm[grepl("record", y.comm$comment_str, ignore.case = TRUE), ]

#>            file_das line_num            DateTime          comment_str
#> 1 airdas_sample.das        1 2015-04-09 12:30:59       Recorder: dd  
#> 2 airdas_sample.das        2 2015-04-09 12:30:59  Not recording molas

# Extract all comments containing "record", but not "recorder"
y.comm <- airdas_comments(y.read)
y.comm[grepl("record", y.comm$comment_str, ignore.case = TRUE) &
         !grepl("recorder", y.comm$comment_str, ignore.case = TRUE), ]

#>            file_das line_num            DateTime          comment_str
#> 2 airdas_sample.das        2 2015-04-09 12:30:59  Not recording molas