Summarize AirDAS effort

Chop AirDAS data into effort segments

airdas_effort(x, ...)

# S3 method for data.frame
airdas_effort(x, ...)

# S3 method for airdas_df
airdas_effort(
  x,
  method,
  sp.codes,
  conditions = NULL,
  dist.method = "vincenty",
  num.cores = NULL,
  ...
)

Arguments

x
airdas_df object; output from airdas_process, or a data frame that can be coerced to a airdas_df object

...
arguments passed to the chopping function specified using method, such as seg.km or seg.min.km

method
character; method to use to chop AirDAS data into effort segments Can be "condition", "equallength", or "section" (case-sensitive) to use airdas_chop_condition, airdas_chop_equal, or airdas_chop_section, respectively

sp.codes
character; species code(s) to include in segdata. These code(s) will be converted to lower case to match airdas_sight

conditions
character vector of names of conditions to include in segdata output. These values must be column names from the output of airdas_process, e.g. 'Bft', 'CCover', etc. If method == "condition", then these also are the conditions which trigger segment chopping when they change.

dist.method
character; method to use to calculate distance between lat/lon coordinates. Can be "greatcircle" to use the great circle distance method (TODO - add ref), or one of "lawofcosines", "haversine", or "vincenty" to use distance. Default is "vincenty"

num.cores
Number of CPUs to over which to distribute computations. Defaults to NULL, which uses one fewer than the number of cores reported by detectCores Using 1 core likely will be faster for smaller datasets

Value

List of three data frames:

segdata: one row for every segment, and columns for information including unique segment number, event code that started the associated continuous effort section, start/end/midpoint coordinates, conditions (e.g. Beaufort), and number of sightings and number of animals on that segment for every species indicated in sp.codes.

siteinfo: details for all sightings in x, including: the unique segment number it is associated with, segment mid points (lat/lon), and whether the sighting was included in the segdata counts (column included), in addition to the other output information described in airdas_sight.

randpicks: see airdas_chop_equal. NULL if using "condition" method.

Details

This is the top-level function for chopping processed AirDAS data into modeling segments (henceforth 'segments'), and assigning sightings and related information (e.g., weather conditions) to each segment. This function returns data frames with all relevant information for the effort segments and associated sightings ('segdata' and 'siteinfo', respectively). Before chopping, the AirDAS data is filtered for events (rows) where either the 'OnEffort' column is TRUE or the 'Event' column is "E" or "O". In other words, the data is filtered for continuous effort sections (henceforth 'effort sections'), where effort sections run from "T"/"R" to "E"/"O" events (inclusive), and then passed to the chopping function specified using method. All on effort events must not have NA Lat or Lon values; note Lat/Lon values for 1 events were 'filled in' in airdas_process.

The following chopping methods are currently available: "condition", "equallength", and "section". When using the "condition" method, effort sections are chopped into segments every time a condition changes, thereby ensuring that the conditions are consistent across the entire segment. See airdas_chop_condition for more details about this method, including arguments that must be passed to it via ....

The "equallength" method consists of chopping effort sections into equal-length segments of length seg.km, and doing a weighted average of the conditions for the length of that segment. See airdas_chop_equal for more details about this method, including arguments that must be passed to it via ....

The "section" method involves 'chopping' the effort into continuous effort sections, i.e. each continuous effort section is a single effort segment. See airdas_chop_section for more details about this method.

The sightings included in the segdata counts are filtered as follows: Beaufort is less than or equal to five, the declination angle is less than 78, it is a standard sighting, and the sighting was made while on effort. Included sightings are those with a TRUE value in the 'included' column in siteinfo (described below). TODO: Allow user to specify this.

The distance between the lat/lon points of subsequent events is calculated using the method specified in dist.method See airdas_sight for how the sightings are processed.

Examples

y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.proc <- airdas_process(y)

# Using "condition" method
airdas_effort(
  y.proc, method = "condition", sp.codes = c("mn", "bm"),
  conditions = "Bft", seg.min.km = 0.05, num.cores = 1
)

#> $segdata
#>   segnum seg_idx event transect              file stlin endlin     lat1
#> 1      1     1_1     T       T1 airdas_sample.das     3     36 39.23550
#> 2      2     1_2     T       T1 airdas_sample.das    36     37 39.23400
#> 3      3     2_1     R       T1 airdas_sample.das    40     44 39.23400
#> 4      4     2_2     R       T1 airdas_sample.das    44     49 39.23317
#> 5      5     3_1     T       T2 airdas_sample.das    53     78 39.22000
#> 6      6     3_2     T       T2 airdas_sample.das    78     88 39.21683
#>        lon1     lat2      lon2     mlat      mlon    dist           mDateTime
#> 1 -123.1563 39.23400 -123.4033 39.23309 -123.2798 21.3548 2015-04-09 12:34:28
#> 2 -123.4033 39.23400 -123.4047 39.23400 -123.4040  0.1151 2015-04-09 12:37:58
#> 3 -123.4915 39.23317 -123.5143 39.23323 -123.4942  3.4703 2015-04-09 12:40:34
#> 4 -123.5143 39.23283 -123.6032 39.23245 -123.5589  7.6861 2015-04-09 12:42:25
#> 5 -123.6078 39.21683 -123.3605 39.21668 -123.4842 21.3813 2015-04-09 12:50:27
#> 6 -123.3605 39.21700 -123.2417 39.21700 -123.3010 10.2657 2015-04-09 12:55:47
#>   year month day    mtime maxdistBft bm_nSI bm_ANI mn_nSI mn_ANI
#> 1 2015     4   9 12:34:28          1      3     11      5     23
#> 2 2015     4   9 12:37:58          2      0      0      0      0
#> 3 2015     4   9 12:40:34          2      0      0      0      0
#> 4 2015     4   9 12:42:25          1      0      0      1      3
#> 5 2015     4   9 12:50:27          1      3     13      1      6
#> 6 2015     4   9 12:55:47          2      0      0      2     12
#> 
#> $siteinfo
#>    Event            DateTime      Lat       Lon OnEffort Trans Bft CCover Jelly
#> 1      S 2015-04-09 12:31:35 39.23367 -123.1783     TRUE    T1   1     10     0
#> 2      S 2015-04-09 12:31:47 39.23350 -123.1857     TRUE    T1   1     10     0
#> 3      S 2015-04-09 12:31:56 39.23333 -123.1917     TRUE    T1   1     10     0
#> 4      S 2015-04-09 12:32:20 39.23233 -123.2053     TRUE    T1   1     10     0
#> 5      S 2015-04-09 12:32:30 39.23233 -123.2108     TRUE    T1   1     10     0
#> 6      S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 7      S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 8      S 2015-04-09 12:33:43 39.23333 -123.2550     TRUE    T1   1     10     0
#> 9      S 2015-04-09 12:35:13 39.23317 -123.3073     TRUE    T1   1     10     2
#> 10     S 2015-04-09 12:36:13 39.23333 -123.3418     TRUE    T1   1     10     2
#> 11     S 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 12     t 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 13     S 2015-04-09 12:36:43 39.23350 -123.3592     TRUE    T1   1     10     2
#> 14     S 2015-04-09 12:37:01 39.23400 -123.3697     TRUE    T1   1     10     2
#> 15     S 2015-04-09 12:37:16 39.23433 -123.3790     TRUE    T1   1     10     2
#> 16     S 2015-04-09 12:37:37 39.23417 -123.3910     TRUE    T1   1     10     2
#> 17     S 2015-04-09 12:42:13 39.23250 -123.5515     TRUE    T1   1     20     0
#> 18     S 2015-04-09 12:51:40 39.21750 -123.4410     TRUE    T2   1     20     0
#> 19     S 2015-04-09 12:51:55 39.21750 -123.4317     TRUE    T2   1     20     0
#> 20     S 2015-04-09 12:52:03 39.21733 -123.4270     TRUE    T2   1     20     0
#> 21     S 2015-04-09 12:52:04 39.21733 -123.4270     TRUE    T2   1     20     0
#> 22     S 2015-04-09 12:52:07 39.21733 -123.4258     TRUE    T2   1     20     0
#> 23     S 2015-04-09 12:52:17 39.21700 -123.4185     TRUE    T2   1     20     0
#> 24     S 2015-04-09 12:52:54 39.21700 -123.3982     TRUE    T2   1     20     0
#> 25     S 2015-04-09 12:54:34 39.21667 -123.3422     TRUE    T2   2     20     0
#> 26     S 2015-04-09 12:55:20 39.21700 -123.3172     TRUE    T2   2     20     0
#> 27     S 2015-04-09 12:56:49 39.21717 -123.2685     TRUE    T2   2     20     0
#> 28     s 2015-04-09 12:57:00 39.21700 -123.2617     TRUE    T2   2     20     0
#>    HorizSun VertSun HKR  Haze  Kelp RedTide AltFt SpKnot ObsL ObsB ObsR Rec VLI
#> 1         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 2         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 3         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 4         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 5         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 6         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 7         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 8         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 9         6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 10        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 11        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 12        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 13        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 14        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 15        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 16        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 17        6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 18       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 19       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 20       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 21       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 22       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 23       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 24       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 25       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 26       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 27       12      NA   r FALSE FALSE    TRUE   680     90   aa   dd   cc  bb   g
#> 28       12      NA   r FALSE FALSE    TRUE   680     90   aa   dd   cc  bb   g
#>    VLO VB VRI VRO EffortDot EventNum          file_das line_num file_type
#> 1    g  g   g   g      TRUE        9 airdas_sample.das        9    turtle
#> 2    g  g   g   g      TRUE       10 airdas_sample.das       10    turtle
#> 3    g  g   g   g      TRUE       12 airdas_sample.das       12    turtle
#> 4    g  g   g   g      TRUE       14 airdas_sample.das       14    turtle
#> 5    g  g   g   g      TRUE       15 airdas_sample.das       15    turtle
#> 6    g  g   g   g      TRUE       16 airdas_sample.das       16    turtle
#> 7    g  g   g   g      TRUE       16 airdas_sample.das       16    turtle
#> 8    g  g   g   g      TRUE       18 airdas_sample.das       19    turtle
#> 9    g  g   g   g      TRUE       23 airdas_sample.das       24    turtle
#> 10   g  g   g   g      TRUE       26 airdas_sample.das       27    turtle
#> 11   g  g   g   g      TRUE       27 airdas_sample.das       28    turtle
#> 12   g  g   g   g      TRUE       28 airdas_sample.das       29    turtle
#> 13   g  g   g   g      TRUE       29 airdas_sample.das       30    turtle
#> 14   g  g   g   g      TRUE       31 airdas_sample.das       32    turtle
#> 15   g  g   g   g      TRUE       32 airdas_sample.das       33    turtle
#> 16   g  g   g   g      TRUE       34 airdas_sample.das       35    turtle
#> 17   g  g   g   g      TRUE       46 airdas_sample.das       47    turtle
#> 18   g  g   g   g      TRUE       65 airdas_sample.das       66    turtle
#> 19   g  g   g   g      TRUE       66 airdas_sample.das       67    turtle
#> 20   g  g   g   g      TRUE       68 airdas_sample.das       69    turtle
#> 21   g  g   g   g      TRUE       69 airdas_sample.das       70    turtle
#> 22   g  g   g   g      TRUE       70 airdas_sample.das       71    turtle
#> 23   g  g   g   g      TRUE       71 airdas_sample.das       72    turtle
#> 24   g  g   g   g      TRUE       73 airdas_sample.das       74    turtle
#> 25   g  p   g   g      TRUE       78 airdas_sample.das       79    turtle
#> 26   g  p   g   g      TRUE       80 airdas_sample.das       81    turtle
#> 27   g  p   g   g      TRUE       84 airdas_sample.das       85    turtle
#> 28   g  p   g   g      TRUE       86 airdas_sample.das       87    turtle
#>    seg_idx segnum     mlat      mlon SightNo  Obs Angle SightStd Mixed GsTotal
#> 1      1_1      1 39.23309 -123.2798       1   cc    67     TRUE FALSE       6
#> 2      1_1      1 39.23309 -123.2798       2   bb   -60     TRUE FALSE       6
#> 3      1_1      1 39.23309 -123.2798       3   cc    38     TRUE FALSE       2
#> 4      1_1      1 39.23309 -123.2798       4   aa   -50     TRUE FALSE       5
#> 5      1_1      1 39.23309 -123.2798       5   cc    49     TRUE FALSE       4
#> 6      1_1      1 39.23309 -123.2798       6   cc    42     TRUE  TRUE      10
#> 7      1_1      1 39.23309 -123.2798       6   cc    42     TRUE  TRUE      10
#> 8      1_1      1 39.23309 -123.2798       7   aa   -53     TRUE FALSE       5
#> 9      1_1      1 39.23309 -123.2798       8   dd    24     TRUE FALSE       2
#> 10     1_1      1 39.23309 -123.2798       9   aa   -57     TRUE FALSE       4
#> 11     1_1      1 39.23309 -123.2798      10   cc    65     TRUE FALSE       5
#> 12     1_1      1 39.23309 -123.2798    <NA>   aa   -20     TRUE FALSE       1
#> 13     1_1      1 39.23309 -123.2798      11   bb    87     TRUE FALSE       3
#> 14     1_1      1 39.23309 -123.2798      12   bb    82     TRUE FALSE       5
#> 15     1_1      1 39.23309 -123.2798      13   bb    54     TRUE FALSE       6
#> 16     1_1      1 39.23309 -123.2798      14   cc    62     TRUE FALSE       2
#> 17     2_2      4 39.23245 -123.5589      15   bb    50     TRUE FALSE       3
#> 18     3_1      5 39.21668 -123.4842      16   aa   -31     TRUE FALSE       6
#> 19     3_1      5 39.21668 -123.4842      17   bb   -59     TRUE FALSE       6
#> 20     3_1      5 39.21668 -123.4842      18   aa   -62     TRUE FALSE       2
#> 21     3_1      5 39.21668 -123.4842      19   cc    42     TRUE FALSE       3
#> 22     3_1      5 39.21668 -123.4842      20   aa   -48     TRUE FALSE       4
#> 23     3_1      5 39.21668 -123.4842      21   cc    63     TRUE FALSE       6
#> 24     3_1      5 39.21668 -123.4842      22   cc    43     TRUE FALSE       2
#> 25     3_2      6 39.21700 -123.3010      23   cc    56     TRUE FALSE       6
#> 26     3_2      6 39.21700 -123.3010      24   cc    37     TRUE FALSE       6
#> 27     3_2      6 39.21700 -123.3010      25   bb   -76     TRUE FALSE       1
#> 28     3_2      6 39.21700 -123.3010      25 <NA>   -70       NA    NA      NA
#>      Sp GsSp TurtleSizeFt TurtleDirection TurtleTail included
#> 1    mn    6           NA              NA       <NA>     TRUE
#> 2    mn    6           NA              NA       <NA>     TRUE
#> 3    mn    2           NA              NA       <NA>     TRUE
#> 4    mn    5           NA              NA       <NA>     TRUE
#> 5    bm    4           NA              NA       <NA>     TRUE
#> 6    gg    8           NA              NA       <NA>     TRUE
#> 7    er    2           NA              NA       <NA>     TRUE
#> 8    bm    5           NA              NA       <NA>     TRUE
#> 9    bm    2           NA              NA       <NA>     TRUE
#> 10   mn    4           NA              NA       <NA>     TRUE
#> 11   gg    5           NA              NA       <NA>     TRUE
#> 12   dc    1            5              90          n     TRUE
#> 13   pp    3           NA              NA       <NA>    FALSE
#> 14   gg    5           NA              NA       <NA>    FALSE
#> 15   pp    6           NA              NA       <NA>     TRUE
#> 16   gg    2           NA              NA       <NA>     TRUE
#> 17   mn    3           NA              NA       <NA>     TRUE
#> 18   mn    6           NA              NA       <NA>     TRUE
#> 19   bm    6           NA              NA       <NA>     TRUE
#> 20   gg    2           NA              NA       <NA>     TRUE
#> 21   bm    3           NA              NA       <NA>     TRUE
#> 22   bm    4           NA              NA       <NA>     TRUE
#> 23   pp    6           NA              NA       <NA>     TRUE
#> 24   pp    2           NA              NA       <NA>     TRUE
#> 25   mn    6           NA              NA       <NA>     TRUE
#> 26   mn    6           NA              NA       <NA>     TRUE
#> 27   pp    1           NA              NA       <NA>     TRUE
#> 28 <NA>   NA           NA              NA       <NA>    FALSE
#> 
#> $randpicks
#> NULL
#> 

# Using "equallength" method
y.rand <- system.file("airdas_sample_randpicks.csv", package = "swfscAirDAS")
airdas_effort(
  y.proc, method = "equallength", sp.codes = c("mn", "bm"),
  conditions = c("Bft", "CCover"),
  seg.km = 3, randpicks.load = y.rand, num.cores = 1
)

#> $segdata
#>    segnum seg_idx event transect              file stlin endlin     lat1
#> 1       1     1_1     T       T1 airdas_sample.das     3     12 39.23550
#> 2       2     1_2     T       T1 airdas_sample.das    12     18 39.23335
#> 3       3     1_3     T       T1 airdas_sample.das    18     20 39.23254
#> 4       4     1_4     T       T1 airdas_sample.das    20     22 39.23313
#> 5       5     1_5     T       T1 airdas_sample.das    22     26 39.23317
#> 6       6     1_6     T       T1 airdas_sample.das    26     31 39.23329
#> 7       7     1_7     T       T1 airdas_sample.das    31     37 39.23369
#> 8       8     2_1     R       T1 airdas_sample.das    40     43 39.23400
#> 9       9     2_2     R       T1 airdas_sample.das    43     45 39.23317
#> 10     10     2_3     R       T1 airdas_sample.das    45     48 39.23283
#> 11     11     2_4     R       T1 airdas_sample.das    48     49 39.23239
#> 12     12     3_1     T       T2 airdas_sample.das    53     59 39.22000
#> 13     13     3_2     T       T2 airdas_sample.das    59     62 39.21808
#> 14     14     3_3     T       T2 airdas_sample.das    62     63 39.21783
#> 15     15     3_4     T       T2 airdas_sample.das    63     65 39.21641
#> 16     16     3_5     T       T2 airdas_sample.das    65     67 39.21720
#> 17     17     3_6     T       T2 airdas_sample.das    67     74 39.21750
#> 18     18     3_7     T       T2 airdas_sample.das    74     76 39.21700
#> 19     19     3_8     T       T2 airdas_sample.das    76     80 39.21687
#> 20     20     3_9     T       T2 airdas_sample.das    80     82 39.21667
#> 21     21    3_10     T       T2 airdas_sample.das    82     83 39.21700
#> 22     22    3_11     T       T2 airdas_sample.das    83     88 39.21728
#>         lon1     lat2      lon2     mlat      mlon   dist           mDateTime
#> 1  -123.1563 39.23335 -123.1909 39.23412 -123.1736 3.0000 2015-04-09 12:31:27
#> 2  -123.1909 39.23254 -123.2256 39.23233 -123.2082 3.0000 2015-04-09 12:32:28
#> 3  -123.2256 39.23313 -123.2603 39.23302 -123.2430 3.0000 2015-04-09 12:33:30
#> 4  -123.2603 39.23317 -123.2950 39.23308 -123.2777 3.0000 2015-04-09 12:34:28
#> 5  -123.2950 39.23329 -123.3297 39.23317 -123.3124 3.0000 2015-04-09 12:35:28
#> 6  -123.3297 39.23369 -123.3645 39.23344 -123.3472 3.0000 2015-04-09 12:36:30
#> 7  -123.3645 39.23400 -123.4047 39.23426 -123.3846 3.4699 2015-04-09 12:37:30
#> 8  -123.4915 39.23317 -123.5088 39.23333 -123.4915 3.0000 2015-04-09 12:40:30
#> 9  -123.5088 39.23283 -123.5338 39.23305 -123.5214 2.1564 2015-04-09 12:41:30
#> 10 -123.5338 39.23239 -123.5685 39.23250 -123.5513 3.0000 2015-04-09 12:42:30
#> 11 -123.5685 39.23283 -123.6032 39.23249 -123.5860 3.0000 2015-04-09 12:43:21
#> 12 -123.6078 39.21808 -123.5733 39.21892 -123.5905 3.0000 2015-04-09 12:47:25
#> 13 -123.5733 39.21783 -123.5385 39.21771 -123.5558 3.0000 2015-04-09 12:48:30
#> 14 -123.5385 39.21641 -123.5039 39.21725 -123.5211 3.0000 2015-04-09 12:49:30
#> 15 -123.5039 39.21720 -123.4691 39.21660 -123.4864 3.0000 2015-04-09 12:50:30
#> 16 -123.4691 39.21750 -123.4344 39.21743 -123.4517 3.0000 2015-04-09 12:51:27
#> 17 -123.4344 39.21700 -123.3997 39.21700 -123.4170 3.0000 2015-04-09 12:52:24
#> 18 -123.3997 39.21687 -123.3650 39.21703 -123.3822 3.0000 2015-04-09 12:53:27
#> 19 -123.3650 39.21667 -123.3302 39.21672 -123.3475 3.0000 2015-04-09 12:54:30
#> 20 -123.3302 39.21700 -123.3111 39.21690 -123.3206 1.6470 2015-04-09 12:55:30
#> 21 -123.3111 39.21728 -123.2764 39.21703 -123.2937 3.0000 2015-04-09 12:56:20
#> 22 -123.2764 39.21700 -123.2417 39.21700 -123.2590 3.0000 2015-04-09 12:57:05
#>    year month day    mtime avgBft avgCCover bm_nSI bm_ANI mn_nSI mn_ANI
#> 1  2015     4   9 12:31:27   1.00     10.00      0      0      2     12
#> 2  2015     4   9 12:32:28   1.00     10.00      1      4      2      7
#> 3  2015     4   9 12:33:30   1.00     10.00      1      5      0      0
#> 4  2015     4   9 12:34:28   1.00     10.00      0      0      0      0
#> 5  2015     4   9 12:35:28   1.00     10.00      1      2      0      0
#> 6  2015     4   9 12:36:30   1.00     10.00      0      0      1      4
#> 7  2015     4   9 12:37:30   1.03     10.00      0      0      0      0
#> 8  2015     4   9 12:40:30   2.00     15.01      0      0      0      0
#> 9  2015     4   9 12:41:30   1.22     20.00      0      0      0      0
#> 10 2015     4   9 12:42:30   1.00     20.00      0      0      1      3
#> 11 2015     4   9 12:43:21   1.00     20.00      0      0      0      0
#> 12 2015     4   9 12:47:25   1.00     20.00      0      0      0      0
#> 13 2015     4   9 12:48:30   1.00     20.00      0      0      0      0
#> 14 2015     4   9 12:49:30   1.00     20.00      0      0      0      0
#> 15 2015     4   9 12:50:30   1.00     20.00      0      0      0      0
#> 16 2015     4   9 12:51:27   1.00     20.00      0      0      1      6
#> 17 2015     4   9 12:52:24   1.00     20.00      3     13      0      0
#> 18 2015     4   9 12:53:27   1.00     20.00      0      0      0      0
#> 19 2015     4   9 12:54:30   1.87     20.00      0      0      1      6
#> 20 2015     4   9 12:55:30   2.00     20.00      0      0      1      6
#> 21 2015     4   9 12:56:20   2.00     20.00      0      0      0      0
#> 22 2015     4   9 12:57:05   2.00     20.00      0      0      0      0
#> 
#> $siteinfo
#>    Event            DateTime      Lat       Lon OnEffort Trans Bft CCover Jelly
#> 1      S 2015-04-09 12:31:35 39.23367 -123.1783     TRUE    T1   1     10     0
#> 2      S 2015-04-09 12:31:47 39.23350 -123.1857     TRUE    T1   1     10     0
#> 3      S 2015-04-09 12:31:56 39.23333 -123.1917     TRUE    T1   1     10     0
#> 4      S 2015-04-09 12:32:20 39.23233 -123.2053     TRUE    T1   1     10     0
#> 5      S 2015-04-09 12:32:30 39.23233 -123.2108     TRUE    T1   1     10     0
#> 6      S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 7      S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 8      S 2015-04-09 12:33:43 39.23333 -123.2550     TRUE    T1   1     10     0
#> 9      S 2015-04-09 12:35:13 39.23317 -123.3073     TRUE    T1   1     10     2
#> 10     S 2015-04-09 12:36:13 39.23333 -123.3418     TRUE    T1   1     10     2
#> 11     S 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 12     t 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 13     S 2015-04-09 12:36:43 39.23350 -123.3592     TRUE    T1   1     10     2
#> 14     S 2015-04-09 12:37:01 39.23400 -123.3697     TRUE    T1   1     10     2
#> 15     S 2015-04-09 12:37:16 39.23433 -123.3790     TRUE    T1   1     10     2
#> 16     S 2015-04-09 12:37:37 39.23417 -123.3910     TRUE    T1   1     10     2
#> 17     S 2015-04-09 12:42:13 39.23250 -123.5515     TRUE    T1   1     20     0
#> 18     S 2015-04-09 12:51:40 39.21750 -123.4410     TRUE    T2   1     20     0
#> 19     S 2015-04-09 12:51:55 39.21750 -123.4317     TRUE    T2   1     20     0
#> 20     S 2015-04-09 12:52:03 39.21733 -123.4270     TRUE    T2   1     20     0
#> 21     S 2015-04-09 12:52:04 39.21733 -123.4270     TRUE    T2   1     20     0
#> 22     S 2015-04-09 12:52:07 39.21733 -123.4258     TRUE    T2   1     20     0
#> 23     S 2015-04-09 12:52:17 39.21700 -123.4185     TRUE    T2   1     20     0
#> 24     S 2015-04-09 12:52:54 39.21700 -123.3982     TRUE    T2   1     20     0
#> 25     S 2015-04-09 12:54:34 39.21667 -123.3422     TRUE    T2   2     20     0
#> 26     S 2015-04-09 12:55:20 39.21700 -123.3172     TRUE    T2   2     20     0
#> 27     S 2015-04-09 12:56:49 39.21717 -123.2685     TRUE    T2   2     20     0
#> 28     s 2015-04-09 12:57:00 39.21700 -123.2617     TRUE    T2   2     20     0
#>    HorizSun VertSun HKR  Haze  Kelp RedTide AltFt SpKnot ObsL ObsB ObsR Rec VLI
#> 1         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 2         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 3         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 4         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 5         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 6         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 7         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 8         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 9         6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 10        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 11        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 12        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 13        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 14        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 15        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 16        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 17        6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 18       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 19       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 20       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 21       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 22       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 23       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 24       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 25       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 26       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 27       12      NA   r FALSE FALSE    TRUE   680     90   aa   dd   cc  bb   g
#> 28       12      NA   r FALSE FALSE    TRUE   680     90   aa   dd   cc  bb   g
#>    VLO VB VRI VRO EffortDot EventNum          file_das line_num file_type
#> 1    g  g   g   g      TRUE        9 airdas_sample.das        9    turtle
#> 2    g  g   g   g      TRUE       10 airdas_sample.das       10    turtle
#> 3    g  g   g   g      TRUE       12 airdas_sample.das       12    turtle
#> 4    g  g   g   g      TRUE       14 airdas_sample.das       14    turtle
#> 5    g  g   g   g      TRUE       15 airdas_sample.das       15    turtle
#> 6    g  g   g   g      TRUE       16 airdas_sample.das       16    turtle
#> 7    g  g   g   g      TRUE       16 airdas_sample.das       16    turtle
#> 8    g  g   g   g      TRUE       18 airdas_sample.das       19    turtle
#> 9    g  g   g   g      TRUE       23 airdas_sample.das       24    turtle
#> 10   g  g   g   g      TRUE       26 airdas_sample.das       27    turtle
#> 11   g  g   g   g      TRUE       27 airdas_sample.das       28    turtle
#> 12   g  g   g   g      TRUE       28 airdas_sample.das       29    turtle
#> 13   g  g   g   g      TRUE       29 airdas_sample.das       30    turtle
#> 14   g  g   g   g      TRUE       31 airdas_sample.das       32    turtle
#> 15   g  g   g   g      TRUE       32 airdas_sample.das       33    turtle
#> 16   g  g   g   g      TRUE       34 airdas_sample.das       35    turtle
#> 17   g  g   g   g      TRUE       46 airdas_sample.das       47    turtle
#> 18   g  g   g   g      TRUE       65 airdas_sample.das       66    turtle
#> 19   g  g   g   g      TRUE       66 airdas_sample.das       67    turtle
#> 20   g  g   g   g      TRUE       68 airdas_sample.das       69    turtle
#> 21   g  g   g   g      TRUE       69 airdas_sample.das       70    turtle
#> 22   g  g   g   g      TRUE       70 airdas_sample.das       71    turtle
#> 23   g  g   g   g      TRUE       71 airdas_sample.das       72    turtle
#> 24   g  g   g   g      TRUE       73 airdas_sample.das       74    turtle
#> 25   g  p   g   g      TRUE       78 airdas_sample.das       79    turtle
#> 26   g  p   g   g      TRUE       80 airdas_sample.das       81    turtle
#> 27   g  p   g   g      TRUE       84 airdas_sample.das       85    turtle
#> 28   g  p   g   g      TRUE       86 airdas_sample.das       87    turtle
#>    seg_idx segnum     mlat      mlon SightNo  Obs Angle SightStd Mixed GsTotal
#> 1      1_1      1 39.23412 -123.1736       1   cc    67     TRUE FALSE       6
#> 2      1_1      1 39.23412 -123.1736       2   bb   -60     TRUE FALSE       6
#> 3      1_2      2 39.23233 -123.2082       3   cc    38     TRUE FALSE       2
#> 4      1_2      2 39.23233 -123.2082       4   aa   -50     TRUE FALSE       5
#> 5      1_2      2 39.23233 -123.2082       5   cc    49     TRUE FALSE       4
#> 6      1_2      2 39.23233 -123.2082       6   cc    42     TRUE  TRUE      10
#> 7      1_2      2 39.23233 -123.2082       6   cc    42     TRUE  TRUE      10
#> 8      1_3      3 39.23302 -123.2430       7   aa   -53     TRUE FALSE       5
#> 9      1_5      5 39.23317 -123.3124       8   dd    24     TRUE FALSE       2
#> 10     1_6      6 39.23344 -123.3472       9   aa   -57     TRUE FALSE       4
#> 11     1_6      6 39.23344 -123.3472      10   cc    65     TRUE FALSE       5
#> 12     1_6      6 39.23344 -123.3472    <NA>   aa   -20     TRUE FALSE       1
#> 13     1_6      6 39.23344 -123.3472      11   bb    87     TRUE FALSE       3
#> 14     1_7      7 39.23426 -123.3846      12   bb    82     TRUE FALSE       5
#> 15     1_7      7 39.23426 -123.3846      13   bb    54     TRUE FALSE       6
#> 16     1_7      7 39.23426 -123.3846      14   cc    62     TRUE FALSE       2
#> 17     2_3     10 39.23250 -123.5513      15   bb    50     TRUE FALSE       3
#> 18     3_5     16 39.21743 -123.4517      16   aa   -31     TRUE FALSE       6
#> 19     3_6     17 39.21700 -123.4170      17   bb   -59     TRUE FALSE       6
#> 20     3_6     17 39.21700 -123.4170      18   aa   -62     TRUE FALSE       2
#> 21     3_6     17 39.21700 -123.4170      19   cc    42     TRUE FALSE       3
#> 22     3_6     17 39.21700 -123.4170      20   aa   -48     TRUE FALSE       4
#> 23     3_6     17 39.21700 -123.4170      21   cc    63     TRUE FALSE       6
#> 24     3_7     18 39.21703 -123.3822      22   cc    43     TRUE FALSE       2
#> 25     3_8     19 39.21672 -123.3475      23   cc    56     TRUE FALSE       6
#> 26     3_9     20 39.21690 -123.3206      24   cc    37     TRUE FALSE       6
#> 27    3_11     22 39.21700 -123.2590      25   bb   -76     TRUE FALSE       1
#> 28    3_11     22 39.21700 -123.2590      25 <NA>   -70       NA    NA      NA
#>      Sp GsSp TurtleSizeFt TurtleDirection TurtleTail included
#> 1    mn    6           NA              NA       <NA>     TRUE
#> 2    mn    6           NA              NA       <NA>     TRUE
#> 3    mn    2           NA              NA       <NA>     TRUE
#> 4    mn    5           NA              NA       <NA>     TRUE
#> 5    bm    4           NA              NA       <NA>     TRUE
#> 6    gg    8           NA              NA       <NA>     TRUE
#> 7    er    2           NA              NA       <NA>     TRUE
#> 8    bm    5           NA              NA       <NA>     TRUE
#> 9    bm    2           NA              NA       <NA>     TRUE
#> 10   mn    4           NA              NA       <NA>     TRUE
#> 11   gg    5           NA              NA       <NA>     TRUE
#> 12   dc    1            5              90          n     TRUE
#> 13   pp    3           NA              NA       <NA>    FALSE
#> 14   gg    5           NA              NA       <NA>    FALSE
#> 15   pp    6           NA              NA       <NA>     TRUE
#> 16   gg    2           NA              NA       <NA>     TRUE
#> 17   mn    3           NA              NA       <NA>     TRUE
#> 18   mn    6           NA              NA       <NA>     TRUE
#> 19   bm    6           NA              NA       <NA>     TRUE
#> 20   gg    2           NA              NA       <NA>     TRUE
#> 21   bm    3           NA              NA       <NA>     TRUE
#> 22   bm    4           NA              NA       <NA>     TRUE
#> 23   pp    6           NA              NA       <NA>     TRUE
#> 24   pp    2           NA              NA       <NA>     TRUE
#> 25   mn    6           NA              NA       <NA>     TRUE
#> 26   mn    6           NA              NA       <NA>     TRUE
#> 27   pp    1           NA              NA       <NA>     TRUE
#> 28 <NA>   NA           NA              NA       <NA>    FALSE
#> 
#> $randpicks
#>   effort_section randpicks
#> 1              1         7
#> 2              2         2
#> 3              3         9
#> 

# Using "section" method
airdas_effort(
  y.proc, method = "section", sp.codes = c("mn", "bm"),
  num.cores = 1
)

#> $segdata
#>   segnum seg_idx event transect              file stlin endlin    lat1
#> 1      1     1_1     T       T1 airdas_sample.das     3     37 39.2355
#> 2      2     2_1     R       T1 airdas_sample.das    40     49 39.2340
#> 3      3     3_1     T       T2 airdas_sample.das    53     88 39.2200
#>        lon1     lat2      lon2     mlat      mlon    dist           mDateTime
#> 1 -123.1563 39.23400 -123.4047 39.23309 -123.2804 21.4699 2015-04-09 12:34:29
#> 2 -123.4915 39.23283 -123.6032 39.23274 -123.5388 11.1564 2015-04-09 12:41:51
#> 3 -123.6078 39.21700 -123.2417 39.21729 -123.4248 31.6470 2015-04-09 12:52:10
#>   year month day    mtime avgBft avgCCover avgJelly avgHorizSun avgVertSun
#> 1 2015     4   9 12:34:29   1.01     10.00     0.88           6         NA
#> 2 2015     4   9 12:41:51   1.31     18.66     0.00           6         NA
#> 3 2015     4   9 12:52:10   1.32     20.00     0.00          12         NA
#>   avgHaze avgKelp avgRedTide avgAltFt avgSpKnot bm_nSI bm_ANI mn_nSI mn_ANI
#> 1       0       0       0.44   650.00     100.0      3     11      5     23
#> 2       0       0       0.00   650.00     100.0      0      0      1      3
#> 3       0       0       0.09   669.87      93.5      3     13      3     18
#> 
#> $siteinfo
#>    Event            DateTime      Lat       Lon OnEffort Trans Bft CCover Jelly
#> 1      S 2015-04-09 12:31:35 39.23367 -123.1783     TRUE    T1   1     10     0
#> 2      S 2015-04-09 12:31:47 39.23350 -123.1857     TRUE    T1   1     10     0
#> 3      S 2015-04-09 12:31:56 39.23333 -123.1917     TRUE    T1   1     10     0
#> 4      S 2015-04-09 12:32:20 39.23233 -123.2053     TRUE    T1   1     10     0
#> 5      S 2015-04-09 12:32:30 39.23233 -123.2108     TRUE    T1   1     10     0
#> 6      S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 7      S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 8      S 2015-04-09 12:33:43 39.23333 -123.2550     TRUE    T1   1     10     0
#> 9      S 2015-04-09 12:35:13 39.23317 -123.3073     TRUE    T1   1     10     2
#> 10     S 2015-04-09 12:36:13 39.23333 -123.3418     TRUE    T1   1     10     2
#> 11     S 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 12     t 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 13     S 2015-04-09 12:36:43 39.23350 -123.3592     TRUE    T1   1     10     2
#> 14     S 2015-04-09 12:37:01 39.23400 -123.3697     TRUE    T1   1     10     2
#> 15     S 2015-04-09 12:37:16 39.23433 -123.3790     TRUE    T1   1     10     2
#> 16     S 2015-04-09 12:37:37 39.23417 -123.3910     TRUE    T1   1     10     2
#> 17     S 2015-04-09 12:42:13 39.23250 -123.5515     TRUE    T1   1     20     0
#> 18     S 2015-04-09 12:51:40 39.21750 -123.4410     TRUE    T2   1     20     0
#> 19     S 2015-04-09 12:51:55 39.21750 -123.4317     TRUE    T2   1     20     0
#> 20     S 2015-04-09 12:52:03 39.21733 -123.4270     TRUE    T2   1     20     0
#> 21     S 2015-04-09 12:52:04 39.21733 -123.4270     TRUE    T2   1     20     0
#> 22     S 2015-04-09 12:52:07 39.21733 -123.4258     TRUE    T2   1     20     0
#> 23     S 2015-04-09 12:52:17 39.21700 -123.4185     TRUE    T2   1     20     0
#> 24     S 2015-04-09 12:52:54 39.21700 -123.3982     TRUE    T2   1     20     0
#> 25     S 2015-04-09 12:54:34 39.21667 -123.3422     TRUE    T2   2     20     0
#> 26     S 2015-04-09 12:55:20 39.21700 -123.3172     TRUE    T2   2     20     0
#> 27     S 2015-04-09 12:56:49 39.21717 -123.2685     TRUE    T2   2     20     0
#> 28     s 2015-04-09 12:57:00 39.21700 -123.2617     TRUE    T2   2     20     0
#>    HorizSun VertSun HKR  Haze  Kelp RedTide AltFt SpKnot ObsL ObsB ObsR Rec VLI
#> 1         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 2         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 3         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 4         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 5         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 6         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 7         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 8         6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 9         6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 10        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 11        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 12        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 13        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 14        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 15        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 16        6      NA   r FALSE FALSE    TRUE   650    100   aa   bb   cc  dd   g
#> 17        6      NA   n FALSE FALSE   FALSE   650    100   aa   bb   cc  dd   g
#> 18       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 19       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 20       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 21       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 22       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 23       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 24       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 25       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 26       12      NA   n FALSE FALSE   FALSE   680     90   aa   dd   cc  bb   g
#> 27       12      NA   r FALSE FALSE    TRUE   680     90   aa   dd   cc  bb   g
#> 28       12      NA   r FALSE FALSE    TRUE   680     90   aa   dd   cc  bb   g
#>    VLO VB VRI VRO EffortDot EventNum          file_das line_num file_type
#> 1    g  g   g   g      TRUE        9 airdas_sample.das        9    turtle
#> 2    g  g   g   g      TRUE       10 airdas_sample.das       10    turtle
#> 3    g  g   g   g      TRUE       12 airdas_sample.das       12    turtle
#> 4    g  g   g   g      TRUE       14 airdas_sample.das       14    turtle
#> 5    g  g   g   g      TRUE       15 airdas_sample.das       15    turtle
#> 6    g  g   g   g      TRUE       16 airdas_sample.das       16    turtle
#> 7    g  g   g   g      TRUE       16 airdas_sample.das       16    turtle
#> 8    g  g   g   g      TRUE       18 airdas_sample.das       19    turtle
#> 9    g  g   g   g      TRUE       23 airdas_sample.das       24    turtle
#> 10   g  g   g   g      TRUE       26 airdas_sample.das       27    turtle
#> 11   g  g   g   g      TRUE       27 airdas_sample.das       28    turtle
#> 12   g  g   g   g      TRUE       28 airdas_sample.das       29    turtle
#> 13   g  g   g   g      TRUE       29 airdas_sample.das       30    turtle
#> 14   g  g   g   g      TRUE       31 airdas_sample.das       32    turtle
#> 15   g  g   g   g      TRUE       32 airdas_sample.das       33    turtle
#> 16   g  g   g   g      TRUE       34 airdas_sample.das       35    turtle
#> 17   g  g   g   g      TRUE       46 airdas_sample.das       47    turtle
#> 18   g  g   g   g      TRUE       65 airdas_sample.das       66    turtle
#> 19   g  g   g   g      TRUE       66 airdas_sample.das       67    turtle
#> 20   g  g   g   g      TRUE       68 airdas_sample.das       69    turtle
#> 21   g  g   g   g      TRUE       69 airdas_sample.das       70    turtle
#> 22   g  g   g   g      TRUE       70 airdas_sample.das       71    turtle
#> 23   g  g   g   g      TRUE       71 airdas_sample.das       72    turtle
#> 24   g  g   g   g      TRUE       73 airdas_sample.das       74    turtle
#> 25   g  p   g   g      TRUE       78 airdas_sample.das       79    turtle
#> 26   g  p   g   g      TRUE       80 airdas_sample.das       81    turtle
#> 27   g  p   g   g      TRUE       84 airdas_sample.das       85    turtle
#> 28   g  p   g   g      TRUE       86 airdas_sample.das       87    turtle
#>    seg_idx segnum     mlat      mlon SightNo  Obs Angle SightStd Mixed GsTotal
#> 1      1_1      1 39.23309 -123.2804       1   cc    67     TRUE FALSE       6
#> 2      1_1      1 39.23309 -123.2804       2   bb   -60     TRUE FALSE       6
#> 3      1_1      1 39.23309 -123.2804       3   cc    38     TRUE FALSE       2
#> 4      1_1      1 39.23309 -123.2804       4   aa   -50     TRUE FALSE       5
#> 5      1_1      1 39.23309 -123.2804       5   cc    49     TRUE FALSE       4
#> 6      1_1      1 39.23309 -123.2804       6   cc    42     TRUE  TRUE      10
#> 7      1_1      1 39.23309 -123.2804       6   cc    42     TRUE  TRUE      10
#> 8      1_1      1 39.23309 -123.2804       7   aa   -53     TRUE FALSE       5
#> 9      1_1      1 39.23309 -123.2804       8   dd    24     TRUE FALSE       2
#> 10     1_1      1 39.23309 -123.2804       9   aa   -57     TRUE FALSE       4
#> 11     1_1      1 39.23309 -123.2804      10   cc    65     TRUE FALSE       5
#> 12     1_1      1 39.23309 -123.2804    <NA>   aa   -20     TRUE FALSE       1
#> 13     1_1      1 39.23309 -123.2804      11   bb    87     TRUE FALSE       3
#> 14     1_1      1 39.23309 -123.2804      12   bb    82     TRUE FALSE       5
#> 15     1_1      1 39.23309 -123.2804      13   bb    54     TRUE FALSE       6
#> 16     1_1      1 39.23309 -123.2804      14   cc    62     TRUE FALSE       2
#> 17     2_1      2 39.23274 -123.5388      15   bb    50     TRUE FALSE       3
#> 18     3_1      3 39.21729 -123.4248      16   aa   -31     TRUE FALSE       6
#> 19     3_1      3 39.21729 -123.4248      17   bb   -59     TRUE FALSE       6
#> 20     3_1      3 39.21729 -123.4248      18   aa   -62     TRUE FALSE       2
#> 21     3_1      3 39.21729 -123.4248      19   cc    42     TRUE FALSE       3
#> 22     3_1      3 39.21729 -123.4248      20   aa   -48     TRUE FALSE       4
#> 23     3_1      3 39.21729 -123.4248      21   cc    63     TRUE FALSE       6
#> 24     3_1      3 39.21729 -123.4248      22   cc    43     TRUE FALSE       2
#> 25     3_1      3 39.21729 -123.4248      23   cc    56     TRUE FALSE       6
#> 26     3_1      3 39.21729 -123.4248      24   cc    37     TRUE FALSE       6
#> 27     3_1      3 39.21729 -123.4248      25   bb   -76     TRUE FALSE       1
#> 28     3_1      3 39.21729 -123.4248      25 <NA>   -70       NA    NA      NA
#>      Sp GsSp TurtleSizeFt TurtleDirection TurtleTail included
#> 1    mn    6           NA              NA       <NA>     TRUE
#> 2    mn    6           NA              NA       <NA>     TRUE
#> 3    mn    2           NA              NA       <NA>     TRUE
#> 4    mn    5           NA              NA       <NA>     TRUE
#> 5    bm    4           NA              NA       <NA>     TRUE
#> 6    gg    8           NA              NA       <NA>     TRUE
#> 7    er    2           NA              NA       <NA>     TRUE
#> 8    bm    5           NA              NA       <NA>     TRUE
#> 9    bm    2           NA              NA       <NA>     TRUE
#> 10   mn    4           NA              NA       <NA>     TRUE
#> 11   gg    5           NA              NA       <NA>     TRUE
#> 12   dc    1            5              90          n     TRUE
#> 13   pp    3           NA              NA       <NA>    FALSE
#> 14   gg    5           NA              NA       <NA>    FALSE
#> 15   pp    6           NA              NA       <NA>     TRUE
#> 16   gg    2           NA              NA       <NA>     TRUE
#> 17   mn    3           NA              NA       <NA>     TRUE
#> 18   mn    6           NA              NA       <NA>     TRUE
#> 19   bm    6           NA              NA       <NA>     TRUE
#> 20   gg    2           NA              NA       <NA>     TRUE
#> 21   bm    3           NA              NA       <NA>     TRUE
#> 22   bm    4           NA              NA       <NA>     TRUE
#> 23   pp    6           NA              NA       <NA>     TRUE
#> 24   pp    2           NA              NA       <NA>     TRUE
#> 25   mn    6           NA              NA       <NA>     TRUE
#> 26   mn    6           NA              NA       <NA>     TRUE
#> 27   pp    1           NA              NA       <NA>     TRUE
#> 28 <NA>   NA           NA              NA       <NA>    FALSE
#> 
#> $randpicks
#>   effort_section randpicks
#> 1              1        NA
#> 2              2        NA
#> 3              3        NA
#>