Summarize AirDAS effort

Chop AirDAS data into effort segments

Usage

airdas_effort(x, ...)

# S3 method for data.frame
airdas_effort(x, ...)

# S3 method for airdas_df
airdas_effort(
  x,
  method = c("condition", "equallength", "section"),
  conditions = NULL,
  distance.method = c("greatcircle", "lawofcosines", "haversine", "vincenty"),
  num.cores = NULL,
  angle.min = 12,
  bft.max = 5,
  ...
)

Arguments

x

airdas_df object; output from airdas_process, or a data frame that can be coerced to a airdas_df object

...

arguments passed to the chopping function specified using method, such as seg.km or seg.min.km

method

character; method to use to chop AirDAS data into effort segments Can be "condition", "equallength", "section", or any partial match thereof (case sensitive)

conditions

character vector of names of conditions to include in segdata output. These values must be column names from the output of airdas_process, e.g. 'Bft', 'CCover', etc. The default is NULL, in which case all relevant conditions will be included. If method == "condition", then these also are the conditions which trigger segment chopping when they change.

distance.method

character; method to use to calculate distance between lat/lon coordinates. Can be "greatcircle", "lawofcosines", "haversine", "vincenty", or any partial match thereof (case sensitive). Default is "greatcircle"

num.cores

Number of CPUs to over which to distribute computations. Defaults to NULL, which uses one fewer than the number of cores reported by detectCores Using 1 core likely will be faster for smaller datasets

angle.min

passed to airdas_sight

bft.max

numeric; the maximum Beaufort (column 'Bft') for which to mark a sighting as TRUE in 'included' (see Details). Default is 5.

Value

List of three data frames:

segdata: one row for every segment, and columns for information including unique segment number, event code that started the associated continuous effort section, the starting and ending line of the segment in the DAS file (stlin, endlin), start/end/midpoint coordinates(lat1/lon1, lat2/lon2, and mlat/mlon, respectively), the start/end/midpoint date/time of the segment (DateTime1, DateTime2, and mDateTime, respectively; mDateTime is the average of DateTime1 and DateTime2), segment length (dist), and conditions (e.g. Beaufort)

sightinfo: details for all sightings in x, including: the unique segment number it is associated with, segment mid points (lat/lon), the 'included' column described in the Details section, and the output information described in airdas_sight

randpicks: see airdas_chop_equallength. NULL if using "condition" method.

Details

This is the top-level function for chopping processed AirDAS data into modeling segments (henceforth 'segments'), and assigning sightings and related information (e.g., weather conditions) to each segment. This function returns data frames with all relevant information for the effort segments and associated sightings ('segdata' and 'sightinfo', respectively). Before chopping, the AirDAS data is filtered for events (rows) where either the 'OnEffort' column is TRUE or the 'Event' column is "E" or "O". In other words, the data is filtered for continuous effort sections (henceforth 'effort sections'), where effort sections run from "T"/"R" to "E"/"O" events (inclusive), and then passed to the chopping function specified using method. All on effort events must not have NA Lat or Lon values; note Lat/Lon values for 1 events were 'filled in' in airdas_process.

The following chopping methods are currently available: "condition", "equallength", and "section". When using the "condition" method, effort sections are chopped into segments every time a condition specified in conditions changes, thereby ensuring that the conditions are consistent across the entire segment. See airdas_chop_condition for more details about this method, including arguments that must be passed to it via ....

The "equallength" method consists of chopping effort sections into equal-length segments of length seg.km, and doing a weighted average of the conditions for the length of that segment. See airdas_chop_equallength for more details about this method, including arguments that must be passed to it via ....

The "section" method involves 'chopping' the effort into continuous effort sections, i.e. each continuous effort section is a single effort segment. See airdas_chop_section for more details about this method.

The distance between the lat/lon points of subsequent events is calculated using the method specified in distance.method. If "greatcircle", distance_greatcircle is used, while distance is used otherwise. See airdas_sight for how the sightings are processed.

The sightinfo data frame includes the column 'included', which is used in airdas_effort_sight when summarizing the number of sightings and animals for selected species. airdas_effort_sight is a separate function to allow users to personalize the 'included' values as desired for their specific analysis. By default, i.e. in the output of this function, 'included' is TRUE if: the sighting was a standard sighting (see airdas_sight) and in a Beaufort sea state less than or equal to 'btf.max'.

Examples

y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.proc <- airdas_process(y)

airdas_effort(
  y.proc, method = "condition", conditions = "Bft", seg.min.km = 0.05, 
  num.cores = 1
)
#> $segdata
#>       segnum section_id section_sub_id event transect              file stlin
#> lat        1          1              1     T       T1 airdas_sample.das     3
#> lat1       2          1              2     T       T1 airdas_sample.das    36
#> lat2       3          2              1     R       T1 airdas_sample.das    40
#> lat11      4          2              2     R       T1 airdas_sample.das    44
#> lat3       5          3              1     T       T2 airdas_sample.das    53
#> lat12      6          3              2     T       T2 airdas_sample.das    78
#>       endlin     lat1      lon1           DateTime1     lat2      lon2
#> lat       36 39.23550 -123.1563 2015-04-09 12:30:59 39.23400 -123.4033
#> lat1      37 39.23400 -123.4033 2015-04-09 12:37:57 39.23400 -123.4047
#> lat2      44 39.23400 -123.4742 2015-04-09 12:40:00 39.23317 -123.5143
#> lat11     49 39.23317 -123.5143 2015-04-09 12:41:08 39.23283 -123.6032
#> lat3      78 39.22000 -123.6078 2015-04-09 12:46:51 39.21683 -123.3605
#> lat12     88 39.21683 -123.3605 2015-04-09 12:54:04 39.21700 -123.2418
#>                 DateTime2     mlat      mlon           mDateTime    dist year
#> lat   2015-04-09 12:37:57 39.23309 -123.2797 2015-04-09 12:34:28 21.2882 2015
#> lat1  2015-04-09 12:38:00 39.23400 -123.4040 2015-04-09 12:37:58  0.1148 2015
#> lat2  2015-04-09 12:41:08 39.23323 -123.4942 2015-04-09 12:40:34  3.4595 2015
#> lat11 2015-04-09 12:43:43 39.23245 -123.5588 2015-04-09 12:42:25  7.6620 2015
#> lat3  2015-04-09 12:54:04 39.21667 -123.4842 2015-04-09 12:50:27 21.3146 2015
#> lat12 2015-04-09 12:57:30 39.21700 -123.3011 2015-04-09 12:55:47 10.2336 2015
#>       month day    mtime maxdistBft
#> lat       4   9 12:34:28          1
#> lat1      4   9 12:37:58          2
#> lat2      4   9 12:40:34          2
#> lat11     4   9 12:42:25          1
#> lat3      4   9 12:50:27          1
#> lat12     4   9 12:55:47          2
#> 
#> $sightinfo
#>    segnum     mlat      mlon Event            DateTime      Lat       Lon
#> 1       1 39.23309 -123.2797     S 2015-04-09 12:31:35 39.23367 -123.1783
#> 2       1 39.23309 -123.2797     S 2015-04-09 12:31:47 39.23350 -123.1857
#> 3       1 39.23309 -123.2797     S 2015-04-09 12:31:56 39.23333 -123.1917
#> 4       1 39.23309 -123.2797     S 2015-04-09 12:32:20 39.23233 -123.2053
#> 5       1 39.23309 -123.2797     S 2015-04-09 12:32:30 39.23233 -123.2108
#> 6       1 39.23309 -123.2797     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 7       1 39.23309 -123.2797     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 8       1 39.23309 -123.2797     S 2015-04-09 12:33:43 39.23333 -123.2550
#> 9       1 39.23309 -123.2797     S 2015-04-09 12:35:13 39.23317 -123.3073
#> 10      1 39.23309 -123.2797     S 2015-04-09 12:36:13 39.23333 -123.3418
#> 11      1 39.23309 -123.2797     S 2015-04-09 12:36:26 39.23350 -123.3500
#> 12      1 39.23309 -123.2797     t 2015-04-09 12:36:26 39.23350 -123.3500
#> 13      1 39.23309 -123.2797     S 2015-04-09 12:36:43 39.23350 -123.3592
#> 14      1 39.23309 -123.2797     S 2015-04-09 12:37:01 39.23400 -123.3697
#> 15      1 39.23309 -123.2797     S 2015-04-09 12:37:16 39.23433 -123.3790
#> 16      1 39.23309 -123.2797     S 2015-04-09 12:37:37 39.23417 -123.3910
#> 17      4 39.23245 -123.5588     S 2015-04-09 12:42:13 39.23250 -123.5515
#> 18      5 39.21667 -123.4842     S 2015-04-09 12:51:40 39.21750 -123.4410
#> 19      5 39.21667 -123.4842     S 2015-04-09 12:51:55 39.21750 -123.4317
#> 20      5 39.21667 -123.4842     S 2015-04-09 12:52:03 39.21733 -123.4270
#> 21      5 39.21667 -123.4842     S 2015-04-09 12:52:04 39.21733 -123.4270
#> 22      5 39.21667 -123.4842     S 2015-04-09 12:52:07 39.21733 -123.4258
#> 23      5 39.21667 -123.4842     S 2015-04-09 12:52:17 39.21700 -123.4185
#> 24      5 39.21667 -123.4842     S 2015-04-09 12:52:54 39.21700 -123.3982
#> 25      6 39.21700 -123.3011     S 2015-04-09 12:54:34 39.21667 -123.3422
#> 26      6 39.21700 -123.3011     S 2015-04-09 12:55:20 39.21700 -123.3172
#> 27      6 39.21700 -123.3011     S 2015-04-09 12:56:49 39.21717 -123.2685
#> 28      6 39.21700 -123.3011     s 2015-04-09 12:57:00 39.21700 -123.2617
#>    OnEffort Trans Bft CCover Jelly HorizSun VertSun HKR  Haze  Kelp RedTide
#> 1      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 2      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 3      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 4      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 5      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 6      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 7      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 8      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 9      TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 10     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 11     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 12     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 13     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 14     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 15     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 16     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 17     TRUE    T1   1     20     0        6      NA   n FALSE FALSE   FALSE
#> 18     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 19     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 20     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 21     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 22     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 23     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 24     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 25     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 26     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 27     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#> 28     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#>    AltFt SpKnot ObsL ObsB ObsR Rec VLI VLO VB VRI VRO EffortDot EventNum
#> 1    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE        9
#> 2    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       10
#> 3    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       12
#> 4    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       14
#> 5    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       15
#> 6    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 7    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 8    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       18
#> 9    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       23
#> 10   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       26
#> 11   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       27
#> 12   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       28
#> 13   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       29
#> 14   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       31
#> 15   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       32
#> 16   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       34
#> 17   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       46
#> 18   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       65
#> 19   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       66
#> 20   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       68
#> 21   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       69
#> 22   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       70
#> 23   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       71
#> 24   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       73
#> 25   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       78
#> 26   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       80
#> 27   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       84
#> 28   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       86
#>             file_das line_num file_type SightNo  Obs Angle ObsStd SightStd
#> 1  airdas_sample.das        9    turtle       1   cc    67   TRUE     TRUE
#> 2  airdas_sample.das       10    turtle       2   bb   -60   TRUE     TRUE
#> 3  airdas_sample.das       12    turtle       3   cc    38   TRUE     TRUE
#> 4  airdas_sample.das       14    turtle       4   aa   -50   TRUE     TRUE
#> 5  airdas_sample.das       15    turtle       5   cc    49   TRUE     TRUE
#> 6  airdas_sample.das       16    turtle       6   cc    42   TRUE     TRUE
#> 7  airdas_sample.das       16    turtle       6   cc    42   TRUE     TRUE
#> 8  airdas_sample.das       19    turtle       7   aa   -53   TRUE     TRUE
#> 9  airdas_sample.das       24    turtle       8   dd    24  FALSE    FALSE
#> 10 airdas_sample.das       27    turtle       9   aa   -57   TRUE     TRUE
#> 11 airdas_sample.das       28    turtle      10   cc    65   TRUE     TRUE
#> 12 airdas_sample.das       29    turtle    <NA>   aa   -20   TRUE     TRUE
#> 13 airdas_sample.das       30    turtle      11   bb    87   TRUE     TRUE
#> 14 airdas_sample.das       32    turtle      12   bb    82   TRUE     TRUE
#> 15 airdas_sample.das       33    turtle      13   bb    54   TRUE     TRUE
#> 16 airdas_sample.das       35    turtle      14   cc    62   TRUE     TRUE
#> 17 airdas_sample.das       47    turtle      15   bb    50   TRUE     TRUE
#> 18 airdas_sample.das       66    turtle      16   aa   -31   TRUE     TRUE
#> 19 airdas_sample.das       67    turtle      17   bb   -59  FALSE    FALSE
#> 20 airdas_sample.das       69    turtle      18   aa   -62   TRUE     TRUE
#> 21 airdas_sample.das       70    turtle      19   cc    42   TRUE     TRUE
#> 22 airdas_sample.das       71    turtle      20   aa   -48   TRUE     TRUE
#> 23 airdas_sample.das       72    turtle      21   cc    63   TRUE     TRUE
#> 24 airdas_sample.das       74    turtle      22   cc    43   TRUE     TRUE
#> 25 airdas_sample.das       79    turtle      23   cc    56   TRUE     TRUE
#> 26 airdas_sample.das       81    turtle      24   cc    37   TRUE     TRUE
#> 27 airdas_sample.das       85    turtle      25   bb   -76  FALSE    FALSE
#> 28 airdas_sample.das       87    turtle      25 <NA>   -70     NA       NA
#>    Mixed SpCode GsTotal GsSp TurtleSize TurtleDirection TurtleTail included
#> 1  FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 2  FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 3  FALSE     mn       2    2         NA              NA       <NA>     TRUE
#> 4  FALSE     mn       5    5         NA              NA       <NA>     TRUE
#> 5  FALSE     bm       4    4         NA              NA       <NA>     TRUE
#> 6   TRUE     gg      10    8         NA              NA       <NA>     TRUE
#> 7   TRUE     er      10    2         NA              NA       <NA>     TRUE
#> 8  FALSE     bm       5    5         NA              NA       <NA>     TRUE
#> 9  FALSE     bm       2    2         NA              NA       <NA>    FALSE
#> 10 FALSE     mn       4    4         NA              NA       <NA>     TRUE
#> 11 FALSE     gg       5    5         NA              NA       <NA>     TRUE
#> 12 FALSE     dc       1    1          5              90          n     TRUE
#> 13 FALSE     pp       3    3         NA              NA       <NA>     TRUE
#> 14 FALSE     gg       5    5         NA              NA       <NA>     TRUE
#> 15 FALSE     pp       6    6         NA              NA       <NA>     TRUE
#> 16 FALSE     gg       2    2         NA              NA       <NA>     TRUE
#> 17 FALSE     mn       3    3         NA              NA       <NA>     TRUE
#> 18 FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 19 FALSE     bm       6    6         NA              NA       <NA>    FALSE
#> 20 FALSE     gg       2    2         NA              NA       <NA>     TRUE
#> 21 FALSE     bm       3    3         NA              NA       <NA>     TRUE
#> 22 FALSE     bm       4    4         NA              NA       <NA>     TRUE
#> 23 FALSE     pp       6    6         NA              NA       <NA>     TRUE
#> 24 FALSE     pp       2    2         NA              NA       <NA>     TRUE
#> 25 FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 26 FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 27 FALSE     pp       1    1         NA              NA       <NA>    FALSE
#> 28    NA   <NA>      NA   NA         NA              NA       <NA>    FALSE
#> 
#> $randpicks
#> NULL
#> 

y.rand <- system.file("airdas_sample_randpicks.csv", package = "swfscAirDAS")
airdas_effort(
  y.proc, method = "equallength", conditions = c("Bft", "CCover"), 
  seg.km = 3, randpicks.load = y.rand, num.cores = 1
)
#> $segdata
#>       segnum section_id section_sub_id event transect              file stlin
#> lat        1          1              1     T       T1 airdas_sample.das     3
#> lat1       2          1              2     T       T1 airdas_sample.das    12
#> lat2       3          1              3     T       T1 airdas_sample.das    18
#> lat3       4          1              4     T       T1 airdas_sample.das    20
#> lat4       5          1              5     T       T1 airdas_sample.das    22
#> lat5       6          1              6     T       T1 airdas_sample.das    26
#> lat6       7          1              7     T       T1 airdas_sample.das    31
#> lat11      8          2              1     R       T1 airdas_sample.das    40
#> lat12      9          2              2     R       T1 airdas_sample.das    43
#> lat21     10          2              3     R       T1 airdas_sample.das    45
#> lat31     11          2              4     R       T1 airdas_sample.das    48
#> lat13     12          3              1     T       T2 airdas_sample.das    53
#> lat14     13          3              2     T       T2 airdas_sample.das    59
#> lat22     14          3              3     T       T2 airdas_sample.das    62
#> lat32     15          3              4     T       T2 airdas_sample.das    63
#> lat41     16          3              5     T       T2 airdas_sample.das    65
#> lat51     17          3              6     T       T2 airdas_sample.das    67
#> lat61     18          3              7     T       T2 airdas_sample.das    74
#> lat7      19          3              8     T       T2 airdas_sample.das    76
#> lat8      20          3              9     T       T2 airdas_sample.das    80
#> lat9      21          3             10     T       T2 airdas_sample.das    82
#> lat10     22          3             11     T       T2 airdas_sample.das    83
#>       endlin     lat1      lon1           DateTime1     lat2      lon2
#> lat       12 39.23550 -123.1563 2015-04-09 12:30:59 39.23335 -123.1910
#> lat1      18 39.23335 -123.1910 2015-04-09 12:31:55 39.23255 -123.2257
#> lat2      20 39.23255 -123.2257 2015-04-09 12:32:53 39.23312 -123.2606
#> lat3      22 39.23312 -123.2606 2015-04-09 12:33:54 39.23317 -123.2955
#> lat4      26 39.23317 -123.2955 2015-04-09 12:34:53 39.23330 -123.3302
#> lat5      31 39.23330 -123.3302 2015-04-09 12:35:54 39.23371 -123.3651
#> lat6      37 39.23371 -123.3651 2015-04-09 12:36:54 39.23400 -123.4047
#> lat11     43 39.23400 -123.4742 2015-04-09 12:40:00 39.23317 -123.5089
#> lat12     45 39.23317 -123.5089 2015-04-09 12:40:59 39.23284 -123.5335
#> lat21     48 39.23284 -123.5335 2015-04-09 12:41:42 39.23239 -123.5684
#> lat31     49 39.23239 -123.5684 2015-04-09 12:42:43 39.23283 -123.6032
#> lat13     59 39.22000 -123.6078 2015-04-09 12:46:51 39.21808 -123.5733
#> lat14     62 39.21808 -123.5733 2015-04-09 12:47:51 39.21783 -123.5383
#> lat22     63 39.21783 -123.5383 2015-04-09 12:48:51 39.21640 -123.5037
#> lat32     65 39.21640 -123.5037 2015-04-09 12:49:51 39.21722 -123.4687
#> lat41     67 39.21722 -123.4687 2015-04-09 12:50:53 39.21750 -123.4339
#> lat51     74 39.21750 -123.4339 2015-04-09 12:51:51 39.21700 -123.3990
#> lat61     76 39.21700 -123.3990 2015-04-09 12:52:52 39.21686 -123.3644
#> lat7      80 39.21686 -123.3644 2015-04-09 12:53:55 39.21667 -123.3294
#> lat8      82 39.21667 -123.3294 2015-04-09 12:54:58 39.21700 -123.3114
#> lat9      83 39.21700 -123.3114 2015-04-09 12:55:31 39.21728 -123.2766
#> lat10     88 39.21728 -123.2766 2015-04-09 12:56:33 39.21700 -123.2418
#>                 DateTime2     mlat      mlon           mDateTime   dist year
#> lat   2015-04-09 12:31:55 39.23412 -123.1736 2015-04-09 12:31:27 3.0000 2015
#> lat1  2015-04-09 12:32:53 39.23233 -123.2084 2015-04-09 12:32:24 3.0000 2015
#> lat2  2015-04-09 12:33:54 39.23303 -123.2432 2015-04-09 12:33:23 3.0000 2015
#> lat3  2015-04-09 12:34:53 39.23308 -123.2781 2015-04-09 12:34:23 3.0000 2015
#> lat4  2015-04-09 12:35:54 39.23317 -123.3129 2015-04-09 12:35:23 3.0000 2015
#> lat5  2015-04-09 12:36:54 39.23345 -123.3478 2015-04-09 12:36:24 3.0000 2015
#> lat6  2015-04-09 12:38:00 39.23425 -123.3849 2015-04-09 12:37:27 3.4029 2015
#> lat11 2015-04-09 12:40:59 39.23333 -123.4916 2015-04-09 12:40:29 3.0000 2015
#> lat12 2015-04-09 12:41:42 39.23305 -123.5213 2015-04-09 12:41:21 2.1215 2015
#> lat21 2015-04-09 12:42:43 39.23251 -123.5511 2015-04-09 12:42:12 3.0000 2015
#> lat31 2015-04-09 12:43:43 39.23249 -123.5859 2015-04-09 12:43:13 3.0000 2015
#> lat13 2015-04-09 12:47:51 39.21892 -123.5905 2015-04-09 12:47:21 3.0000 2015
#> lat14 2015-04-09 12:48:51 39.21771 -123.5557 2015-04-09 12:48:21 3.0000 2015
#> lat22 2015-04-09 12:49:51 39.21723 -123.5208 2015-04-09 12:49:21 3.0000 2015
#> lat32 2015-04-09 12:50:53 39.21661 -123.4860 2015-04-09 12:50:22 3.0000 2015
#> lat41 2015-04-09 12:51:51 39.21743 -123.4512 2015-04-09 12:51:22 3.0000 2015
#> lat51 2015-04-09 12:52:52 39.21700 -123.4164 2015-04-09 12:52:22 3.0000 2015
#> lat61 2015-04-09 12:53:55 39.21703 -123.3816 2015-04-09 12:53:24 3.0000 2015
#> lat7  2015-04-09 12:54:58 39.21671 -123.3467 2015-04-09 12:54:27 3.0000 2015
#> lat8  2015-04-09 12:55:31 39.21691 -123.3203 2015-04-09 12:55:14 1.5482 2015
#> lat9  2015-04-09 12:56:33 39.21703 -123.2939 2015-04-09 12:56:02 3.0000 2015
#> lat10 2015-04-09 12:57:30 39.21700 -123.2591 2015-04-09 12:57:01 3.0000 2015
#>       month day    mtime   avgBft avgCCover
#> lat       4   9 12:31:27 1.000000  10.00000
#> lat1      4   9 12:32:24 1.000000  10.00000
#> lat2      4   9 12:33:23 1.000000  10.00000
#> lat3      4   9 12:34:23 1.000000  10.00000
#> lat4      4   9 12:35:23 1.000000  10.00000
#> lat5      4   9 12:36:24 1.000000  10.00000
#> lat6      4   9 12:37:27 1.033724  10.00000
#> lat11     4   9 12:40:29 2.000000  15.02091
#> lat12     4   9 12:41:21 1.216574  20.00000
#> lat21     4   9 12:42:12 1.000000  20.00000
#> lat31     4   9 12:43:13 1.000000  20.00000
#> lat13     4   9 12:47:21 1.000000  20.00000
#> lat14     4   9 12:48:21 1.000000  20.00000
#> lat22     4   9 12:49:21 1.000000  20.00000
#> lat32     4   9 12:50:22 1.000000  20.00000
#> lat41     4   9 12:51:22 1.000000  20.00000
#> lat51     4   9 12:52:22 1.000000  20.00000
#> lat61     4   9 12:53:24 1.000000  20.00000
#> lat7      4   9 12:54:27 1.895145  20.00000
#> lat8      4   9 12:55:14 2.000000  20.00000
#> lat9      4   9 12:56:02 2.000000  20.00000
#> lat10     4   9 12:57:01 2.000000  20.00000
#> 
#> $sightinfo
#>    segnum     mlat      mlon Event            DateTime      Lat       Lon
#> 1       1 39.23412 -123.1736     S 2015-04-09 12:31:35 39.23367 -123.1783
#> 2       1 39.23412 -123.1736     S 2015-04-09 12:31:47 39.23350 -123.1857
#> 3       2 39.23233 -123.2084     S 2015-04-09 12:31:56 39.23333 -123.1917
#> 4       2 39.23233 -123.2084     S 2015-04-09 12:32:20 39.23233 -123.2053
#> 5       2 39.23233 -123.2084     S 2015-04-09 12:32:30 39.23233 -123.2108
#> 6       2 39.23233 -123.2084     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 7       2 39.23233 -123.2084     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 8       3 39.23303 -123.2432     S 2015-04-09 12:33:43 39.23333 -123.2550
#> 9       5 39.23317 -123.3129     S 2015-04-09 12:35:13 39.23317 -123.3073
#> 10      6 39.23345 -123.3478     S 2015-04-09 12:36:13 39.23333 -123.3418
#> 11      6 39.23345 -123.3478     S 2015-04-09 12:36:26 39.23350 -123.3500
#> 12      6 39.23345 -123.3478     t 2015-04-09 12:36:26 39.23350 -123.3500
#> 13      6 39.23345 -123.3478     S 2015-04-09 12:36:43 39.23350 -123.3592
#> 14      7 39.23425 -123.3849     S 2015-04-09 12:37:01 39.23400 -123.3697
#> 15      7 39.23425 -123.3849     S 2015-04-09 12:37:16 39.23433 -123.3790
#> 16      7 39.23425 -123.3849     S 2015-04-09 12:37:37 39.23417 -123.3910
#> 17     10 39.23251 -123.5511     S 2015-04-09 12:42:13 39.23250 -123.5515
#> 18     16 39.21743 -123.4512     S 2015-04-09 12:51:40 39.21750 -123.4410
#> 19     17 39.21700 -123.4164     S 2015-04-09 12:51:55 39.21750 -123.4317
#> 20     17 39.21700 -123.4164     S 2015-04-09 12:52:03 39.21733 -123.4270
#> 21     17 39.21700 -123.4164     S 2015-04-09 12:52:04 39.21733 -123.4270
#> 22     17 39.21700 -123.4164     S 2015-04-09 12:52:07 39.21733 -123.4258
#> 23     17 39.21700 -123.4164     S 2015-04-09 12:52:17 39.21700 -123.4185
#> 24     18 39.21703 -123.3816     S 2015-04-09 12:52:54 39.21700 -123.3982
#> 25     19 39.21671 -123.3467     S 2015-04-09 12:54:34 39.21667 -123.3422
#> 26     20 39.21691 -123.3203     S 2015-04-09 12:55:20 39.21700 -123.3172
#> 27     22 39.21700 -123.2591     S 2015-04-09 12:56:49 39.21717 -123.2685
#> 28     22 39.21700 -123.2591     s 2015-04-09 12:57:00 39.21700 -123.2617
#>    OnEffort Trans Bft CCover Jelly HorizSun VertSun HKR  Haze  Kelp RedTide
#> 1      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 2      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 3      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 4      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 5      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 6      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 7      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 8      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 9      TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 10     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 11     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 12     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 13     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 14     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 15     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 16     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 17     TRUE    T1   1     20     0        6      NA   n FALSE FALSE   FALSE
#> 18     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 19     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 20     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 21     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 22     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 23     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 24     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 25     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 26     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 27     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#> 28     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#>    AltFt SpKnot ObsL ObsB ObsR Rec VLI VLO VB VRI VRO EffortDot EventNum
#> 1    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE        9
#> 2    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       10
#> 3    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       12
#> 4    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       14
#> 5    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       15
#> 6    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 7    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 8    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       18
#> 9    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       23
#> 10   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       26
#> 11   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       27
#> 12   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       28
#> 13   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       29
#> 14   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       31
#> 15   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       32
#> 16   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       34
#> 17   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       46
#> 18   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       65
#> 19   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       66
#> 20   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       68
#> 21   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       69
#> 22   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       70
#> 23   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       71
#> 24   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       73
#> 25   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       78
#> 26   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       80
#> 27   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       84
#> 28   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       86
#>             file_das line_num file_type SightNo  Obs Angle ObsStd SightStd
#> 1  airdas_sample.das        9    turtle       1   cc    67   TRUE     TRUE
#> 2  airdas_sample.das       10    turtle       2   bb   -60   TRUE     TRUE
#> 3  airdas_sample.das       12    turtle       3   cc    38   TRUE     TRUE
#> 4  airdas_sample.das       14    turtle       4   aa   -50   TRUE     TRUE
#> 5  airdas_sample.das       15    turtle       5   cc    49   TRUE     TRUE
#> 6  airdas_sample.das       16    turtle       6   cc    42   TRUE     TRUE
#> 7  airdas_sample.das       16    turtle       6   cc    42   TRUE     TRUE
#> 8  airdas_sample.das       19    turtle       7   aa   -53   TRUE     TRUE
#> 9  airdas_sample.das       24    turtle       8   dd    24  FALSE    FALSE
#> 10 airdas_sample.das       27    turtle       9   aa   -57   TRUE     TRUE
#> 11 airdas_sample.das       28    turtle      10   cc    65   TRUE     TRUE
#> 12 airdas_sample.das       29    turtle    <NA>   aa   -20   TRUE     TRUE
#> 13 airdas_sample.das       30    turtle      11   bb    87   TRUE     TRUE
#> 14 airdas_sample.das       32    turtle      12   bb    82   TRUE     TRUE
#> 15 airdas_sample.das       33    turtle      13   bb    54   TRUE     TRUE
#> 16 airdas_sample.das       35    turtle      14   cc    62   TRUE     TRUE
#> 17 airdas_sample.das       47    turtle      15   bb    50   TRUE     TRUE
#> 18 airdas_sample.das       66    turtle      16   aa   -31   TRUE     TRUE
#> 19 airdas_sample.das       67    turtle      17   bb   -59  FALSE    FALSE
#> 20 airdas_sample.das       69    turtle      18   aa   -62   TRUE     TRUE
#> 21 airdas_sample.das       70    turtle      19   cc    42   TRUE     TRUE
#> 22 airdas_sample.das       71    turtle      20   aa   -48   TRUE     TRUE
#> 23 airdas_sample.das       72    turtle      21   cc    63   TRUE     TRUE
#> 24 airdas_sample.das       74    turtle      22   cc    43   TRUE     TRUE
#> 25 airdas_sample.das       79    turtle      23   cc    56   TRUE     TRUE
#> 26 airdas_sample.das       81    turtle      24   cc    37   TRUE     TRUE
#> 27 airdas_sample.das       85    turtle      25   bb   -76  FALSE    FALSE
#> 28 airdas_sample.das       87    turtle      25 <NA>   -70     NA       NA
#>    Mixed SpCode GsTotal GsSp TurtleSize TurtleDirection TurtleTail included
#> 1  FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 2  FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 3  FALSE     mn       2    2         NA              NA       <NA>     TRUE
#> 4  FALSE     mn       5    5         NA              NA       <NA>     TRUE
#> 5  FALSE     bm       4    4         NA              NA       <NA>     TRUE
#> 6   TRUE     gg      10    8         NA              NA       <NA>     TRUE
#> 7   TRUE     er      10    2         NA              NA       <NA>     TRUE
#> 8  FALSE     bm       5    5         NA              NA       <NA>     TRUE
#> 9  FALSE     bm       2    2         NA              NA       <NA>    FALSE
#> 10 FALSE     mn       4    4         NA              NA       <NA>     TRUE
#> 11 FALSE     gg       5    5         NA              NA       <NA>     TRUE
#> 12 FALSE     dc       1    1          5              90          n     TRUE
#> 13 FALSE     pp       3    3         NA              NA       <NA>     TRUE
#> 14 FALSE     gg       5    5         NA              NA       <NA>     TRUE
#> 15 FALSE     pp       6    6         NA              NA       <NA>     TRUE
#> 16 FALSE     gg       2    2         NA              NA       <NA>     TRUE
#> 17 FALSE     mn       3    3         NA              NA       <NA>     TRUE
#> 18 FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 19 FALSE     bm       6    6         NA              NA       <NA>    FALSE
#> 20 FALSE     gg       2    2         NA              NA       <NA>     TRUE
#> 21 FALSE     bm       3    3         NA              NA       <NA>     TRUE
#> 22 FALSE     bm       4    4         NA              NA       <NA>     TRUE
#> 23 FALSE     pp       6    6         NA              NA       <NA>     TRUE
#> 24 FALSE     pp       2    2         NA              NA       <NA>     TRUE
#> 25 FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 26 FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 27 FALSE     pp       1    1         NA              NA       <NA>    FALSE
#> 28    NA   <NA>      NA   NA         NA              NA       <NA>    FALSE
#> 
#> $randpicks
#>   effort_section randpicks
#> 1              1         7
#> 2              2         2
#> 3              3         9
#> 

airdas_effort(y.proc, method = "section", num.cores = 1)
#> $segdata
#>      segnum section_id section_sub_id event transect              file stlin
#> lat       1          1              1     T       T1 airdas_sample.das     3
#> lat1      2          2              1     R       T1 airdas_sample.das    40
#> lat2      3          3              1     T       T2 airdas_sample.das    53
#>      endlin    lat1      lon1           DateTime1     lat2      lon2
#> lat      37 39.2355 -123.1563 2015-04-09 12:30:59 39.23400 -123.4047
#> lat1     49 39.2340 -123.4742 2015-04-09 12:40:00 39.23283 -123.6032
#> lat2     88 39.2200 -123.6078 2015-04-09 12:46:51 39.21700 -123.2418
#>                DateTime2     mlat      mlon           mDateTime    dist year
#> lat  2015-04-09 12:38:00 39.23309 -123.2804 2015-04-09 12:34:29 21.4029 2015
#> lat1 2015-04-09 12:43:43 39.23275 -123.5388 2015-04-09 12:41:51 11.1215 2015
#> lat2 2015-04-09 12:57:30 39.21729 -123.4248 2015-04-09 12:52:10 31.5482 2015
#>      month day    mtime   avgBft avgCCover  avgJelly avgHorizSun avgVertSun
#> lat      4   9 12:34:29 1.005362   10.0000 0.8757162           6         NA
#> lat1     4   9 12:41:51 1.311061   18.6569 0.0000000           6         NA
#> lat2     4   9 12:52:10 1.324381   20.0000 0.0000000          12         NA
#>      avgHaze avgKelp avgRedTide avgAltFt avgSpKnot
#> lat        0       0 0.43785812 650.0000 100.00000
#> lat1       0       0 0.00000000 650.0000 100.00000
#> lat2       0       0 0.08556751 669.8738  93.50289
#> 
#> $sightinfo
#>    segnum     mlat      mlon Event            DateTime      Lat       Lon
#> 1       1 39.23309 -123.2804     S 2015-04-09 12:31:35 39.23367 -123.1783
#> 2       1 39.23309 -123.2804     S 2015-04-09 12:31:47 39.23350 -123.1857
#> 3       1 39.23309 -123.2804     S 2015-04-09 12:31:56 39.23333 -123.1917
#> 4       1 39.23309 -123.2804     S 2015-04-09 12:32:20 39.23233 -123.2053
#> 5       1 39.23309 -123.2804     S 2015-04-09 12:32:30 39.23233 -123.2108
#> 6       1 39.23309 -123.2804     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 7       1 39.23309 -123.2804     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 8       1 39.23309 -123.2804     S 2015-04-09 12:33:43 39.23333 -123.2550
#> 9       1 39.23309 -123.2804     S 2015-04-09 12:35:13 39.23317 -123.3073
#> 10      1 39.23309 -123.2804     S 2015-04-09 12:36:13 39.23333 -123.3418
#> 11      1 39.23309 -123.2804     S 2015-04-09 12:36:26 39.23350 -123.3500
#> 12      1 39.23309 -123.2804     t 2015-04-09 12:36:26 39.23350 -123.3500
#> 13      1 39.23309 -123.2804     S 2015-04-09 12:36:43 39.23350 -123.3592
#> 14      1 39.23309 -123.2804     S 2015-04-09 12:37:01 39.23400 -123.3697
#> 15      1 39.23309 -123.2804     S 2015-04-09 12:37:16 39.23433 -123.3790
#> 16      1 39.23309 -123.2804     S 2015-04-09 12:37:37 39.23417 -123.3910
#> 17      2 39.23275 -123.5388     S 2015-04-09 12:42:13 39.23250 -123.5515
#> 18      3 39.21729 -123.4248     S 2015-04-09 12:51:40 39.21750 -123.4410
#> 19      3 39.21729 -123.4248     S 2015-04-09 12:51:55 39.21750 -123.4317
#> 20      3 39.21729 -123.4248     S 2015-04-09 12:52:03 39.21733 -123.4270
#> 21      3 39.21729 -123.4248     S 2015-04-09 12:52:04 39.21733 -123.4270
#> 22      3 39.21729 -123.4248     S 2015-04-09 12:52:07 39.21733 -123.4258
#> 23      3 39.21729 -123.4248     S 2015-04-09 12:52:17 39.21700 -123.4185
#> 24      3 39.21729 -123.4248     S 2015-04-09 12:52:54 39.21700 -123.3982
#> 25      3 39.21729 -123.4248     S 2015-04-09 12:54:34 39.21667 -123.3422
#> 26      3 39.21729 -123.4248     S 2015-04-09 12:55:20 39.21700 -123.3172
#> 27      3 39.21729 -123.4248     S 2015-04-09 12:56:49 39.21717 -123.2685
#> 28      3 39.21729 -123.4248     s 2015-04-09 12:57:00 39.21700 -123.2617
#>    OnEffort Trans Bft CCover Jelly HorizSun VertSun HKR  Haze  Kelp RedTide
#> 1      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 2      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 3      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 4      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 5      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 6      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 7      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 8      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 9      TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 10     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 11     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 12     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 13     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 14     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 15     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 16     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 17     TRUE    T1   1     20     0        6      NA   n FALSE FALSE   FALSE
#> 18     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 19     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 20     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 21     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 22     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 23     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 24     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 25     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 26     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 27     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#> 28     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#>    AltFt SpKnot ObsL ObsB ObsR Rec VLI VLO VB VRI VRO EffortDot EventNum
#> 1    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE        9
#> 2    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       10
#> 3    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       12
#> 4    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       14
#> 5    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       15
#> 6    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 7    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 8    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       18
#> 9    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       23
#> 10   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       26
#> 11   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       27
#> 12   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       28
#> 13   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       29
#> 14   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       31
#> 15   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       32
#> 16   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       34
#> 17   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       46
#> 18   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       65
#> 19   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       66
#> 20   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       68
#> 21   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       69
#> 22   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       70
#> 23   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       71
#> 24   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       73
#> 25   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       78
#> 26   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       80
#> 27   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       84
#> 28   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       86
#>             file_das line_num file_type SightNo  Obs Angle ObsStd SightStd
#> 1  airdas_sample.das        9    turtle       1   cc    67   TRUE     TRUE
#> 2  airdas_sample.das       10    turtle       2   bb   -60   TRUE     TRUE
#> 3  airdas_sample.das       12    turtle       3   cc    38   TRUE     TRUE
#> 4  airdas_sample.das       14    turtle       4   aa   -50   TRUE     TRUE
#> 5  airdas_sample.das       15    turtle       5   cc    49   TRUE     TRUE
#> 6  airdas_sample.das       16    turtle       6   cc    42   TRUE     TRUE
#> 7  airdas_sample.das       16    turtle       6   cc    42   TRUE     TRUE
#> 8  airdas_sample.das       19    turtle       7   aa   -53   TRUE     TRUE
#> 9  airdas_sample.das       24    turtle       8   dd    24  FALSE    FALSE
#> 10 airdas_sample.das       27    turtle       9   aa   -57   TRUE     TRUE
#> 11 airdas_sample.das       28    turtle      10   cc    65   TRUE     TRUE
#> 12 airdas_sample.das       29    turtle    <NA>   aa   -20   TRUE     TRUE
#> 13 airdas_sample.das       30    turtle      11   bb    87   TRUE     TRUE
#> 14 airdas_sample.das       32    turtle      12   bb    82   TRUE     TRUE
#> 15 airdas_sample.das       33    turtle      13   bb    54   TRUE     TRUE
#> 16 airdas_sample.das       35    turtle      14   cc    62   TRUE     TRUE
#> 17 airdas_sample.das       47    turtle      15   bb    50   TRUE     TRUE
#> 18 airdas_sample.das       66    turtle      16   aa   -31   TRUE     TRUE
#> 19 airdas_sample.das       67    turtle      17   bb   -59  FALSE    FALSE
#> 20 airdas_sample.das       69    turtle      18   aa   -62   TRUE     TRUE
#> 21 airdas_sample.das       70    turtle      19   cc    42   TRUE     TRUE
#> 22 airdas_sample.das       71    turtle      20   aa   -48   TRUE     TRUE
#> 23 airdas_sample.das       72    turtle      21   cc    63   TRUE     TRUE
#> 24 airdas_sample.das       74    turtle      22   cc    43   TRUE     TRUE
#> 25 airdas_sample.das       79    turtle      23   cc    56   TRUE     TRUE
#> 26 airdas_sample.das       81    turtle      24   cc    37   TRUE     TRUE
#> 27 airdas_sample.das       85    turtle      25   bb   -76  FALSE    FALSE
#> 28 airdas_sample.das       87    turtle      25 <NA>   -70     NA       NA
#>    Mixed SpCode GsTotal GsSp TurtleSize TurtleDirection TurtleTail included
#> 1  FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 2  FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 3  FALSE     mn       2    2         NA              NA       <NA>     TRUE
#> 4  FALSE     mn       5    5         NA              NA       <NA>     TRUE
#> 5  FALSE     bm       4    4         NA              NA       <NA>     TRUE
#> 6   TRUE     gg      10    8         NA              NA       <NA>     TRUE
#> 7   TRUE     er      10    2         NA              NA       <NA>     TRUE
#> 8  FALSE     bm       5    5         NA              NA       <NA>     TRUE
#> 9  FALSE     bm       2    2         NA              NA       <NA>    FALSE
#> 10 FALSE     mn       4    4         NA              NA       <NA>     TRUE
#> 11 FALSE     gg       5    5         NA              NA       <NA>     TRUE
#> 12 FALSE     dc       1    1          5              90          n     TRUE
#> 13 FALSE     pp       3    3         NA              NA       <NA>     TRUE
#> 14 FALSE     gg       5    5         NA              NA       <NA>     TRUE
#> 15 FALSE     pp       6    6         NA              NA       <NA>     TRUE
#> 16 FALSE     gg       2    2         NA              NA       <NA>     TRUE
#> 17 FALSE     mn       3    3         NA              NA       <NA>     TRUE
#> 18 FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 19 FALSE     bm       6    6         NA              NA       <NA>    FALSE
#> 20 FALSE     gg       2    2         NA              NA       <NA>     TRUE
#> 21 FALSE     bm       3    3         NA              NA       <NA>     TRUE
#> 22 FALSE     bm       4    4         NA              NA       <NA>     TRUE
#> 23 FALSE     pp       6    6         NA              NA       <NA>     TRUE
#> 24 FALSE     pp       2    2         NA              NA       <NA>     TRUE
#> 25 FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 26 FALSE     mn       6    6         NA              NA       <NA>     TRUE
#> 27 FALSE     pp       1    1         NA              NA       <NA>    FALSE
#> 28    NA   <NA>      NA   NA         NA              NA       <NA>    FALSE
#> 
#> $randpicks
#>   effort_section randpicks
#> 1              1        NA
#> 2              2        NA
#> 3              3        NA
#>