Summarize AirDAS effort

Chop AirDAS data into effort segments

airdas_effort(x, ...)

# S3 method for data.frame
airdas_effort(x, ...)

# S3 method for airdas_df
airdas_effort(
  x,
  method,
  sp.codes,
  conditions = NULL,
  dist.method = "greatcircle",
  num.cores = NULL,
  ...
)

Arguments

x
airdas_df object; output from airdas_process, or a data frame that can be coerced to a airdas_df object

...
arguments passed to the chopping function specified using method, such as seg.km or seg.min.km

method
character; method to use to chop AirDAS data into effort segments Can be "condition", "equallength", or "section" (case-sensitive) to use airdas_chop_condition, airdas_chop_equal, or airdas_chop_section, respectively

sp.codes
character; species code(s) to include in segdata. These code(s) will be converted to lower case to match airdas_sight

conditions
character vector of names of conditions to include in segdata output. These values must be column names from the output of airdas_process, e.g. 'Bft', 'CCover', etc. The default is NULL, in which case all relevant conditions will be included. If method == "condition", then these also are the conditions which trigger segment chopping when they change.

dist.method
character; method to use to calculate distance between lat/lon coordinates. Can be "greatcircle" to use the great circle distance method (TODO - add ref), or one of "lawofcosines", "haversine", or "vincenty" to use distance. Default is "vincenty"

num.cores
Number of CPUs to over which to distribute computations. Defaults to NULL, which uses one fewer than the number of cores reported by detectCores Using 1 core likely will be faster for smaller datasets

Value

List of three data frames:

segdata: one row for every segment, and columns for information including unique segment number, event code that started the associated continuous effort section, start/end/midpoint coordinates, conditions (e.g. Beaufort), and number of sightings and number of animals on that segment for every species indicated in sp.codes.

siteinfo: details for all sightings in x, including: the unique segment number it is associated with, segment mid points (lat/lon), and whether the sighting was included in the segdata counts (column included), in addition to the other output information described in airdas_sight.

randpicks: see airdas_chop_equal. NULL if using "condition" method.

Details

This is the top-level function for chopping processed AirDAS data into modeling segments (henceforth 'segments'), and assigning sightings and related information (e.g., weather conditions) to each segment. This function returns data frames with all relevant information for the effort segments and associated sightings ('segdata' and 'siteinfo', respectively). Before chopping, the AirDAS data is filtered for events (rows) where either the 'OnEffort' column is TRUE or the 'Event' column is "E" or "O". In other words, the data is filtered for continuous effort sections (henceforth 'effort sections'), where effort sections run from "T"/"R" to "E"/"O" events (inclusive), and then passed to the chopping function specified using method. All on effort events must not have NA Lat or Lon values; note Lat/Lon values for 1 events were 'filled in' in airdas_process.

The following chopping methods are currently available: "condition", "equallength", and "section". When using the "condition" method, effort sections are chopped into segments every time a condition changes, thereby ensuring that the conditions are consistent across the entire segment. See airdas_chop_condition for more details about this method, including arguments that must be passed to it via ....

The "equallength" method consists of chopping effort sections into equal-length segments of length seg.km, and doing a weighted average of the conditions for the length of that segment. See airdas_chop_equal for more details about this method, including arguments that must be passed to it via ....

The "section" method involves 'chopping' the effort into continuous effort sections, i.e. each continuous effort section is a single effort segment. See airdas_chop_section for more details about this method.

The sightings included in the segdata counts are filtered as follows: Beaufort is less than or equal to five, the declination angle is less than 78, it is a standard sighting, and the sighting was made while on effort. Included sightings are those with a TRUE value in the 'included' column in siteinfo (described below). TODO: Allow user to specify this.

The distance between the lat/lon points of subsequent events is calculated using the method specified in dist.method See airdas_sight for how the sightings are processed.

Examples

y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.proc <- airdas_process(y)

# Using "condition" method
airdas_effort(
  y.proc, method = "condition", sp.codes = c("mn", "bm"),
  conditions = "Bft", seg.min.km = 0.05, num.cores = 1
)

#> $segdata
#>   segnum section_id section_sub_id event transect              file stlin
#> 1      1          1              1     T       T1 airdas_sample.das     3
#> 2      2          1              2     T       T1 airdas_sample.das    36
#> 3      3          2              1     R       T1 airdas_sample.das    40
#> 4      4          2              2     R       T1 airdas_sample.das    44
#> 5      5          3              1     T       T2 airdas_sample.das    53
#> 6      6          3              2     T       T2 airdas_sample.das    78
#>   endlin     lat1      lon1     lat2      lon2     mlat      mlon    dist
#> 1     36 39.23550 -123.1563 39.23400 -123.4033 39.23309 -123.2797 21.2882
#> 2     37 39.23400 -123.4033 39.23400 -123.4047 39.23400 -123.4040  0.1148
#> 3     44 39.23400 -123.4915 39.23317 -123.5143 39.23323 -123.4942  3.4595
#> 4     49 39.23317 -123.5143 39.23283 -123.6032 39.23245 -123.5588  7.6620
#> 5     78 39.22000 -123.6078 39.21683 -123.3605 39.21667 -123.4842 21.3146
#> 6     88 39.21683 -123.3605 39.21700 -123.2418 39.21700 -123.3011 10.2336
#>             mDateTime year month day    mtime maxdistBft bm_nSI bm_ANI mn_nSI
#> 1 2015-04-09 12:34:28 2015     4   9 12:34:28          1      3     11      5
#> 2 2015-04-09 12:37:58 2015     4   9 12:37:58          2      0      0      0
#> 3 2015-04-09 12:40:34 2015     4   9 12:40:34          2      0      0      0
#> 4 2015-04-09 12:42:25 2015     4   9 12:42:25          1      0      0      1
#> 5 2015-04-09 12:50:27 2015     4   9 12:50:27          1      3     13      1
#> 6 2015-04-09 12:55:47 2015     4   9 12:55:47          2      0      0      2
#>   mn_ANI
#> 1     23
#> 2      0
#> 3      0
#> 4      3
#> 5      6
#> 6     12
#> 
#> $siteinfo
#>    segnum     mlat      mlon Event            DateTime      Lat       Lon
#> 1       1 39.23309 -123.2797     S 2015-04-09 12:31:35 39.23367 -123.1783
#> 2       1 39.23309 -123.2797     S 2015-04-09 12:31:47 39.23350 -123.1857
#> 3       1 39.23309 -123.2797     S 2015-04-09 12:31:56 39.23333 -123.1917
#> 4       1 39.23309 -123.2797     S 2015-04-09 12:32:20 39.23233 -123.2053
#> 5       1 39.23309 -123.2797     S 2015-04-09 12:32:30 39.23233 -123.2108
#> 6       1 39.23309 -123.2797     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 7       1 39.23309 -123.2797     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 8       1 39.23309 -123.2797     S 2015-04-09 12:33:43 39.23333 -123.2550
#> 9       1 39.23309 -123.2797     S 2015-04-09 12:35:13 39.23317 -123.3073
#> 10      1 39.23309 -123.2797     S 2015-04-09 12:36:13 39.23333 -123.3418
#> 11      1 39.23309 -123.2797     S 2015-04-09 12:36:26 39.23350 -123.3500
#> 12      1 39.23309 -123.2797     t 2015-04-09 12:36:26 39.23350 -123.3500
#> 13      1 39.23309 -123.2797     S 2015-04-09 12:36:43 39.23350 -123.3592
#> 14      1 39.23309 -123.2797     S 2015-04-09 12:37:01 39.23400 -123.3697
#> 15      1 39.23309 -123.2797     S 2015-04-09 12:37:16 39.23433 -123.3790
#> 16      1 39.23309 -123.2797     S 2015-04-09 12:37:37 39.23417 -123.3910
#> 17      4 39.23245 -123.5588     S 2015-04-09 12:42:13 39.23250 -123.5515
#> 18      5 39.21667 -123.4842     S 2015-04-09 12:51:40 39.21750 -123.4410
#> 19      5 39.21667 -123.4842     S 2015-04-09 12:51:55 39.21750 -123.4317
#> 20      5 39.21667 -123.4842     S 2015-04-09 12:52:03 39.21733 -123.4270
#> 21      5 39.21667 -123.4842     S 2015-04-09 12:52:04 39.21733 -123.4270
#> 22      5 39.21667 -123.4842     S 2015-04-09 12:52:07 39.21733 -123.4258
#> 23      5 39.21667 -123.4842     S 2015-04-09 12:52:17 39.21700 -123.4185
#> 24      5 39.21667 -123.4842     S 2015-04-09 12:52:54 39.21700 -123.3982
#> 25      6 39.21700 -123.3011     S 2015-04-09 12:54:34 39.21667 -123.3422
#> 26      6 39.21700 -123.3011     S 2015-04-09 12:55:20 39.21700 -123.3172
#> 27      6 39.21700 -123.3011     S 2015-04-09 12:56:49 39.21717 -123.2685
#> 28      6 39.21700 -123.3011     s 2015-04-09 12:57:00 39.21700 -123.2617
#>    OnEffort Trans Bft CCover Jelly HorizSun VertSun HKR  Haze  Kelp RedTide
#> 1      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 2      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 3      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 4      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 5      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 6      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 7      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 8      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 9      TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 10     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 11     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 12     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 13     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 14     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 15     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 16     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 17     TRUE    T1   1     20     0        6      NA   n FALSE FALSE   FALSE
#> 18     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 19     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 20     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 21     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 22     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 23     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 24     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 25     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 26     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 27     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#> 28     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#>    AltFt SpKnot ObsL ObsB ObsR Rec VLI VLO VB VRI VRO EffortDot EventNum
#> 1    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE        9
#> 2    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       10
#> 3    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       12
#> 4    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       14
#> 5    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       15
#> 6    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 7    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 8    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       18
#> 9    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       23
#> 10   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       26
#> 11   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       27
#> 12   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       28
#> 13   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       29
#> 14   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       31
#> 15   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       32
#> 16   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       34
#> 17   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       46
#> 18   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       65
#> 19   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       66
#> 20   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       68
#> 21   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       69
#> 22   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       70
#> 23   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       71
#> 24   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       73
#> 25   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       78
#> 26   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       80
#> 27   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       84
#> 28   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       86
#>             file_das line_num file_type SightNo  Obs Angle SightStd Mixed
#> 1  airdas_sample.das        9    turtle       1   cc    67     TRUE FALSE
#> 2  airdas_sample.das       10    turtle       2   bb   -60     TRUE FALSE
#> 3  airdas_sample.das       12    turtle       3   cc    38     TRUE FALSE
#> 4  airdas_sample.das       14    turtle       4   aa   -50     TRUE FALSE
#> 5  airdas_sample.das       15    turtle       5   cc    49     TRUE FALSE
#> 6  airdas_sample.das       16    turtle       6   cc    42     TRUE  TRUE
#> 7  airdas_sample.das       16    turtle       6   cc    42     TRUE  TRUE
#> 8  airdas_sample.das       19    turtle       7   aa   -53     TRUE FALSE
#> 9  airdas_sample.das       24    turtle       8   dd    24     TRUE FALSE
#> 10 airdas_sample.das       27    turtle       9   aa   -57     TRUE FALSE
#> 11 airdas_sample.das       28    turtle      10   cc    65     TRUE FALSE
#> 12 airdas_sample.das       29    turtle    <NA>   aa   -20     TRUE FALSE
#> 13 airdas_sample.das       30    turtle      11   bb    87     TRUE FALSE
#> 14 airdas_sample.das       32    turtle      12   bb    82     TRUE FALSE
#> 15 airdas_sample.das       33    turtle      13   bb    54     TRUE FALSE
#> 16 airdas_sample.das       35    turtle      14   cc    62     TRUE FALSE
#> 17 airdas_sample.das       47    turtle      15   bb    50     TRUE FALSE
#> 18 airdas_sample.das       66    turtle      16   aa   -31     TRUE FALSE
#> 19 airdas_sample.das       67    turtle      17   bb   -59     TRUE FALSE
#> 20 airdas_sample.das       69    turtle      18   aa   -62     TRUE FALSE
#> 21 airdas_sample.das       70    turtle      19   cc    42     TRUE FALSE
#> 22 airdas_sample.das       71    turtle      20   aa   -48     TRUE FALSE
#> 23 airdas_sample.das       72    turtle      21   cc    63     TRUE FALSE
#> 24 airdas_sample.das       74    turtle      22   cc    43     TRUE FALSE
#> 25 airdas_sample.das       79    turtle      23   cc    56     TRUE FALSE
#> 26 airdas_sample.das       81    turtle      24   cc    37     TRUE FALSE
#> 27 airdas_sample.das       85    turtle      25   bb   -76     TRUE FALSE
#> 28 airdas_sample.das       87    turtle      25 <NA>   -70       NA    NA
#>    GsTotal   Sp GsSp TurtleSizeFt TurtleDirection TurtleTail included
#> 1        6   mn    6           NA              NA       <NA>     TRUE
#> 2        6   mn    6           NA              NA       <NA>     TRUE
#> 3        2   mn    2           NA              NA       <NA>     TRUE
#> 4        5   mn    5           NA              NA       <NA>     TRUE
#> 5        4   bm    4           NA              NA       <NA>     TRUE
#> 6       10   gg    8           NA              NA       <NA>     TRUE
#> 7       10   er    2           NA              NA       <NA>     TRUE
#> 8        5   bm    5           NA              NA       <NA>     TRUE
#> 9        2   bm    2           NA              NA       <NA>     TRUE
#> 10       4   mn    4           NA              NA       <NA>     TRUE
#> 11       5   gg    5           NA              NA       <NA>     TRUE
#> 12       1   dc    1            5              90          n     TRUE
#> 13       3   pp    3           NA              NA       <NA>    FALSE
#> 14       5   gg    5           NA              NA       <NA>    FALSE
#> 15       6   pp    6           NA              NA       <NA>     TRUE
#> 16       2   gg    2           NA              NA       <NA>     TRUE
#> 17       3   mn    3           NA              NA       <NA>     TRUE
#> 18       6   mn    6           NA              NA       <NA>     TRUE
#> 19       6   bm    6           NA              NA       <NA>     TRUE
#> 20       2   gg    2           NA              NA       <NA>     TRUE
#> 21       3   bm    3           NA              NA       <NA>     TRUE
#> 22       4   bm    4           NA              NA       <NA>     TRUE
#> 23       6   pp    6           NA              NA       <NA>     TRUE
#> 24       2   pp    2           NA              NA       <NA>     TRUE
#> 25       6   mn    6           NA              NA       <NA>     TRUE
#> 26       6   mn    6           NA              NA       <NA>     TRUE
#> 27       1   pp    1           NA              NA       <NA>     TRUE
#> 28      NA <NA>   NA           NA              NA       <NA>    FALSE
#> 
#> $randpicks
#> NULL
#> 

# Using "equallength" method
y.rand <- system.file("airdas_sample_randpicks.csv", package = "swfscAirDAS")
airdas_effort(
  y.proc, method = "equallength", sp.codes = c("mn", "bm"),
  conditions = c("Bft", "CCover"),
  seg.km = 3, randpicks.load = y.rand, num.cores = 1
)

#> $segdata
#>    segnum section_id section_sub_id event transect              file stlin
#> 1       1          1              1     T       T1 airdas_sample.das     3
#> 2       2          1              2     T       T1 airdas_sample.das    12
#> 3       3          1              3     T       T1 airdas_sample.das    18
#> 4       4          1              4     T       T1 airdas_sample.das    20
#> 5       5          1              5     T       T1 airdas_sample.das    22
#> 6       6          1              6     T       T1 airdas_sample.das    26
#> 7       7          1              7     T       T1 airdas_sample.das    31
#> 8       8          2              1     R       T1 airdas_sample.das    40
#> 9       9          2              2     R       T1 airdas_sample.das    43
#> 10     10          2              3     R       T1 airdas_sample.das    45
#> 11     11          2              4     R       T1 airdas_sample.das    48
#> 12     12          3              1     T       T2 airdas_sample.das    53
#> 13     13          3              2     T       T2 airdas_sample.das    59
#> 14     14          3              3     T       T2 airdas_sample.das    62
#> 15     15          3              4     T       T2 airdas_sample.das    63
#> 16     16          3              5     T       T2 airdas_sample.das    65
#> 17     17          3              6     T       T2 airdas_sample.das    67
#> 18     18          3              7     T       T2 airdas_sample.das    74
#> 19     19          3              8     T       T2 airdas_sample.das    76
#> 20     20          3              9     T       T2 airdas_sample.das    80
#> 21     21          3             10     T       T2 airdas_sample.das    82
#> 22     22          3             11     T       T2 airdas_sample.das    83
#>    endlin     lat1      lon1     lat2      lon2     mlat      mlon   dist
#> 1      12 39.23550 -123.1563 39.23335 -123.1910 39.23412 -123.1736 3.0000
#> 2      18 39.23335 -123.1910 39.23255 -123.2257 39.23233 -123.2084 3.0000
#> 3      20 39.23255 -123.2257 39.23312 -123.2606 39.23303 -123.2432 3.0000
#> 4      22 39.23312 -123.2606 39.23317 -123.2955 39.23308 -123.2781 3.0000
#> 5      26 39.23317 -123.2955 39.23330 -123.3302 39.23317 -123.3129 3.0000
#> 6      31 39.23330 -123.3302 39.23371 -123.3651 39.23345 -123.3478 3.0000
#> 7      37 39.23371 -123.3651 39.23400 -123.4047 39.23425 -123.3849 3.4029
#> 8      43 39.23400 -123.4915 39.23317 -123.5089 39.23333 -123.4916 3.0000
#> 9      45 39.23317 -123.5089 39.23284 -123.5335 39.23305 -123.5213 2.1215
#> 10     48 39.23284 -123.5335 39.23239 -123.5684 39.23251 -123.5511 3.0000
#> 11     49 39.23239 -123.5684 39.23283 -123.6032 39.23249 -123.5859 3.0000
#> 12     59 39.22000 -123.6078 39.21808 -123.5733 39.21892 -123.5905 3.0000
#> 13     62 39.21808 -123.5733 39.21783 -123.5383 39.21771 -123.5557 3.0000
#> 14     63 39.21783 -123.5383 39.21640 -123.5037 39.21723 -123.5208 3.0000
#> 15     65 39.21640 -123.5037 39.21722 -123.4687 39.21661 -123.4860 3.0000
#> 16     67 39.21722 -123.4687 39.21750 -123.4339 39.21743 -123.4512 3.0000
#> 17     74 39.21750 -123.4339 39.21700 -123.3990 39.21700 -123.4164 3.0000
#> 18     76 39.21700 -123.3990 39.21686 -123.3644 39.21703 -123.3816 3.0000
#> 19     80 39.21686 -123.3644 39.21667 -123.3294 39.21671 -123.3467 3.0000
#> 20     82 39.21667 -123.3294 39.21700 -123.3114 39.21691 -123.3203 1.5482
#> 21     83 39.21700 -123.3114 39.21728 -123.2766 39.21703 -123.2939 3.0000
#> 22     88 39.21728 -123.2766 39.21700 -123.2418 39.21700 -123.2591 3.0000
#>              mDateTime year month day    mtime avgBft avgCCover bm_nSI bm_ANI
#> 1  2015-04-09 12:31:27 2015     4   9 12:31:27   1.00     10.00      0      0
#> 2  2015-04-09 12:32:28 2015     4   9 12:32:28   1.00     10.00      1      4
#> 3  2015-04-09 12:33:30 2015     4   9 12:33:30   1.00     10.00      1      5
#> 4  2015-04-09 12:34:28 2015     4   9 12:34:28   1.00     10.00      0      0
#> 5  2015-04-09 12:35:28 2015     4   9 12:35:28   1.00     10.00      1      2
#> 6  2015-04-09 12:36:30 2015     4   9 12:36:30   1.00     10.00      0      0
#> 7  2015-04-09 12:37:30 2015     4   9 12:37:30   1.03     10.00      0      0
#> 8  2015-04-09 12:40:30 2015     4   9 12:40:30   2.00     15.02      0      0
#> 9  2015-04-09 12:41:30 2015     4   9 12:41:30   1.22     20.00      0      0
#> 10 2015-04-09 12:42:30 2015     4   9 12:42:30   1.00     20.00      0      0
#> 11 2015-04-09 12:43:21 2015     4   9 12:43:21   1.00     20.00      0      0
#> 12 2015-04-09 12:47:25 2015     4   9 12:47:25   1.00     20.00      0      0
#> 13 2015-04-09 12:48:30 2015     4   9 12:48:30   1.00     20.00      0      0
#> 14 2015-04-09 12:49:30 2015     4   9 12:49:30   1.00     20.00      0      0
#> 15 2015-04-09 12:50:30 2015     4   9 12:50:30   1.00     20.00      0      0
#> 16 2015-04-09 12:51:27 2015     4   9 12:51:27   1.00     20.00      0      0
#> 17 2015-04-09 12:52:24 2015     4   9 12:52:24   1.00     20.00      3     13
#> 18 2015-04-09 12:53:27 2015     4   9 12:53:27   1.00     20.00      0      0
#> 19 2015-04-09 12:54:30 2015     4   9 12:54:30   1.90     20.00      0      0
#> 20 2015-04-09 12:55:30 2015     4   9 12:55:30   2.00     20.00      0      0
#> 21 2015-04-09 12:56:20 2015     4   9 12:56:20   2.00     20.00      0      0
#> 22 2015-04-09 12:57:05 2015     4   9 12:57:05   2.00     20.00      0      0
#>    mn_nSI mn_ANI
#> 1       2     12
#> 2       2      7
#> 3       0      0
#> 4       0      0
#> 5       0      0
#> 6       1      4
#> 7       0      0
#> 8       0      0
#> 9       0      0
#> 10      1      3
#> 11      0      0
#> 12      0      0
#> 13      0      0
#> 14      0      0
#> 15      0      0
#> 16      1      6
#> 17      0      0
#> 18      0      0
#> 19      1      6
#> 20      1      6
#> 21      0      0
#> 22      0      0
#> 
#> $siteinfo
#>    segnum     mlat      mlon Event            DateTime      Lat       Lon
#> 1       1 39.23412 -123.1736     S 2015-04-09 12:31:35 39.23367 -123.1783
#> 2       1 39.23412 -123.1736     S 2015-04-09 12:31:47 39.23350 -123.1857
#> 3       2 39.23233 -123.2084     S 2015-04-09 12:31:56 39.23333 -123.1917
#> 4       2 39.23233 -123.2084     S 2015-04-09 12:32:20 39.23233 -123.2053
#> 5       2 39.23233 -123.2084     S 2015-04-09 12:32:30 39.23233 -123.2108
#> 6       2 39.23233 -123.2084     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 7       2 39.23233 -123.2084     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 8       3 39.23303 -123.2432     S 2015-04-09 12:33:43 39.23333 -123.2550
#> 9       5 39.23317 -123.3129     S 2015-04-09 12:35:13 39.23317 -123.3073
#> 10      6 39.23345 -123.3478     S 2015-04-09 12:36:13 39.23333 -123.3418
#> 11      6 39.23345 -123.3478     S 2015-04-09 12:36:26 39.23350 -123.3500
#> 12      6 39.23345 -123.3478     t 2015-04-09 12:36:26 39.23350 -123.3500
#> 13      6 39.23345 -123.3478     S 2015-04-09 12:36:43 39.23350 -123.3592
#> 14      7 39.23425 -123.3849     S 2015-04-09 12:37:01 39.23400 -123.3697
#> 15      7 39.23425 -123.3849     S 2015-04-09 12:37:16 39.23433 -123.3790
#> 16      7 39.23425 -123.3849     S 2015-04-09 12:37:37 39.23417 -123.3910
#> 17     10 39.23251 -123.5511     S 2015-04-09 12:42:13 39.23250 -123.5515
#> 18     16 39.21743 -123.4512     S 2015-04-09 12:51:40 39.21750 -123.4410
#> 19     17 39.21700 -123.4164     S 2015-04-09 12:51:55 39.21750 -123.4317
#> 20     17 39.21700 -123.4164     S 2015-04-09 12:52:03 39.21733 -123.4270
#> 21     17 39.21700 -123.4164     S 2015-04-09 12:52:04 39.21733 -123.4270
#> 22     17 39.21700 -123.4164     S 2015-04-09 12:52:07 39.21733 -123.4258
#> 23     17 39.21700 -123.4164     S 2015-04-09 12:52:17 39.21700 -123.4185
#> 24     18 39.21703 -123.3816     S 2015-04-09 12:52:54 39.21700 -123.3982
#> 25     19 39.21671 -123.3467     S 2015-04-09 12:54:34 39.21667 -123.3422
#> 26     20 39.21691 -123.3203     S 2015-04-09 12:55:20 39.21700 -123.3172
#> 27     22 39.21700 -123.2591     S 2015-04-09 12:56:49 39.21717 -123.2685
#> 28     22 39.21700 -123.2591     s 2015-04-09 12:57:00 39.21700 -123.2617
#>    OnEffort Trans Bft CCover Jelly HorizSun VertSun HKR  Haze  Kelp RedTide
#> 1      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 2      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 3      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 4      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 5      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 6      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 7      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 8      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 9      TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 10     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 11     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 12     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 13     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 14     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 15     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 16     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 17     TRUE    T1   1     20     0        6      NA   n FALSE FALSE   FALSE
#> 18     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 19     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 20     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 21     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 22     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 23     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 24     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 25     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 26     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 27     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#> 28     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#>    AltFt SpKnot ObsL ObsB ObsR Rec VLI VLO VB VRI VRO EffortDot EventNum
#> 1    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE        9
#> 2    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       10
#> 3    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       12
#> 4    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       14
#> 5    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       15
#> 6    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 7    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 8    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       18
#> 9    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       23
#> 10   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       26
#> 11   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       27
#> 12   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       28
#> 13   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       29
#> 14   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       31
#> 15   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       32
#> 16   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       34
#> 17   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       46
#> 18   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       65
#> 19   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       66
#> 20   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       68
#> 21   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       69
#> 22   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       70
#> 23   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       71
#> 24   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       73
#> 25   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       78
#> 26   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       80
#> 27   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       84
#> 28   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       86
#>             file_das line_num file_type SightNo  Obs Angle SightStd Mixed
#> 1  airdas_sample.das        9    turtle       1   cc    67     TRUE FALSE
#> 2  airdas_sample.das       10    turtle       2   bb   -60     TRUE FALSE
#> 3  airdas_sample.das       12    turtle       3   cc    38     TRUE FALSE
#> 4  airdas_sample.das       14    turtle       4   aa   -50     TRUE FALSE
#> 5  airdas_sample.das       15    turtle       5   cc    49     TRUE FALSE
#> 6  airdas_sample.das       16    turtle       6   cc    42     TRUE  TRUE
#> 7  airdas_sample.das       16    turtle       6   cc    42     TRUE  TRUE
#> 8  airdas_sample.das       19    turtle       7   aa   -53     TRUE FALSE
#> 9  airdas_sample.das       24    turtle       8   dd    24     TRUE FALSE
#> 10 airdas_sample.das       27    turtle       9   aa   -57     TRUE FALSE
#> 11 airdas_sample.das       28    turtle      10   cc    65     TRUE FALSE
#> 12 airdas_sample.das       29    turtle    <NA>   aa   -20     TRUE FALSE
#> 13 airdas_sample.das       30    turtle      11   bb    87     TRUE FALSE
#> 14 airdas_sample.das       32    turtle      12   bb    82     TRUE FALSE
#> 15 airdas_sample.das       33    turtle      13   bb    54     TRUE FALSE
#> 16 airdas_sample.das       35    turtle      14   cc    62     TRUE FALSE
#> 17 airdas_sample.das       47    turtle      15   bb    50     TRUE FALSE
#> 18 airdas_sample.das       66    turtle      16   aa   -31     TRUE FALSE
#> 19 airdas_sample.das       67    turtle      17   bb   -59     TRUE FALSE
#> 20 airdas_sample.das       69    turtle      18   aa   -62     TRUE FALSE
#> 21 airdas_sample.das       70    turtle      19   cc    42     TRUE FALSE
#> 22 airdas_sample.das       71    turtle      20   aa   -48     TRUE FALSE
#> 23 airdas_sample.das       72    turtle      21   cc    63     TRUE FALSE
#> 24 airdas_sample.das       74    turtle      22   cc    43     TRUE FALSE
#> 25 airdas_sample.das       79    turtle      23   cc    56     TRUE FALSE
#> 26 airdas_sample.das       81    turtle      24   cc    37     TRUE FALSE
#> 27 airdas_sample.das       85    turtle      25   bb   -76     TRUE FALSE
#> 28 airdas_sample.das       87    turtle      25 <NA>   -70       NA    NA
#>    GsTotal   Sp GsSp TurtleSizeFt TurtleDirection TurtleTail included
#> 1        6   mn    6           NA              NA       <NA>     TRUE
#> 2        6   mn    6           NA              NA       <NA>     TRUE
#> 3        2   mn    2           NA              NA       <NA>     TRUE
#> 4        5   mn    5           NA              NA       <NA>     TRUE
#> 5        4   bm    4           NA              NA       <NA>     TRUE
#> 6       10   gg    8           NA              NA       <NA>     TRUE
#> 7       10   er    2           NA              NA       <NA>     TRUE
#> 8        5   bm    5           NA              NA       <NA>     TRUE
#> 9        2   bm    2           NA              NA       <NA>     TRUE
#> 10       4   mn    4           NA              NA       <NA>     TRUE
#> 11       5   gg    5           NA              NA       <NA>     TRUE
#> 12       1   dc    1            5              90          n     TRUE
#> 13       3   pp    3           NA              NA       <NA>    FALSE
#> 14       5   gg    5           NA              NA       <NA>    FALSE
#> 15       6   pp    6           NA              NA       <NA>     TRUE
#> 16       2   gg    2           NA              NA       <NA>     TRUE
#> 17       3   mn    3           NA              NA       <NA>     TRUE
#> 18       6   mn    6           NA              NA       <NA>     TRUE
#> 19       6   bm    6           NA              NA       <NA>     TRUE
#> 20       2   gg    2           NA              NA       <NA>     TRUE
#> 21       3   bm    3           NA              NA       <NA>     TRUE
#> 22       4   bm    4           NA              NA       <NA>     TRUE
#> 23       6   pp    6           NA              NA       <NA>     TRUE
#> 24       2   pp    2           NA              NA       <NA>     TRUE
#> 25       6   mn    6           NA              NA       <NA>     TRUE
#> 26       6   mn    6           NA              NA       <NA>     TRUE
#> 27       1   pp    1           NA              NA       <NA>     TRUE
#> 28      NA <NA>   NA           NA              NA       <NA>    FALSE
#> 
#> $randpicks
#>   effort_section randpicks
#> 1              1         7
#> 2              2         2
#> 3              3         9
#> 

# Using "section" method
airdas_effort(
  y.proc, method = "section", sp.codes = c("mn", "bm"),
  num.cores = 1
)

#> $segdata
#>   segnum section_id section_sub_id event transect              file stlin
#> 1      1          1              1     T       T1 airdas_sample.das     3
#> 2      2          2              1     R       T1 airdas_sample.das    40
#> 3      3          3              1     T       T2 airdas_sample.das    53
#>   endlin    lat1      lon1     lat2      lon2     mlat      mlon    dist
#> 1     37 39.2355 -123.1563 39.23400 -123.4047 39.23309 -123.2804 21.4029
#> 2     49 39.2340 -123.4915 39.23283 -123.6032 39.23275 -123.5388 11.1215
#> 3     88 39.2200 -123.6078 39.21700 -123.2418 39.21729 -123.4248 31.5482
#>             mDateTime year month day    mtime avgBft avgCCover avgJelly
#> 1 2015-04-09 12:34:29 2015     4   9 12:34:29   1.01     10.00     0.88
#> 2 2015-04-09 12:41:51 2015     4   9 12:41:51   1.31     18.66     0.00
#> 3 2015-04-09 12:52:10 2015     4   9 12:52:10   1.32     20.00     0.00
#>   avgHorizSun avgVertSun avgHaze avgKelp avgRedTide avgAltFt avgSpKnot bm_nSI
#> 1           6         NA       0       0       0.44   650.00     100.0      3
#> 2           6         NA       0       0       0.00   650.00     100.0      0
#> 3          12         NA       0       0       0.09   669.87      93.5      3
#>   bm_ANI mn_nSI mn_ANI
#> 1     11      5     23
#> 2      0      1      3
#> 3     13      3     18
#> 
#> $siteinfo
#>    segnum     mlat      mlon Event            DateTime      Lat       Lon
#> 1       1 39.23309 -123.2804     S 2015-04-09 12:31:35 39.23367 -123.1783
#> 2       1 39.23309 -123.2804     S 2015-04-09 12:31:47 39.23350 -123.1857
#> 3       1 39.23309 -123.2804     S 2015-04-09 12:31:56 39.23333 -123.1917
#> 4       1 39.23309 -123.2804     S 2015-04-09 12:32:20 39.23233 -123.2053
#> 5       1 39.23309 -123.2804     S 2015-04-09 12:32:30 39.23233 -123.2108
#> 6       1 39.23309 -123.2804     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 7       1 39.23309 -123.2804     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 8       1 39.23309 -123.2804     S 2015-04-09 12:33:43 39.23333 -123.2550
#> 9       1 39.23309 -123.2804     S 2015-04-09 12:35:13 39.23317 -123.3073
#> 10      1 39.23309 -123.2804     S 2015-04-09 12:36:13 39.23333 -123.3418
#> 11      1 39.23309 -123.2804     S 2015-04-09 12:36:26 39.23350 -123.3500
#> 12      1 39.23309 -123.2804     t 2015-04-09 12:36:26 39.23350 -123.3500
#> 13      1 39.23309 -123.2804     S 2015-04-09 12:36:43 39.23350 -123.3592
#> 14      1 39.23309 -123.2804     S 2015-04-09 12:37:01 39.23400 -123.3697
#> 15      1 39.23309 -123.2804     S 2015-04-09 12:37:16 39.23433 -123.3790
#> 16      1 39.23309 -123.2804     S 2015-04-09 12:37:37 39.23417 -123.3910
#> 17      2 39.23275 -123.5388     S 2015-04-09 12:42:13 39.23250 -123.5515
#> 18      3 39.21729 -123.4248     S 2015-04-09 12:51:40 39.21750 -123.4410
#> 19      3 39.21729 -123.4248     S 2015-04-09 12:51:55 39.21750 -123.4317
#> 20      3 39.21729 -123.4248     S 2015-04-09 12:52:03 39.21733 -123.4270
#> 21      3 39.21729 -123.4248     S 2015-04-09 12:52:04 39.21733 -123.4270
#> 22      3 39.21729 -123.4248     S 2015-04-09 12:52:07 39.21733 -123.4258
#> 23      3 39.21729 -123.4248     S 2015-04-09 12:52:17 39.21700 -123.4185
#> 24      3 39.21729 -123.4248     S 2015-04-09 12:52:54 39.21700 -123.3982
#> 25      3 39.21729 -123.4248     S 2015-04-09 12:54:34 39.21667 -123.3422
#> 26      3 39.21729 -123.4248     S 2015-04-09 12:55:20 39.21700 -123.3172
#> 27      3 39.21729 -123.4248     S 2015-04-09 12:56:49 39.21717 -123.2685
#> 28      3 39.21729 -123.4248     s 2015-04-09 12:57:00 39.21700 -123.2617
#>    OnEffort Trans Bft CCover Jelly HorizSun VertSun HKR  Haze  Kelp RedTide
#> 1      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 2      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 3      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 4      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 5      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 6      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 7      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 8      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 9      TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 10     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 11     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 12     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 13     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 14     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 15     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 16     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 17     TRUE    T1   1     20     0        6      NA   n FALSE FALSE   FALSE
#> 18     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 19     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 20     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 21     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 22     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 23     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 24     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 25     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 26     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 27     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#> 28     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#>    AltFt SpKnot ObsL ObsB ObsR Rec VLI VLO VB VRI VRO EffortDot EventNum
#> 1    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE        9
#> 2    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       10
#> 3    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       12
#> 4    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       14
#> 5    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       15
#> 6    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 7    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 8    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       18
#> 9    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       23
#> 10   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       26
#> 11   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       27
#> 12   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       28
#> 13   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       29
#> 14   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       31
#> 15   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       32
#> 16   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       34
#> 17   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       46
#> 18   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       65
#> 19   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       66
#> 20   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       68
#> 21   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       69
#> 22   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       70
#> 23   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       71
#> 24   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       73
#> 25   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       78
#> 26   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       80
#> 27   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       84
#> 28   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       86
#>             file_das line_num file_type SightNo  Obs Angle SightStd Mixed
#> 1  airdas_sample.das        9    turtle       1   cc    67     TRUE FALSE
#> 2  airdas_sample.das       10    turtle       2   bb   -60     TRUE FALSE
#> 3  airdas_sample.das       12    turtle       3   cc    38     TRUE FALSE
#> 4  airdas_sample.das       14    turtle       4   aa   -50     TRUE FALSE
#> 5  airdas_sample.das       15    turtle       5   cc    49     TRUE FALSE
#> 6  airdas_sample.das       16    turtle       6   cc    42     TRUE  TRUE
#> 7  airdas_sample.das       16    turtle       6   cc    42     TRUE  TRUE
#> 8  airdas_sample.das       19    turtle       7   aa   -53     TRUE FALSE
#> 9  airdas_sample.das       24    turtle       8   dd    24     TRUE FALSE
#> 10 airdas_sample.das       27    turtle       9   aa   -57     TRUE FALSE
#> 11 airdas_sample.das       28    turtle      10   cc    65     TRUE FALSE
#> 12 airdas_sample.das       29    turtle    <NA>   aa   -20     TRUE FALSE
#> 13 airdas_sample.das       30    turtle      11   bb    87     TRUE FALSE
#> 14 airdas_sample.das       32    turtle      12   bb    82     TRUE FALSE
#> 15 airdas_sample.das       33    turtle      13   bb    54     TRUE FALSE
#> 16 airdas_sample.das       35    turtle      14   cc    62     TRUE FALSE
#> 17 airdas_sample.das       47    turtle      15   bb    50     TRUE FALSE
#> 18 airdas_sample.das       66    turtle      16   aa   -31     TRUE FALSE
#> 19 airdas_sample.das       67    turtle      17   bb   -59     TRUE FALSE
#> 20 airdas_sample.das       69    turtle      18   aa   -62     TRUE FALSE
#> 21 airdas_sample.das       70    turtle      19   cc    42     TRUE FALSE
#> 22 airdas_sample.das       71    turtle      20   aa   -48     TRUE FALSE
#> 23 airdas_sample.das       72    turtle      21   cc    63     TRUE FALSE
#> 24 airdas_sample.das       74    turtle      22   cc    43     TRUE FALSE
#> 25 airdas_sample.das       79    turtle      23   cc    56     TRUE FALSE
#> 26 airdas_sample.das       81    turtle      24   cc    37     TRUE FALSE
#> 27 airdas_sample.das       85    turtle      25   bb   -76     TRUE FALSE
#> 28 airdas_sample.das       87    turtle      25 <NA>   -70       NA    NA
#>    GsTotal   Sp GsSp TurtleSizeFt TurtleDirection TurtleTail included
#> 1        6   mn    6           NA              NA       <NA>     TRUE
#> 2        6   mn    6           NA              NA       <NA>     TRUE
#> 3        2   mn    2           NA              NA       <NA>     TRUE
#> 4        5   mn    5           NA              NA       <NA>     TRUE
#> 5        4   bm    4           NA              NA       <NA>     TRUE
#> 6       10   gg    8           NA              NA       <NA>     TRUE
#> 7       10   er    2           NA              NA       <NA>     TRUE
#> 8        5   bm    5           NA              NA       <NA>     TRUE
#> 9        2   bm    2           NA              NA       <NA>     TRUE
#> 10       4   mn    4           NA              NA       <NA>     TRUE
#> 11       5   gg    5           NA              NA       <NA>     TRUE
#> 12       1   dc    1            5              90          n     TRUE
#> 13       3   pp    3           NA              NA       <NA>    FALSE
#> 14       5   gg    5           NA              NA       <NA>    FALSE
#> 15       6   pp    6           NA              NA       <NA>     TRUE
#> 16       2   gg    2           NA              NA       <NA>     TRUE
#> 17       3   mn    3           NA              NA       <NA>     TRUE
#> 18       6   mn    6           NA              NA       <NA>     TRUE
#> 19       6   bm    6           NA              NA       <NA>     TRUE
#> 20       2   gg    2           NA              NA       <NA>     TRUE
#> 21       3   bm    3           NA              NA       <NA>     TRUE
#> 22       4   bm    4           NA              NA       <NA>     TRUE
#> 23       6   pp    6           NA              NA       <NA>     TRUE
#> 24       2   pp    2           NA              NA       <NA>     TRUE
#> 25       6   mn    6           NA              NA       <NA>     TRUE
#> 26       6   mn    6           NA              NA       <NA>     TRUE
#> 27       1   pp    1           NA              NA       <NA>     TRUE
#> 28      NA <NA>   NA           NA              NA       <NA>    FALSE
#> 
#> $randpicks
#>   effort_section randpicks
#> 1              1        NA
#> 2              2        NA
#> 3              3        NA
#>