Summarize AirDAS sightings by effort segment

Summarize number of sightings and animals for selected species by segment

airdas_effort_sight(x.list, sp.codes, sp.events = c("S", "t"))

Arguments

x.list
list; output of airdas_effort

sp.codes
character; species code(s) to include in segdata. These code(s) will be converted to lower case to match airdas_sight

sp.events
character; event code(s) to include in the sightinfo output. This argument supersedes the 'included' value when determining whether a sighting is included in the segment summaries. Must be one or more of: "S", "t" (case-sensitive). The default is that all of these event codes are kept

Value

A list, identical to x.list except for 1) the nSI and ANI columns added to x.list$segdata, one each for each element of sp.codes, and 2) the included column of x.list$sightinfo, which has been set as FALSE for sightings of species not listed in sp.codes

Details

This function takes the output of airdas_effort and adds columns for the number of sightings (nSI) and number of animals (ANI) for selected species (selected via sp.codes) for each segment to the segdata element of x.list. However, only sightings with an included value of TRUE (included is a column in sightinfo) are included in the summaries. Having this step separate from airdas_effort allows users to personalize the included values as desired for their analysis.

Examples

y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.proc <- airdas_process(y)
y.cond <- airdas_effort(
  y.proc, method = "condition", conditions = "Bft", seg.min.km = 0.05,
  num.cores = 1
)

airdas_effort_sight(y.cond, sp.codes = c("mn", "bm"))

#> $segdata
#>   segnum section_id section_sub_id event transect              file stlin
#> 1      1          1              1     T       T1 airdas_sample.das     3
#> 2      2          1              2     T       T1 airdas_sample.das    36
#> 3      3          2              1     R       T1 airdas_sample.das    40
#> 4      4          2              2     R       T1 airdas_sample.das    44
#> 5      5          3              1     T       T2 airdas_sample.das    53
#> 6      6          3              2     T       T2 airdas_sample.das    78
#>   endlin     lat1      lon1     lat2      lon2     mlat      mlon    dist
#> 1     36 39.23550 -123.1563 39.23400 -123.4033 39.23309 -123.2797 21.2882
#> 2     37 39.23400 -123.4033 39.23400 -123.4047 39.23400 -123.4040  0.1148
#> 3     44 39.23400 -123.4742 39.23317 -123.5143 39.23323 -123.4942  3.4595
#> 4     49 39.23317 -123.5143 39.23283 -123.6032 39.23245 -123.5588  7.6620
#> 5     78 39.22000 -123.6078 39.21683 -123.3605 39.21667 -123.4842 21.3146
#> 6     88 39.21683 -123.3605 39.21700 -123.2418 39.21700 -123.3011 10.2336
#>             mDateTime year month day    mtime maxdistBft nSI_bm ANI_bm nSI_mn
#> 1 2015-04-09 12:34:28 2015     4   9 12:34:28          1      3     11      5
#> 2 2015-04-09 12:37:58 2015     4   9 12:37:58          2      0      0      0
#> 3 2015-04-09 12:40:34 2015     4   9 12:40:34          2      0      0      0
#> 4 2015-04-09 12:42:25 2015     4   9 12:42:25          1      0      0      1
#> 5 2015-04-09 12:50:27 2015     4   9 12:50:27          1      3     13      1
#> 6 2015-04-09 12:55:47 2015     4   9 12:55:47          2      0      0      2
#>   ANI_mn
#> 1     23
#> 2      0
#> 3      0
#> 4      3
#> 5      6
#> 6     12
#> 
#> $sightinfo
#>    segnum     mlat      mlon Event            DateTime      Lat       Lon
#> 1       1 39.23309 -123.2797     S 2015-04-09 12:31:35 39.23367 -123.1783
#> 2       1 39.23309 -123.2797     S 2015-04-09 12:31:47 39.23350 -123.1857
#> 3       1 39.23309 -123.2797     S 2015-04-09 12:31:56 39.23333 -123.1917
#> 4       1 39.23309 -123.2797     S 2015-04-09 12:32:20 39.23233 -123.2053
#> 5       1 39.23309 -123.2797     S 2015-04-09 12:32:30 39.23233 -123.2108
#> 6       1 39.23309 -123.2797     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 7       1 39.23309 -123.2797     S 2015-04-09 12:32:42 39.23233 -123.2195
#> 8       1 39.23309 -123.2797     S 2015-04-09 12:33:43 39.23333 -123.2550
#> 9       1 39.23309 -123.2797     S 2015-04-09 12:35:13 39.23317 -123.3073
#> 10      1 39.23309 -123.2797     S 2015-04-09 12:36:13 39.23333 -123.3418
#> 11      1 39.23309 -123.2797     S 2015-04-09 12:36:26 39.23350 -123.3500
#> 12      1 39.23309 -123.2797     t 2015-04-09 12:36:26 39.23350 -123.3500
#> 13      1 39.23309 -123.2797     S 2015-04-09 12:36:43 39.23350 -123.3592
#> 14      1 39.23309 -123.2797     S 2015-04-09 12:37:01 39.23400 -123.3697
#> 15      1 39.23309 -123.2797     S 2015-04-09 12:37:16 39.23433 -123.3790
#> 16      1 39.23309 -123.2797     S 2015-04-09 12:37:37 39.23417 -123.3910
#> 17      4 39.23245 -123.5588     S 2015-04-09 12:42:13 39.23250 -123.5515
#> 18      5 39.21667 -123.4842     S 2015-04-09 12:51:40 39.21750 -123.4410
#> 19      5 39.21667 -123.4842     S 2015-04-09 12:51:55 39.21750 -123.4317
#> 20      5 39.21667 -123.4842     S 2015-04-09 12:52:03 39.21733 -123.4270
#> 21      5 39.21667 -123.4842     S 2015-04-09 12:52:04 39.21733 -123.4270
#> 22      5 39.21667 -123.4842     S 2015-04-09 12:52:07 39.21733 -123.4258
#> 23      5 39.21667 -123.4842     S 2015-04-09 12:52:17 39.21700 -123.4185
#> 24      5 39.21667 -123.4842     S 2015-04-09 12:52:54 39.21700 -123.3982
#> 25      6 39.21700 -123.3011     S 2015-04-09 12:54:34 39.21667 -123.3422
#> 26      6 39.21700 -123.3011     S 2015-04-09 12:55:20 39.21700 -123.3172
#> 27      6 39.21700 -123.3011     S 2015-04-09 12:56:49 39.21717 -123.2685
#>    OnEffort Trans Bft CCover Jelly HorizSun VertSun HKR  Haze  Kelp RedTide
#> 1      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 2      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 3      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 4      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 5      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 6      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 7      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 8      TRUE    T1   1     10     0        6      NA   n FALSE FALSE   FALSE
#> 9      TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 10     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 11     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 12     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 13     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 14     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 15     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 16     TRUE    T1   1     10     2        6      NA   r FALSE FALSE    TRUE
#> 17     TRUE    T1   1     20     0        6      NA   n FALSE FALSE   FALSE
#> 18     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 19     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 20     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 21     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 22     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 23     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 24     TRUE    T2   1     20     0       12      NA   n FALSE FALSE   FALSE
#> 25     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 26     TRUE    T2   2     20     0       12      NA   n FALSE FALSE   FALSE
#> 27     TRUE    T2   2     20     0       12      NA   r FALSE FALSE    TRUE
#>    AltFt SpKnot ObsL ObsB ObsR Rec VLI VLO VB VRI VRO EffortDot EventNum
#> 1    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE        9
#> 2    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       10
#> 3    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       12
#> 4    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       14
#> 5    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       15
#> 6    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 7    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       16
#> 8    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       18
#> 9    650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       23
#> 10   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       26
#> 11   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       27
#> 12   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       28
#> 13   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       29
#> 14   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       31
#> 15   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       32
#> 16   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       34
#> 17   650    100   aa   bb   cc  dd   g   g  g   g   g      TRUE       46
#> 18   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       65
#> 19   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       66
#> 20   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       68
#> 21   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       69
#> 22   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       70
#> 23   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       71
#> 24   680     90   aa   dd   cc  bb   g   g  g   g   g      TRUE       73
#> 25   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       78
#> 26   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       80
#> 27   680     90   aa   dd   cc  bb   g   g  p   g   g      TRUE       84
#>             file_das line_num file_type SightNo Obs Angle SightStd Mixed SpCode
#> 1  airdas_sample.das        9    turtle       1  cc    67     TRUE FALSE     mn
#> 2  airdas_sample.das       10    turtle       2  bb   -60     TRUE FALSE     mn
#> 3  airdas_sample.das       12    turtle       3  cc    38     TRUE FALSE     mn
#> 4  airdas_sample.das       14    turtle       4  aa   -50     TRUE FALSE     mn
#> 5  airdas_sample.das       15    turtle       5  cc    49     TRUE FALSE     bm
#> 6  airdas_sample.das       16    turtle       6  cc    42     TRUE  TRUE     gg
#> 7  airdas_sample.das       16    turtle       6  cc    42     TRUE  TRUE     er
#> 8  airdas_sample.das       19    turtle       7  aa   -53     TRUE FALSE     bm
#> 9  airdas_sample.das       24    turtle       8  dd    24     TRUE FALSE     bm
#> 10 airdas_sample.das       27    turtle       9  aa   -57     TRUE FALSE     mn
#> 11 airdas_sample.das       28    turtle      10  cc    65     TRUE FALSE     gg
#> 12 airdas_sample.das       29    turtle    <NA>  aa   -20     TRUE FALSE     dc
#> 13 airdas_sample.das       30    turtle      11  bb    87     TRUE FALSE     pp
#> 14 airdas_sample.das       32    turtle      12  bb    82     TRUE FALSE     gg
#> 15 airdas_sample.das       33    turtle      13  bb    54     TRUE FALSE     pp
#> 16 airdas_sample.das       35    turtle      14  cc    62     TRUE FALSE     gg
#> 17 airdas_sample.das       47    turtle      15  bb    50     TRUE FALSE     mn
#> 18 airdas_sample.das       66    turtle      16  aa   -31     TRUE FALSE     mn
#> 19 airdas_sample.das       67    turtle      17  bb   -59     TRUE FALSE     bm
#> 20 airdas_sample.das       69    turtle      18  aa   -62     TRUE FALSE     gg
#> 21 airdas_sample.das       70    turtle      19  cc    42     TRUE FALSE     bm
#> 22 airdas_sample.das       71    turtle      20  aa   -48     TRUE FALSE     bm
#> 23 airdas_sample.das       72    turtle      21  cc    63     TRUE FALSE     pp
#> 24 airdas_sample.das       74    turtle      22  cc    43     TRUE FALSE     pp
#> 25 airdas_sample.das       79    turtle      23  cc    56     TRUE FALSE     mn
#> 26 airdas_sample.das       81    turtle      24  cc    37     TRUE FALSE     mn
#> 27 airdas_sample.das       85    turtle      25  bb   -76     TRUE FALSE     pp
#>    GsTotal GsSp TurtleSize TurtleDirection TurtleTail included
#> 1        6    6         NA              NA       <NA>     TRUE
#> 2        6    6         NA              NA       <NA>     TRUE
#> 3        2    2         NA              NA       <NA>     TRUE
#> 4        5    5         NA              NA       <NA>     TRUE
#> 5        4    4         NA              NA       <NA>     TRUE
#> 6       10    8         NA              NA       <NA>    FALSE
#> 7       10    2         NA              NA       <NA>    FALSE
#> 8        5    5         NA              NA       <NA>     TRUE
#> 9        2    2         NA              NA       <NA>     TRUE
#> 10       4    4         NA              NA       <NA>     TRUE
#> 11       5    5         NA              NA       <NA>    FALSE
#> 12       1    1          5              90          n    FALSE
#> 13       3    3         NA              NA       <NA>    FALSE
#> 14       5    5         NA              NA       <NA>    FALSE
#> 15       6    6         NA              NA       <NA>    FALSE
#> 16       2    2         NA              NA       <NA>    FALSE
#> 17       3    3         NA              NA       <NA>     TRUE
#> 18       6    6         NA              NA       <NA>     TRUE
#> 19       6    6         NA              NA       <NA>     TRUE
#> 20       2    2         NA              NA       <NA>    FALSE
#> 21       3    3         NA              NA       <NA>     TRUE
#> 22       4    4         NA              NA       <NA>     TRUE
#> 23       6    6         NA              NA       <NA>    FALSE
#> 24       2    2         NA              NA       <NA>    FALSE
#> 25       6    6         NA              NA       <NA>     TRUE
#> 26       6    6         NA              NA       <NA>     TRUE
#> 27       1    1         NA              NA       <NA>    FALSE
#> 
#> $randpicks
#> NULL
#>