Aerial DAS format requirements

Access and save local PDF documents describing the data format of the different file types supported by swfscAirDAS

Usage

airdas_format_pdf(file, file.type = c("phocoena", "caretta", "turtle"), ...)

Arguments

file

character; the name of the file where the PDF will be saved

file.type

character; indicates which data format PDF to extract. Must be one of: "turtle", "caretta", "survey", or "phocoena" (case sensitive)

...

passed to file.copy; might included named argument overwrite

Value

output of file.copy: TRUE if writing of file was successful, and FALSE otherwise

Details

This function is a wrapper function for file.copy. It saves a PDF document describing the specified aerial DAS data format requirements by copying the PDF document to file

The PDF files can also be manually copied or downloaded from:

PHOCOENA

Can be copied from: system.file("AirDAS_Format_PHOCOENA.pdf", package = "swfscAirDAS")

Can be downloaded from: https://github.com/swfsc/swfscAirDAS/blob/master/inst/AirDAS_Format_PHOCOENA.pdf

CARETTA

Can be copied from: system.file("AirDAS_Format_CARETTA.pdf", package = "swfscAirDAS")

Can be downloaded from: https://github.com/swfsc/swfscAirDAS/blob/master/inst/AirDAS_Format_CARETTA.pdf

TURTLE

Can be copied from: system.file("AirDAS_Format_TURTLE.pdf", package = "swfscAirDAS")

Can be downloaded from: https://github.com/swfsc/swfscAirDAS/blob/master/inst/AirDAS_Format_TURTLE.pdf

See also

Examples

if (interactive()) {
  airdas_format_pdf(
    "AirDAS_Format_TURTLE.pdf", file.type  = "turtle", 
    overwrite = FALSE
  )
}