Process aerial survey DAS data

Process AirDAS data (the output of airdas_read), including extracting state and condition information for each AirDAS event

Usage

airdas_process(x, ...)

# S3 method for character
airdas_process(x, ...)

# S3 method for data.frame
airdas_process(x, ...)

# S3 method for airdas_dfr
airdas_process(
  x,
  days.gap.part = 0.5/24,
  days.gap.full = 12/24,
  gap.message = FALSE,
  reset.transect = TRUE,
  trans.upper = FALSE,
  ...
)

Arguments

x

an object of class airdas_dfr object, an object that can be coerced to class airdas_dfr, or a character (filepath) which is first passed to airdas_read

...

passed to airdas_read if x is a character. Otherwise ignored

days.gap.part

numeric of length 1; time gap (in days) used to identify when a 'partial reset' is performed, i.e. when propagated info (weather, observers, etc) is reset. Default is 30 minutes; must be less than or equal to days.gap.full

days.gap.full

numeric of length 1; time gap (in days) used to identify when a 'full reset; is performed, i.e. when all info (transect number and propagated info) is reset. Default is 12 hours; must be greater than days.gap.part

gap.message

logical; default is FALSE. Indicates if messages should be printed detailing which row(s) of the output data frame were partially or fully reset

reset.transect

logical; default is TRUE. Indicates if propagated info (weather, observers, etc) should be reset to NA when beginning a new transect. See Details section

trans.upper

logical; indicates if all transect codes should be capitalized using toupper. Default is FALSE

Value

An airdas_df object, which is also a data frame. It consists of the input data frame, i.e. the output of airdas_read, with the following columns added:

State/condition	Column name	Notes

On/off effort	OnEffort	Transect code

Trans	Beaufort sea state	Bft

Percent overcast (cloud cover)	CCover	Jellyfish code

Jelly	not in PHOCOENA data	Horizontal sun (clock system)

HorizSun	Vertical sun (clock system)	VertSun

only in PHOCOENA data	Haze/Kelp/Red tide code	HKR

Haze (from HKR code)	Haze	Kelp (from HKR code)

Kelp	Red tide (from HKR code)	RedTide

Altitude (feet)	AltFt	Speed (knots)

SpKnot	Left observer	ObsL

Belly observer	ObsB	Right observer

ObsR	Data recorder	Rec

Viewing condition - left inside	VLI	Viewing condition - left outside

VLO	Viewing condition - belly	VB

Viewing condition - right inside	VRI	Viewing condition - right outside

See airdas_format_pdf for which data columns the condition information is extracted form for each file type. In addition, warnings are printed with line numbers of unexpected event codes

Details

If x is a character, it is assumed to be a filepath and first passed to airdas_read. This output is then processed.

This function cannot handle concatenated airdas_dfr objects of multiple file types. In other words, AirDAS data must be processed and then concatenated.

AirDAS data is event-based, meaning most events indicate when a state or weather condition changes. For instance, a 'W' event indicates when one or more weather conditions (such as Beaufort sea state) change, and the weather conditions are the same for subsequent events until the next 'W' event. For each state/condition: a new column is created, the state/condition information is extracted from relevant events, and extracted information is propagated to appropriate subsequent rows (events). Thus, each row in the output data frame contains all pertinent state/condition information for that row.

The following assumptions/decisions are made during processing:

All '#' events (deleted events) are removed

'DateTime', 'Lat', and 'Lon' information are added to '1' events where applicable

Effort is determined as follows: T/R events turns effort on, and O/E events turn effort off. T/R events themselves will be on effort, while O/E events will be off effort. The 'EffortDot' column is ignored

'HKR' values are converted to lower case. "Y" values are considered to be "H" values

Observer ('ObsL', 'ObsB', 'ObsR', 'Rec') values are converted to lower case

Viewing condition ('VLI', 'VLO', 'VB', 'VRI', 'VRO') values are converted to lower case

Missing values are NA rather than -1

Normally, a T event (to indicate starting/resuming a transect) is immediately followed by a VPAW event series, creating a TVPAW event series. The reset.transect argument causes the conditions set in the VPAW event series (Beaufort, viewing conditions, altitude, etc.) to be reset to NA at each T event

Examples

y <- system.file("airdas_sample.das", package = "swfscAirDAS")
airdas_process(y, trans.upper = FALSE)
#>    Event            DateTime      Lat       Lon OnEffort Trans Bft CCover Jelly
#> 1      C 2015-04-09 12:30:59 39.23550 -123.1563    FALSE  <NA>  NA     NA    NA
#> 2      C 2015-04-09 12:30:59 39.23550 -123.1563    FALSE  <NA>  NA     NA    NA
#> 3      T 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 4      V 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 5      P 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 6      A 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 7      W 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1   1     10     0
#> 8      * 2015-04-09 12:31:00 39.23567 -123.1575     TRUE    T1   1     10     0
#> 9      S 2015-04-09 12:31:35 39.23367 -123.1783     TRUE    T1   1     10     0
#> 10     S 2015-04-09 12:31:47 39.23350 -123.1857     TRUE    T1   1     10     0
#> 11     C 2015-04-09 12:31:48 39.23350 -123.1857     TRUE    T1   1     10     0
#> 12     S 2015-04-09 12:31:56 39.23333 -123.1917     TRUE    T1   1     10     0
#> 13     * 2015-04-09 12:32:00 39.23317 -123.1930     TRUE    T1   1     10     0
#> 14     S 2015-04-09 12:32:20 39.23233 -123.2053     TRUE    T1   1     10     0
#> 15     S 2015-04-09 12:32:30 39.23233 -123.2108     TRUE    T1   1     10     0
#> 16     S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 17     1 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 18     * 2015-04-09 12:33:00 39.23267 -123.2292     TRUE    T1   1     10     0
#> 19     S 2015-04-09 12:33:43 39.23333 -123.2550     TRUE    T1   1     10     0
#> 20     * 2015-04-09 12:34:00 39.23300 -123.2637     TRUE    T1   1     10     0
#> 21     W 2015-04-09 12:34:52 39.23317 -123.2945     TRUE    T1   1     10     2
#> 22     C 2015-04-09 12:34:56 39.23317 -123.2968     TRUE    T1   1     10     2
#> 23     * 2015-04-09 12:35:00 39.23317 -123.2985     TRUE    T1   1     10     2
#> 24     S 2015-04-09 12:35:13 39.23317 -123.3073     TRUE    T1   1     10     2
#> 25     C 2015-04-09 12:35:32 39.23317 -123.3178     TRUE    T1   1     10     2
#> 26     * 2015-04-09 12:36:00 39.23333 -123.3337     TRUE    T1   1     10     2
#> 27     S 2015-04-09 12:36:13 39.23333 -123.3418     TRUE    T1   1     10     2
#> 28     S 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 29     t 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 30     S 2015-04-09 12:36:43 39.23350 -123.3592     TRUE    T1   1     10     2
#> 31     * 2015-04-09 12:37:01 39.23383 -123.3685     TRUE    T1   1     10     2
#> 32     S 2015-04-09 12:37:01 39.23400 -123.3697     TRUE    T1   1     10     2
#> 33     S 2015-04-09 12:37:16 39.23433 -123.3790     TRUE    T1   1     10     2
#> 34     S 2015-04-09 12:37:37 39.23417 -123.3910     TRUE    T1   1     10     2
#> 35     W 2015-04-09 12:37:57 39.23400 -123.4033     TRUE    T1   2     10     0
#> 36     E 2015-04-09 12:38:00 39.23400 -123.4047    FALSE    T1   2     10     0
#> 37     C 2015-04-09 12:38:00 39.23400 -123.4047    FALSE    T1   2     10     0
#> 38     * 2015-04-09 12:39:00 39.23417 -123.4400    FALSE    T1   2     10     0
#> 39     R 2015-04-09 12:40:00 39.23400 -123.4742     TRUE    T1   2     10     0
#> 40     W 2015-04-09 12:40:28 39.23333 -123.4915     TRUE    T1   2     20     0
#> 41     C 2015-04-09 12:40:36 39.23317 -123.4960     TRUE    T1   2     20     0
#> 42     * 2015-04-09 12:41:00 39.23317 -123.5092     TRUE    T1   2     20     0
#> 43     W 2015-04-09 12:41:08 39.23317 -123.5143     TRUE    T1   1     20     0
#> 44     * 2015-04-09 12:42:00 39.23267 -123.5433     TRUE    T1   1     20     0
#> 45     C 2015-04-09 12:42:01 39.23267 -123.5433     TRUE    T1   1     20     0
#> 46     S 2015-04-09 12:42:13 39.23250 -123.5515     TRUE    T1   1     20     0
#> 47     * 2015-04-09 12:43:00 39.23233 -123.5778     TRUE    T1   1     20     0
#> 48     O 2015-04-09 12:43:43 39.23283 -123.6033    FALSE  <NA>   1     20     0
#> 49     * 2015-04-09 12:44:00 39.23300 -123.6118    FALSE  <NA>   1     20     0
#> 50     * 2015-04-09 12:45:00 39.24450 -123.6415    FALSE  <NA>   1     20     0
#> 51     * 2015-04-09 12:46:00 39.22167 -123.6373    FALSE  <NA>   1     20     0
#> 52     T 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2  NA     NA    NA
#> 53     V 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2  NA     NA    NA
#> 54     P 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2  NA     NA    NA
#> 55     A 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2  NA     NA    NA
#> 56     W 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 57     * 2015-04-09 12:47:00 39.21950 -123.6025     TRUE    T2   1     20     0
#> 58     * 2015-04-09 12:48:00 39.21783 -123.5682     TRUE    T2   1     20     0
#> 59     A 2015-04-09 12:48:29 39.21767 -123.5515     TRUE    T2   1     20     0
#> 60     A 2015-04-09 12:48:36 39.21783 -123.5468     TRUE    T2   1     20     0
#> 61     * 2015-04-09 12:49:00 39.21783 -123.5333     TRUE    T2   1     20     0
#> 62     * 2015-04-09 12:50:00 39.21617 -123.4987     TRUE    T2   1     20     0
#> 63     A 2015-04-09 12:50:32 39.21683 -123.4797     TRUE    T2   1     20     0
#> 64     * 2015-04-09 12:51:00 39.21733 -123.4653     TRUE    T2   1     20     0
#> 65     S 2015-04-09 12:51:40 39.21750 -123.4410     TRUE    T2   1     20     0
#> 66     S 2015-04-09 12:51:55 39.21750 -123.4317     TRUE    T2   1     20     0
#> 67     * 2015-04-09 12:52:00 39.21733 -123.4293     TRUE    T2   1     20     0
#> 68     S 2015-04-09 12:52:03 39.21733 -123.4270     TRUE    T2   1     20     0
#> 69     S 2015-04-09 12:52:04 39.21733 -123.4270     TRUE    T2   1     20     0
#> 70     S 2015-04-09 12:52:07 39.21733 -123.4258     TRUE    T2   1     20     0
#> 71     S 2015-04-09 12:52:17 39.21700 -123.4185     TRUE    T2   1     20     0
#> 72     C 2015-04-09 12:52:39 39.21700 -123.4072     TRUE    T2   1     20     0
#> 73     S 2015-04-09 12:52:54 39.21700 -123.3982     TRUE    T2   1     20     0
#> 74     * 2015-04-09 12:53:00 39.21717 -123.3960     TRUE    T2   1     20     0
#> 75     * 2015-04-09 12:54:00 39.21683 -123.3617     TRUE    T2   1     20     0
#> 76     V 2015-04-09 12:54:01 39.21683 -123.3605     TRUE    T2   1     20     0
#> 77     W 2015-04-09 12:54:04 39.21683 -123.3605     TRUE    T2   2     20     0
#> 78     S 2015-04-09 12:54:34 39.21667 -123.3422     TRUE    T2   2     20     0
#> 79     * 2015-04-09 12:55:00 39.21667 -123.3285     TRUE    T2   2     20     0
#> 80     S 2015-04-09 12:55:20 39.21700 -123.3172     TRUE    T2   2     20     0
#> 81     * 2015-04-09 12:56:00 39.21700 -123.2960     TRUE    T2   2     20     0
#> 82     W 2015-04-09 12:56:40 39.21733 -123.2730     TRUE    T2   2     20     0
#> 83     C 2015-04-09 12:56:43 39.21717 -123.2718     TRUE    T2   2     20     0
#> 84     S 2015-04-09 12:56:49 39.21717 -123.2685     TRUE    T2   2     20     0
#> 85     * 2015-04-09 12:57:00 39.21700 -123.2617     TRUE    T2   2     20     0
#> 86     s 2015-04-09 12:57:00 39.21700 -123.2617     TRUE    T2   2     20     0
#> 87     O 2015-04-09 12:57:30 39.21700 -123.2417    FALSE  <NA>   2     20     0
#>    HorizSun VertSun  HKR  Haze  Kelp RedTide AltFt SpKnot ObsL ObsB ObsR  Rec
#> 1        NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 2        NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 3        NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 4        NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 5        NA      NA <NA> FALSE FALSE   FALSE    NA     NA   aa   bb   cc   dd
#> 6        NA      NA <NA> FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 7         6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 8         6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 9         6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 10        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 11        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 12        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 13        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 14        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 15        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 16        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 17        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 18        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 19        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 20        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 21        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 22        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 23        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 24        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 25        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 26        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 27        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 28        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 29        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 30        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 31        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 32        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 33        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 34        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 35        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 36        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 37        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 38        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 39        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 40        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 41        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 42        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 43        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 44        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 45        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 46        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 47        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 48        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 49        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 50        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 51        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 52       NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 53       NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 54       NA      NA <NA> FALSE FALSE   FALSE    NA     NA   aa   dd   cc   bb
#> 55       NA      NA <NA> FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 56       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 57       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 58       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 59       12      NA    n FALSE FALSE   FALSE   680    100   aa   dd   cc   bb
#> 60       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 61       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 62       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 63       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 64       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 65       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 66       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 67       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 68       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 69       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 70       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 71       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 72       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 73       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 74       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 75       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 76       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 77       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 78       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 79       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 80       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 81       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 82       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 83       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 84       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 85       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 86       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 87       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#>     VLI  VLO   VB  VRI  VRO Data1 Data2 Data3 Data4 Data5 Data6           Data7
#> 1  <NA> <NA> <NA> <NA> <NA>  Reco rder:  dd    <NA>  <NA>  <NA>            <NA>
#> 2  <NA> <NA> <NA> <NA> <NA>  Not  recor ding  molas  <NA>  <NA>            <NA>
#> 3  <NA> <NA> <NA> <NA> <NA>    T1  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 4     g    g    g    g    g     g     g     g     g     g  <NA>            <NA>
#> 5     g    g    g    g    g    aa    bb    cc    dd  <NA>  <NA>            <NA>
#> 6     g    g    g    g    g   650   100  <NA>  <NA>  <NA>  <NA>            <NA>
#> 7     g    g    g    g    g     n    10     1     0     6  <NA>            <NA>
#> 8     g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 9     g    g    g    g    g     1    cc    67     6    mn  <NA>            <NA>
#> 10    g    g    g    g    g     2    bb   -60     6    mn  <NA>            <NA>
#> 11    g    g    g    g    g  2 cp  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 12    g    g    g    g    g     3    cc    38     2    mn  <NA>            <NA>
#> 13    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 14    g    g    g    g    g     4    aa   -50     5    mn  <NA>            <NA>
#> 15    g    g    g    g    g     5    cc    49     4    bm  <NA>            <NA>
#> 16    g    g    g    g    g     6    cc    42    10    gg    er            <NA>
#> 17    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>    80    20            <NA>
#> 18    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 19    g    g    g    g    g     7    aa   -53     5    bm  <NA>            <NA>
#> 20    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 21    g    g    g    g    g     r    10     1     2     6  <NA>            <NA>
#> 22    g    g    g    g    g  stre aky r ed ti de     <NA>  <NA>            <NA>
#> 23    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 24    g    g    g    g    g     8    dd    24     2    bm  <NA>            <NA>
#> 25    g    g    g    g    g  shar k 11f t      <NA>  <NA>  <NA>            <NA>
#> 26    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 27    g    g    g    g    g     9    aa   -57     4    mn  <NA>            <NA>
#> 28    g    g    g    g    g    10    cc    65     5    gg  <NA>            <NA>
#> 29    g    g    g    g    g    aa   -20    dc     5    90     n            <NA>
#> 30    g    g    g    g    g    11    bb    87     3    pp  <NA>            <NA>
#> 31    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 32    g    g    g    g    g    12    bb    82     5    gg  <NA>            <NA>
#> 33    g    g    g    g    g    13    bb    54     6    pp  <NA>            <NA>
#> 34    g    g    g    g    g    14    cc    62     2    gg  <NA>            <NA>
#> 35    g    g    g    g    g     n    10     2     0     6  <NA>            <NA>
#> 36    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 37    g    g    g    g    g  off  effor t to  circl e on  unide ntifed object  
#> 38    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 39    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 40    g    g    g    g    g     n    20     2     0     6  <NA>            <NA>
#> 41    g    g    g    g    g  wate r got  more  blue  <NA>  <NA>            <NA>
#> 42    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 43    g    g    g    g    g     n    20     1     0     6  <NA>            <NA>
#> 44    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 45    g    g    g    g    g  fb2s  fb1m  <NA>  <NA>  <NA>  <NA>            <NA>
#> 46    g    g    g    g    g    15    bb    50     3    mn  <NA>            <NA>
#> 47    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 48    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 49    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 50    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 51    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 52 <NA> <NA> <NA> <NA> <NA>    T2  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 53    g    g    g    g    g     g     g     g     g     g  <NA>            <NA>
#> 54    g    g    g    g    g    aa    dd    cc    bb  <NA>  <NA>            <NA>
#> 55    g    g    g    g    g   650   100  <NA>  <NA>  <NA>  <NA>            <NA>
#> 56    g    g    g    g    g     n    20     1     0    12  <NA>            <NA>
#> 57    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 58    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 59    g    g    g    g    g   680   100  <NA>  <NA>  <NA>  <NA>            <NA>
#> 60    g    g    g    g    g   650   100  <NA>  <NA>  <NA>  <NA>            <NA>
#> 61    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 62    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 63    g    g    g    g    g   680    90  <NA>  <NA>  <NA>  <NA>            <NA>
#> 64    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 65    g    g    g    g    g    16    aa   -31     6    mn  <NA>            <NA>
#> 66    g    g    g    g    g    17    bb   -59     6    bm  <NA>            <NA>
#> 67    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 68    g    g    g    g    g    18    aa   -62     2    gg  <NA>            <NA>
#> 69    g    g    g    g    g    19    cc    42     3    bm  <NA>            <NA>
#> 70    g    g    g    g    g    20    aa   -48     4    bm  <NA>            <NA>
#> 71    g    g    g    g    g    21    cc    63     6    pp  <NA>            <NA>
#> 72    g    g    g    g    g  bell y has  some  glar e      <NA>            <NA>
#> 73    g    g    g    g    g    22    cc    43     2    pp  <NA>            <NA>
#> 74    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 75    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 76    g    g    p    g    g     g     g     p     g     g  <NA>            <NA>
#> 77    g    g    p    g    g     n    20     2     0    12  <NA>            <NA>
#> 78    g    g    p    g    g    23    cc    56     6    mn  <NA>            <NA>
#> 79    g    g    p    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 80    g    g    p    g    g    24    cc    37     6    mn  <NA>            <NA>
#> 81    g    g    p    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 82    g    g    p    g    g     r    20     2     0    12  <NA>            <NA>
#> 83    g    g    p    g    g  patc hy re d tid e      <NA>  <NA>            <NA>
#> 84    g    g    p    g    g    25    bb   -76     1    pp  <NA>            <NA>
#> 85    g    g    p    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 86    g    g    p    g    g    25   -70  <NA>  <NA>  <NA>  <NA>            <NA>
#> 87    g    g    p    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#>    EffortDot EventNum          file_das line_num file_type
#> 1      FALSE        1 airdas_sample.das        1    turtle
#> 2      FALSE        2 airdas_sample.das        2    turtle
#> 3       TRUE        3 airdas_sample.das        3    turtle
#> 4       TRUE        4 airdas_sample.das        4    turtle
#> 5       TRUE        5 airdas_sample.das        5    turtle
#> 6       TRUE        6 airdas_sample.das        6    turtle
#> 7       TRUE        7 airdas_sample.das        7    turtle
#> 8       TRUE        8 airdas_sample.das        8    turtle
#> 9       TRUE        9 airdas_sample.das        9    turtle
#> 10      TRUE       10 airdas_sample.das       10    turtle
#> 11      TRUE       11 airdas_sample.das       11    turtle
#> 12      TRUE       12 airdas_sample.das       12    turtle
#> 13      TRUE       13 airdas_sample.das       13    turtle
#> 14      TRUE       14 airdas_sample.das       14    turtle
#> 15      TRUE       15 airdas_sample.das       15    turtle
#> 16      TRUE       16 airdas_sample.das       16    turtle
#> 17     FALSE     <NA> airdas_sample.das       17    turtle
#> 18      TRUE       17 airdas_sample.das       18    turtle
#> 19      TRUE       18 airdas_sample.das       19    turtle
#> 20      TRUE       19 airdas_sample.das       20    turtle
#> 21      TRUE       20 airdas_sample.das       21    turtle
#> 22      TRUE       21 airdas_sample.das       22    turtle
#> 23      TRUE       22 airdas_sample.das       23    turtle
#> 24      TRUE       23 airdas_sample.das       24    turtle
#> 25      TRUE       24 airdas_sample.das       25    turtle
#> 26      TRUE       25 airdas_sample.das       26    turtle
#> 27      TRUE       26 airdas_sample.das       27    turtle
#> 28      TRUE       27 airdas_sample.das       28    turtle
#> 29      TRUE       28 airdas_sample.das       29    turtle
#> 30      TRUE       29 airdas_sample.das       30    turtle
#> 31      TRUE       30 airdas_sample.das       31    turtle
#> 32      TRUE       31 airdas_sample.das       32    turtle
#> 33      TRUE       32 airdas_sample.das       33    turtle
#> 34      TRUE       34 airdas_sample.das       35    turtle
#> 35      TRUE       35 airdas_sample.das       36    turtle
#> 36     FALSE       36 airdas_sample.das       37    turtle
#> 37     FALSE       37 airdas_sample.das       38    turtle
#> 38     FALSE       38 airdas_sample.das       39    turtle
#> 39      TRUE       39 airdas_sample.das       40    turtle
#> 40      TRUE       40 airdas_sample.das       41    turtle
#> 41      TRUE       41 airdas_sample.das       42    turtle
#> 42      TRUE       42 airdas_sample.das       43    turtle
#> 43      TRUE       43 airdas_sample.das       44    turtle
#> 44      TRUE       44 airdas_sample.das       45    turtle
#> 45      TRUE       45 airdas_sample.das       46    turtle
#> 46      TRUE       46 airdas_sample.das       47    turtle
#> 47      TRUE       47 airdas_sample.das       48    turtle
#> 48     FALSE       48 airdas_sample.das       49    turtle
#> 49     FALSE       49 airdas_sample.das       50    turtle
#> 50     FALSE       50 airdas_sample.das       51    turtle
#> 51     FALSE       51 airdas_sample.das       52    turtle
#> 52      TRUE       52 airdas_sample.das       53    turtle
#> 53      TRUE       53 airdas_sample.das       54    turtle
#> 54      TRUE       54 airdas_sample.das       55    turtle
#> 55      TRUE       55 airdas_sample.das       56    turtle
#> 56      TRUE       56 airdas_sample.das       57    turtle
#> 57      TRUE       57 airdas_sample.das       58    turtle
#> 58      TRUE       58 airdas_sample.das       59    turtle
#> 59      TRUE       59 airdas_sample.das       60    turtle
#> 60      TRUE       60 airdas_sample.das       61    turtle
#> 61      TRUE       61 airdas_sample.das       62    turtle
#> 62      TRUE       62 airdas_sample.das       63    turtle
#> 63      TRUE       63 airdas_sample.das       64    turtle
#> 64      TRUE       64 airdas_sample.das       65    turtle
#> 65      TRUE       65 airdas_sample.das       66    turtle
#> 66      TRUE       66 airdas_sample.das       67    turtle
#> 67      TRUE       67 airdas_sample.das       68    turtle
#> 68      TRUE       68 airdas_sample.das       69    turtle
#> 69      TRUE       69 airdas_sample.das       70    turtle
#> 70      TRUE       70 airdas_sample.das       71    turtle
#> 71      TRUE       71 airdas_sample.das       72    turtle
#> 72      TRUE       72 airdas_sample.das       73    turtle
#> 73      TRUE       73 airdas_sample.das       74    turtle
#> 74      TRUE       74 airdas_sample.das       75    turtle
#> 75      TRUE       75 airdas_sample.das       76    turtle
#> 76      TRUE       76 airdas_sample.das       77    turtle
#> 77      TRUE       77 airdas_sample.das       78    turtle
#> 78      TRUE       78 airdas_sample.das       79    turtle
#> 79      TRUE       79 airdas_sample.das       80    turtle
#> 80      TRUE       80 airdas_sample.das       81    turtle
#> 81      TRUE       81 airdas_sample.das       82    turtle
#> 82      TRUE       82 airdas_sample.das       83    turtle
#> 83      TRUE       83 airdas_sample.das       84    turtle
#> 84      TRUE       84 airdas_sample.das       85    turtle
#> 85      TRUE       85 airdas_sample.das       86    turtle
#> 86      TRUE       86 airdas_sample.das       87    turtle
#> 87     FALSE       87 airdas_sample.das       88    turtle

y.read <- airdas_read(y)
airdas_process(y.read)
#>    Event            DateTime      Lat       Lon OnEffort Trans Bft CCover Jelly
#> 1      C 2015-04-09 12:30:59 39.23550 -123.1563    FALSE  <NA>  NA     NA    NA
#> 2      C 2015-04-09 12:30:59 39.23550 -123.1563    FALSE  <NA>  NA     NA    NA
#> 3      T 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 4      V 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 5      P 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 6      A 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1  NA     NA    NA
#> 7      W 2015-04-09 12:30:59 39.23550 -123.1563     TRUE    T1   1     10     0
#> 8      * 2015-04-09 12:31:00 39.23567 -123.1575     TRUE    T1   1     10     0
#> 9      S 2015-04-09 12:31:35 39.23367 -123.1783     TRUE    T1   1     10     0
#> 10     S 2015-04-09 12:31:47 39.23350 -123.1857     TRUE    T1   1     10     0
#> 11     C 2015-04-09 12:31:48 39.23350 -123.1857     TRUE    T1   1     10     0
#> 12     S 2015-04-09 12:31:56 39.23333 -123.1917     TRUE    T1   1     10     0
#> 13     * 2015-04-09 12:32:00 39.23317 -123.1930     TRUE    T1   1     10     0
#> 14     S 2015-04-09 12:32:20 39.23233 -123.2053     TRUE    T1   1     10     0
#> 15     S 2015-04-09 12:32:30 39.23233 -123.2108     TRUE    T1   1     10     0
#> 16     S 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 17     1 2015-04-09 12:32:42 39.23233 -123.2195     TRUE    T1   1     10     0
#> 18     * 2015-04-09 12:33:00 39.23267 -123.2292     TRUE    T1   1     10     0
#> 19     S 2015-04-09 12:33:43 39.23333 -123.2550     TRUE    T1   1     10     0
#> 20     * 2015-04-09 12:34:00 39.23300 -123.2637     TRUE    T1   1     10     0
#> 21     W 2015-04-09 12:34:52 39.23317 -123.2945     TRUE    T1   1     10     2
#> 22     C 2015-04-09 12:34:56 39.23317 -123.2968     TRUE    T1   1     10     2
#> 23     * 2015-04-09 12:35:00 39.23317 -123.2985     TRUE    T1   1     10     2
#> 24     S 2015-04-09 12:35:13 39.23317 -123.3073     TRUE    T1   1     10     2
#> 25     C 2015-04-09 12:35:32 39.23317 -123.3178     TRUE    T1   1     10     2
#> 26     * 2015-04-09 12:36:00 39.23333 -123.3337     TRUE    T1   1     10     2
#> 27     S 2015-04-09 12:36:13 39.23333 -123.3418     TRUE    T1   1     10     2
#> 28     S 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 29     t 2015-04-09 12:36:26 39.23350 -123.3500     TRUE    T1   1     10     2
#> 30     S 2015-04-09 12:36:43 39.23350 -123.3592     TRUE    T1   1     10     2
#> 31     * 2015-04-09 12:37:01 39.23383 -123.3685     TRUE    T1   1     10     2
#> 32     S 2015-04-09 12:37:01 39.23400 -123.3697     TRUE    T1   1     10     2
#> 33     S 2015-04-09 12:37:16 39.23433 -123.3790     TRUE    T1   1     10     2
#> 34     S 2015-04-09 12:37:37 39.23417 -123.3910     TRUE    T1   1     10     2
#> 35     W 2015-04-09 12:37:57 39.23400 -123.4033     TRUE    T1   2     10     0
#> 36     E 2015-04-09 12:38:00 39.23400 -123.4047    FALSE    T1   2     10     0
#> 37     C 2015-04-09 12:38:00 39.23400 -123.4047    FALSE    T1   2     10     0
#> 38     * 2015-04-09 12:39:00 39.23417 -123.4400    FALSE    T1   2     10     0
#> 39     R 2015-04-09 12:40:00 39.23400 -123.4742     TRUE    T1   2     10     0
#> 40     W 2015-04-09 12:40:28 39.23333 -123.4915     TRUE    T1   2     20     0
#> 41     C 2015-04-09 12:40:36 39.23317 -123.4960     TRUE    T1   2     20     0
#> 42     * 2015-04-09 12:41:00 39.23317 -123.5092     TRUE    T1   2     20     0
#> 43     W 2015-04-09 12:41:08 39.23317 -123.5143     TRUE    T1   1     20     0
#> 44     * 2015-04-09 12:42:00 39.23267 -123.5433     TRUE    T1   1     20     0
#> 45     C 2015-04-09 12:42:01 39.23267 -123.5433     TRUE    T1   1     20     0
#> 46     S 2015-04-09 12:42:13 39.23250 -123.5515     TRUE    T1   1     20     0
#> 47     * 2015-04-09 12:43:00 39.23233 -123.5778     TRUE    T1   1     20     0
#> 48     O 2015-04-09 12:43:43 39.23283 -123.6033    FALSE  <NA>   1     20     0
#> 49     * 2015-04-09 12:44:00 39.23300 -123.6118    FALSE  <NA>   1     20     0
#> 50     * 2015-04-09 12:45:00 39.24450 -123.6415    FALSE  <NA>   1     20     0
#> 51     * 2015-04-09 12:46:00 39.22167 -123.6373    FALSE  <NA>   1     20     0
#> 52     T 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2  NA     NA    NA
#> 53     V 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2  NA     NA    NA
#> 54     P 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2  NA     NA    NA
#> 55     A 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2  NA     NA    NA
#> 56     W 2015-04-09 12:46:51 39.22000 -123.6078     TRUE    T2   1     20     0
#> 57     * 2015-04-09 12:47:00 39.21950 -123.6025     TRUE    T2   1     20     0
#> 58     * 2015-04-09 12:48:00 39.21783 -123.5682     TRUE    T2   1     20     0
#> 59     A 2015-04-09 12:48:29 39.21767 -123.5515     TRUE    T2   1     20     0
#> 60     A 2015-04-09 12:48:36 39.21783 -123.5468     TRUE    T2   1     20     0
#> 61     * 2015-04-09 12:49:00 39.21783 -123.5333     TRUE    T2   1     20     0
#> 62     * 2015-04-09 12:50:00 39.21617 -123.4987     TRUE    T2   1     20     0
#> 63     A 2015-04-09 12:50:32 39.21683 -123.4797     TRUE    T2   1     20     0
#> 64     * 2015-04-09 12:51:00 39.21733 -123.4653     TRUE    T2   1     20     0
#> 65     S 2015-04-09 12:51:40 39.21750 -123.4410     TRUE    T2   1     20     0
#> 66     S 2015-04-09 12:51:55 39.21750 -123.4317     TRUE    T2   1     20     0
#> 67     * 2015-04-09 12:52:00 39.21733 -123.4293     TRUE    T2   1     20     0
#> 68     S 2015-04-09 12:52:03 39.21733 -123.4270     TRUE    T2   1     20     0
#> 69     S 2015-04-09 12:52:04 39.21733 -123.4270     TRUE    T2   1     20     0
#> 70     S 2015-04-09 12:52:07 39.21733 -123.4258     TRUE    T2   1     20     0
#> 71     S 2015-04-09 12:52:17 39.21700 -123.4185     TRUE    T2   1     20     0
#> 72     C 2015-04-09 12:52:39 39.21700 -123.4072     TRUE    T2   1     20     0
#> 73     S 2015-04-09 12:52:54 39.21700 -123.3982     TRUE    T2   1     20     0
#> 74     * 2015-04-09 12:53:00 39.21717 -123.3960     TRUE    T2   1     20     0
#> 75     * 2015-04-09 12:54:00 39.21683 -123.3617     TRUE    T2   1     20     0
#> 76     V 2015-04-09 12:54:01 39.21683 -123.3605     TRUE    T2   1     20     0
#> 77     W 2015-04-09 12:54:04 39.21683 -123.3605     TRUE    T2   2     20     0
#> 78     S 2015-04-09 12:54:34 39.21667 -123.3422     TRUE    T2   2     20     0
#> 79     * 2015-04-09 12:55:00 39.21667 -123.3285     TRUE    T2   2     20     0
#> 80     S 2015-04-09 12:55:20 39.21700 -123.3172     TRUE    T2   2     20     0
#> 81     * 2015-04-09 12:56:00 39.21700 -123.2960     TRUE    T2   2     20     0
#> 82     W 2015-04-09 12:56:40 39.21733 -123.2730     TRUE    T2   2     20     0
#> 83     C 2015-04-09 12:56:43 39.21717 -123.2718     TRUE    T2   2     20     0
#> 84     S 2015-04-09 12:56:49 39.21717 -123.2685     TRUE    T2   2     20     0
#> 85     * 2015-04-09 12:57:00 39.21700 -123.2617     TRUE    T2   2     20     0
#> 86     s 2015-04-09 12:57:00 39.21700 -123.2617     TRUE    T2   2     20     0
#> 87     O 2015-04-09 12:57:30 39.21700 -123.2417    FALSE  <NA>   2     20     0
#>    HorizSun VertSun  HKR  Haze  Kelp RedTide AltFt SpKnot ObsL ObsB ObsR  Rec
#> 1        NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 2        NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 3        NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 4        NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 5        NA      NA <NA> FALSE FALSE   FALSE    NA     NA   aa   bb   cc   dd
#> 6        NA      NA <NA> FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 7         6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 8         6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 9         6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 10        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 11        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 12        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 13        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 14        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 15        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 16        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 17        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 18        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 19        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 20        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 21        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 22        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 23        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 24        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 25        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 26        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 27        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 28        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 29        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 30        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 31        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 32        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 33        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 34        6      NA    r FALSE FALSE    TRUE   650    100   aa   bb   cc   dd
#> 35        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 36        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 37        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 38        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 39        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 40        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 41        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 42        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 43        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 44        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 45        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 46        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 47        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 48        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 49        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 50        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 51        6      NA    n FALSE FALSE   FALSE   650    100   aa   bb   cc   dd
#> 52       NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 53       NA      NA <NA> FALSE FALSE   FALSE    NA     NA <NA> <NA> <NA> <NA>
#> 54       NA      NA <NA> FALSE FALSE   FALSE    NA     NA   aa   dd   cc   bb
#> 55       NA      NA <NA> FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 56       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 57       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 58       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 59       12      NA    n FALSE FALSE   FALSE   680    100   aa   dd   cc   bb
#> 60       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 61       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 62       12      NA    n FALSE FALSE   FALSE   650    100   aa   dd   cc   bb
#> 63       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 64       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 65       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 66       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 67       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 68       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 69       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 70       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 71       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 72       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 73       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 74       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 75       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 76       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 77       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 78       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 79       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 80       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 81       12      NA    n FALSE FALSE   FALSE   680     90   aa   dd   cc   bb
#> 82       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 83       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 84       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 85       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 86       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#> 87       12      NA    r FALSE FALSE    TRUE   680     90   aa   dd   cc   bb
#>     VLI  VLO   VB  VRI  VRO Data1 Data2 Data3 Data4 Data5 Data6           Data7
#> 1  <NA> <NA> <NA> <NA> <NA>  Reco rder:  dd    <NA>  <NA>  <NA>            <NA>
#> 2  <NA> <NA> <NA> <NA> <NA>  Not  recor ding  molas  <NA>  <NA>            <NA>
#> 3  <NA> <NA> <NA> <NA> <NA>    T1  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 4     g    g    g    g    g     g     g     g     g     g  <NA>            <NA>
#> 5     g    g    g    g    g    aa    bb    cc    dd  <NA>  <NA>            <NA>
#> 6     g    g    g    g    g   650   100  <NA>  <NA>  <NA>  <NA>            <NA>
#> 7     g    g    g    g    g     n    10     1     0     6  <NA>            <NA>
#> 8     g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 9     g    g    g    g    g     1    cc    67     6    mn  <NA>            <NA>
#> 10    g    g    g    g    g     2    bb   -60     6    mn  <NA>            <NA>
#> 11    g    g    g    g    g  2 cp  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 12    g    g    g    g    g     3    cc    38     2    mn  <NA>            <NA>
#> 13    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 14    g    g    g    g    g     4    aa   -50     5    mn  <NA>            <NA>
#> 15    g    g    g    g    g     5    cc    49     4    bm  <NA>            <NA>
#> 16    g    g    g    g    g     6    cc    42    10    gg    er            <NA>
#> 17    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>    80    20            <NA>
#> 18    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 19    g    g    g    g    g     7    aa   -53     5    bm  <NA>            <NA>
#> 20    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 21    g    g    g    g    g     r    10     1     2     6  <NA>            <NA>
#> 22    g    g    g    g    g  stre aky r ed ti de     <NA>  <NA>            <NA>
#> 23    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 24    g    g    g    g    g     8    dd    24     2    bm  <NA>            <NA>
#> 25    g    g    g    g    g  shar k 11f t      <NA>  <NA>  <NA>            <NA>
#> 26    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 27    g    g    g    g    g     9    aa   -57     4    mn  <NA>            <NA>
#> 28    g    g    g    g    g    10    cc    65     5    gg  <NA>            <NA>
#> 29    g    g    g    g    g    aa   -20    dc     5    90     n            <NA>
#> 30    g    g    g    g    g    11    bb    87     3    pp  <NA>            <NA>
#> 31    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 32    g    g    g    g    g    12    bb    82     5    gg  <NA>            <NA>
#> 33    g    g    g    g    g    13    bb    54     6    pp  <NA>            <NA>
#> 34    g    g    g    g    g    14    cc    62     2    gg  <NA>            <NA>
#> 35    g    g    g    g    g     n    10     2     0     6  <NA>            <NA>
#> 36    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 37    g    g    g    g    g  off  effor t to  circl e on  unide ntifed object  
#> 38    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 39    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 40    g    g    g    g    g     n    20     2     0     6  <NA>            <NA>
#> 41    g    g    g    g    g  wate r got  more  blue  <NA>  <NA>            <NA>
#> 42    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 43    g    g    g    g    g     n    20     1     0     6  <NA>            <NA>
#> 44    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 45    g    g    g    g    g  fb2s  fb1m  <NA>  <NA>  <NA>  <NA>            <NA>
#> 46    g    g    g    g    g    15    bb    50     3    mn  <NA>            <NA>
#> 47    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 48    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 49    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 50    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 51    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 52 <NA> <NA> <NA> <NA> <NA>    T2  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 53    g    g    g    g    g     g     g     g     g     g  <NA>            <NA>
#> 54    g    g    g    g    g    aa    dd    cc    bb  <NA>  <NA>            <NA>
#> 55    g    g    g    g    g   650   100  <NA>  <NA>  <NA>  <NA>            <NA>
#> 56    g    g    g    g    g     n    20     1     0    12  <NA>            <NA>
#> 57    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 58    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 59    g    g    g    g    g   680   100  <NA>  <NA>  <NA>  <NA>            <NA>
#> 60    g    g    g    g    g   650   100  <NA>  <NA>  <NA>  <NA>            <NA>
#> 61    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 62    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 63    g    g    g    g    g   680    90  <NA>  <NA>  <NA>  <NA>            <NA>
#> 64    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 65    g    g    g    g    g    16    aa   -31     6    mn  <NA>            <NA>
#> 66    g    g    g    g    g    17    bb   -59     6    bm  <NA>            <NA>
#> 67    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 68    g    g    g    g    g    18    aa   -62     2    gg  <NA>            <NA>
#> 69    g    g    g    g    g    19    cc    42     3    bm  <NA>            <NA>
#> 70    g    g    g    g    g    20    aa   -48     4    bm  <NA>            <NA>
#> 71    g    g    g    g    g    21    cc    63     6    pp  <NA>            <NA>
#> 72    g    g    g    g    g  bell y has  some  glar e      <NA>            <NA>
#> 73    g    g    g    g    g    22    cc    43     2    pp  <NA>            <NA>
#> 74    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 75    g    g    g    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 76    g    g    p    g    g     g     g     p     g     g  <NA>            <NA>
#> 77    g    g    p    g    g     n    20     2     0    12  <NA>            <NA>
#> 78    g    g    p    g    g    23    cc    56     6    mn  <NA>            <NA>
#> 79    g    g    p    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 80    g    g    p    g    g    24    cc    37     6    mn  <NA>            <NA>
#> 81    g    g    p    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 82    g    g    p    g    g     r    20     2     0    12  <NA>            <NA>
#> 83    g    g    p    g    g  patc hy re d tid e      <NA>  <NA>            <NA>
#> 84    g    g    p    g    g    25    bb   -76     1    pp  <NA>            <NA>
#> 85    g    g    p    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#> 86    g    g    p    g    g    25   -70  <NA>  <NA>  <NA>  <NA>            <NA>
#> 87    g    g    p    g    g  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>            <NA>
#>    EffortDot EventNum          file_das line_num file_type
#> 1      FALSE        1 airdas_sample.das        1    turtle
#> 2      FALSE        2 airdas_sample.das        2    turtle
#> 3       TRUE        3 airdas_sample.das        3    turtle
#> 4       TRUE        4 airdas_sample.das        4    turtle
#> 5       TRUE        5 airdas_sample.das        5    turtle
#> 6       TRUE        6 airdas_sample.das        6    turtle
#> 7       TRUE        7 airdas_sample.das        7    turtle
#> 8       TRUE        8 airdas_sample.das        8    turtle
#> 9       TRUE        9 airdas_sample.das        9    turtle
#> 10      TRUE       10 airdas_sample.das       10    turtle
#> 11      TRUE       11 airdas_sample.das       11    turtle
#> 12      TRUE       12 airdas_sample.das       12    turtle
#> 13      TRUE       13 airdas_sample.das       13    turtle
#> 14      TRUE       14 airdas_sample.das       14    turtle
#> 15      TRUE       15 airdas_sample.das       15    turtle
#> 16      TRUE       16 airdas_sample.das       16    turtle
#> 17     FALSE     <NA> airdas_sample.das       17    turtle
#> 18      TRUE       17 airdas_sample.das       18    turtle
#> 19      TRUE       18 airdas_sample.das       19    turtle
#> 20      TRUE       19 airdas_sample.das       20    turtle
#> 21      TRUE       20 airdas_sample.das       21    turtle
#> 22      TRUE       21 airdas_sample.das       22    turtle
#> 23      TRUE       22 airdas_sample.das       23    turtle
#> 24      TRUE       23 airdas_sample.das       24    turtle
#> 25      TRUE       24 airdas_sample.das       25    turtle
#> 26      TRUE       25 airdas_sample.das       26    turtle
#> 27      TRUE       26 airdas_sample.das       27    turtle
#> 28      TRUE       27 airdas_sample.das       28    turtle
#> 29      TRUE       28 airdas_sample.das       29    turtle
#> 30      TRUE       29 airdas_sample.das       30    turtle
#> 31      TRUE       30 airdas_sample.das       31    turtle
#> 32      TRUE       31 airdas_sample.das       32    turtle
#> 33      TRUE       32 airdas_sample.das       33    turtle
#> 34      TRUE       34 airdas_sample.das       35    turtle
#> 35      TRUE       35 airdas_sample.das       36    turtle
#> 36     FALSE       36 airdas_sample.das       37    turtle
#> 37     FALSE       37 airdas_sample.das       38    turtle
#> 38     FALSE       38 airdas_sample.das       39    turtle
#> 39      TRUE       39 airdas_sample.das       40    turtle
#> 40      TRUE       40 airdas_sample.das       41    turtle
#> 41      TRUE       41 airdas_sample.das       42    turtle
#> 42      TRUE       42 airdas_sample.das       43    turtle
#> 43      TRUE       43 airdas_sample.das       44    turtle
#> 44      TRUE       44 airdas_sample.das       45    turtle
#> 45      TRUE       45 airdas_sample.das       46    turtle
#> 46      TRUE       46 airdas_sample.das       47    turtle
#> 47      TRUE       47 airdas_sample.das       48    turtle
#> 48     FALSE       48 airdas_sample.das       49    turtle
#> 49     FALSE       49 airdas_sample.das       50    turtle
#> 50     FALSE       50 airdas_sample.das       51    turtle
#> 51     FALSE       51 airdas_sample.das       52    turtle
#> 52      TRUE       52 airdas_sample.das       53    turtle
#> 53      TRUE       53 airdas_sample.das       54    turtle
#> 54      TRUE       54 airdas_sample.das       55    turtle
#> 55      TRUE       55 airdas_sample.das       56    turtle
#> 56      TRUE       56 airdas_sample.das       57    turtle
#> 57      TRUE       57 airdas_sample.das       58    turtle
#> 58      TRUE       58 airdas_sample.das       59    turtle
#> 59      TRUE       59 airdas_sample.das       60    turtle
#> 60      TRUE       60 airdas_sample.das       61    turtle
#> 61      TRUE       61 airdas_sample.das       62    turtle
#> 62      TRUE       62 airdas_sample.das       63    turtle
#> 63      TRUE       63 airdas_sample.das       64    turtle
#> 64      TRUE       64 airdas_sample.das       65    turtle
#> 65      TRUE       65 airdas_sample.das       66    turtle
#> 66      TRUE       66 airdas_sample.das       67    turtle
#> 67      TRUE       67 airdas_sample.das       68    turtle
#> 68      TRUE       68 airdas_sample.das       69    turtle
#> 69      TRUE       69 airdas_sample.das       70    turtle
#> 70      TRUE       70 airdas_sample.das       71    turtle
#> 71      TRUE       71 airdas_sample.das       72    turtle
#> 72      TRUE       72 airdas_sample.das       73    turtle
#> 73      TRUE       73 airdas_sample.das       74    turtle
#> 74      TRUE       74 airdas_sample.das       75    turtle
#> 75      TRUE       75 airdas_sample.das       76    turtle
#> 76      TRUE       76 airdas_sample.das       77    turtle
#> 77      TRUE       77 airdas_sample.das       78    turtle
#> 78      TRUE       78 airdas_sample.das       79    turtle
#> 79      TRUE       79 airdas_sample.das       80    turtle
#> 80      TRUE       80 airdas_sample.das       81    turtle
#> 81      TRUE       81 airdas_sample.das       82    turtle
#> 82      TRUE       82 airdas_sample.das       83    turtle
#> 83      TRUE       83 airdas_sample.das       84    turtle
#> 84      TRUE       84 airdas_sample.das       85    turtle
#> 85      TRUE       85 airdas_sample.das       86    turtle
#> 86      TRUE       86 airdas_sample.das       87    turtle
#> 87     FALSE       87 airdas_sample.das       88    turtle